### 2020-04-29 #81 @whedon
[PRE REVIEW]: ClimateEstimate.net: A tutorial on climate econometrics
**Submitting author:** @jrising (<a href="http://orcid.org/0000-0001-8514-4748">James Rising</a>)
**Repository:** <a href="https://github.com/atrisovic/weather-panel.github.io" target ="_blank">https://github.com/atrisovic/weather-panel.github.io</a>
**Version:** v1.2

@jrising if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
