+++ paper.bib
@@ -6,7 +6,8 @@
   number={6304},
   pages={aad9837},
   year={2016},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.aad9837}
 }
 @article{hsiang2018economist,
   title={An economist's guide to climate change science},
@@ -15,7 +16,8 @@
   volume={32},
   number={4},
   pages={3--32},
-  year={2018}
+  year={2018},
+  doi={10.3386/w25189}
 }
 @article{hsiang2016climate,
   title={Climate econometrics},
@@ -24,7 +26,8 @@
   volume={8},
   pages={43--75},
   year={2016},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.3386/w22181}
 }
 @article{ciscar2019assessing,
   title={Assessing future climate change impacts in the EU and the USA: insights and lessons from two continental-scale projects},
@@ -34,7 +37,8 @@
   number={8},
   pages={084010},
   year={2019},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1748-9326/ab281e}
 }
 @article{nissan2019use,
   title={On the use and misuse of climate change projections in international development},
@@ -44,5 +48,6 @@
   number={3},
   pages={e579},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wcc.579}
 }
