### 2020-05-03 4b96bd1
Edits to the reduced form specification
