+++ content/reduced-form-specification.md
@@ -4,15 +4,43 @@ This section describes some of the considerations that go into developing a redu
 
 ## 2.1. Choosing weather variables
 
-The choice of weather variables depends on the question we are trying to answer, and there are many forms to represent any given variable. For example, in the case of temperature, we can use $T_{avg}$, $T_{min}$, $T_{max}$, days above 30 C, heating and cooling degree-days, or growing degree-days. Morevoer, it is very important to first think about possible *mechanism(s)* behind a change witnessed in the environment, and then only make choices for variables that can explain that mechanism. A few of the important and frequently-used weather variables are listed below:
+The choice of weather variables depends on the question we are trying
+to answer, and there are many forms to represent any given
+variable. For example, in the case of temperature, we can use
+$T_{avg}$, $T_{min}$, $T_{max}$, days above 30 C, heating and cooling
+degree-days, or growing degree-days. Morevoer, it is very important to
+first think about possible *mechanism(s)* behind a change witnessed in
+the environment, and then only make choices for variables that can
+explain that mechanism. A few of the important and frequently-used
+weather variables are listed below, and why you might choose them:
 
-- **Temperature:** There are various measures of temperature that can be used. Some of them are listed below:
-    1. *$T_{min}$, $T_{max}$:*  Many socioeconomic processes are more sensitive to extreme temperatures than to variation in the average. This is also useful when temperature variation is large, leading to significant differences in cold end and hot end response. These are important metrics when heterogeneity between each time unit matters, and may better capture heat waves and cold spells. Also, note that $T_{min}$ better reflects nighttime temperatures while $T_{max}$ better reflects daytime temperatures. Not all datasets include $T_{min}$ or $T_{max}$. 
-    2. *$T_{avg}$:*  A good mean metric for seeing average response over the temperature support, when there is not much variation in temperature across time unit considered in the study. $T_{avg}$ is most appropriate when there is some natural inertia in the response, so that the dependent variable is responding to a kind of average over the last 24 hours. Note that $T_{avg}$ is often just $(T_{min} + T_{max}) / 2$, unless calculated from sub-daily data.
-    3. [*HDD/CDD & GDD:*](https://www.degreedays.net/introduction) Degree days (DD) are a measure of ’how much’ and for ’how long’ the outside air temperature was above or below a certain level.
+- **Temperature:** Temperature relationships are often preferred in
+  climate impacts research, because temperature is more predictable
+  than many other weather variables. There are various measures of temperature that can be used. Some of them are listed below:
+    1. *$T_{min}$, $T_{max}$:*  Many socioeconomic processes are more
+       sensitive to extreme temperatures than to variation in the
+       average. This is also useful when temperature variation is
+       large, leading to significant differences in the cold end and hot end responses. These are important metrics when heterogeneity within time units matters, and may better capture heat waves and cold spells. Also, note that $T_{min}$ better reflects nighttime temperatures while $T_{max}$ better reflects daytime temperatures. Not all datasets include $T_{min}$ or $T_{max}$. 
+    2. *$T_{avg}$:*  A good mean metric for seeing average response
+       over the temperature support, when there is not much variation
+       in temperature within each time unit considered in the
+       study. $T_{avg}$ is most appropriate when there is some natural
+       inertia in the response, so that the dependent variable is
+       responding to a kind of average over the last 24 hours. Note
+       that $T_{avg}$ is often just equalt to $(T_{min} + T_{max}) / 2$, unless calculated from sub-daily data.
+    3. [*HDD/CDD & GDD:*](https://www.degreedays.net/introduction)
+       Degree days (DD) are a measure of ’how much’ and for ’how long’
[…]
