### 2020-05-25 2265ee3
edits to scale section
