+++ content/reduced-form-specification.md
@@ -86,11 +86,119 @@ Some datasets (such as [HadEX2](https://climatedataguide.ucar.edu/climate-data/h
          throughout East Africa, South and Southeast Asia, and
          Oceania.
 
-## 2.2. Common functional forms (pros, cons, and methods)
+## 2.2. Dealing with the spatial and temporal scales of economic processes
 
-Different functional forms serve different purposes. First, think about the "true model" that relates your dependent variable to your weather variables, and then try to turn it into a linear expression that you can estimate. Some of the frequently used functional forms along with a good reference for understanding them in detail are listed below.  
+The process of developing a reduced-form specification starts with a
+study of the "true model", or data-generating process, that relates
+your dependent variable to your weather variables. A crucial aspect of
+that relationship is the question of scale.
 
-For understanding the methods part we are using transformation-before-aggregation method, which is explained in section 2.4 below.  
+Weather data products are generally available in *gridded* form,
+developed through careful interpolation and/or reanalysis. The grids
+used can vary in size across datasets, but they can be aggregated to
+administrative units like county, city, etc., using appropriate
+weighted aggregation methods. Think about the scale of your
+administrative units, relative to the scale of the grid cells. If the
+regions are much bigger than the grid cells, a weighted average across
+included cells is appropriate. If the regions are much smaller than
+the cells, it will probably be necessary to aggregate the regions,
+since the level of variation is only at the grid cell level. If the
+two are of similar sizes, it may be necessary to account for the
+amount of each grid cell lies within each region. This can be
+calculated as a transformation matrix, with a row for each region and
+a column for each cell. Once the matrix is calculated, it can be
+reused for each time step. More details for this process are described
+in sections 3 and 5.
+
+Typically, relating weather to a dependent variable requires some kind
+of non-linear transformation. For example, estimating a polynomial
+functional form requires raising the temperatures to various powers. Importantly, the square of a weighted average of grid-level temperatures is not the same as the weighted average of the square of grid-level temperatures.
+
+While doing the spatial aggregation, we need to decide whether we want
+to transform the data first and then aggregate it
+(transformation-before-aggregation) or aggregate it and then transform
+it (aggregation-before-transformation). This decision is based on the
+whether the phenomenon in consideration is occurring at the local
+(grid or individual) scale or at the larger administrative units
+(country, state, county, etc.) scale. Also, it matters what variable
+is being consideration. For example, doing
+aggregation-before-transformation for temperature will distort the
+signal less that doing it for precipitation. This is because
+precipitation is highly local both temporally and spatially; it could
+rain for <1 min in <1 km radius area.
+
+### Transformation-before-aggregation
+
+When an economic process is occurring at the local level (for example,
+for individuals or households), we need to first do our estimation at
+the grid level. For example, to estimate the effect of temperature on
+human mortality at the county level, we should reckon that the effect
+of temperature on mortality is a local phenomenon, so the estimation
+should happen at the lowest possible level.  Since the dependent
+variable is a sum of individual-level outcomes, we should write down
+the reduced-form specification for an individual experiencing
+high-resolution weather, and then sum across all of those reduced
[…]
