### 2020-05-25 fd988c4
edits to scale section
