+++ content/reduced-form-specification.md
@@ -154,13 +154,13 @@ $$f(T_{ps}) = \beta_1 g_1(T_{ps}) + \beta_2 g_2(T_{ps}) + \cdots + \beta_k g_k(T
 where $g_k(T_{ps})$ is a transformation of the weather variables. For example, for a cubic response curve, $g_1(T_{ps}) = T_{ps}$, $g_2(T_{ps}) = T_{ps}^2$, and $g_3(T_{ps}) = T_{ps}^3$.  
 
 We know that  
-$$y_{it} = \sum_{s\in t} \sum_{j\in i} y_{js} = \sum_{s\in t} \sum_{j\in i}$$
+$$y_{it} = \sum_{s\in t} \sum_{j\in i} y_{js} = \sum_{s\in t} \sum_{p\in i}$$
 $$\beta_1 g_1(T_{ps}) + \beta_2 g_2(T_{ps}) + \cdots + \beta_k g_k(T_{ps})$$  
 
 We can rearrange this to  
-$$y_{it} = \beta_1 (\sum_{s\in t} \sum_{j\in i} g_1(T_{ps})) + $$
-$$\beta_2 (\sum_{s\in t} \sum_{j\in i} g_2(T_{ps})) + \cdots + $$
-$$\beta_k (\sum_{s\in t} \sum_{j\in i} g_k(T_{ps}))$$  
+$$y_{it} = \beta_1 (\sum_{s\in t} \sum_{p\in i} g_1(T_{ps})) + $$
+$$\beta_2 (\sum_{s\in t} \sum_{p\in i} g_2(T_{ps})) + \cdots + $$
+$$\beta_k (\sum_{s\in t} \sum_{p\in i} g_k(T_{ps}))$$  
 
 That is, the variables used in the regression should be the sum over
 weather data that has been transformed at the grid level.
@@ -182,9 +182,9 @@ offices, and extremely large values for those (very few) cells, which
 do.
 
 Using the formulation above, here we would regress:
-$$y_{it} = \beta_1 g_1(\sum_{s\in t} \sum_{j\in i} T_{ps}) + $$
-$$\beta_2 g_2(\sum_{s\in t} \sum_{j\in i} T_{ps}) + \cdots + $$
-$$\beta_k g_k(\sum_{s\in t} \sum_{j\in i} T_{ps})$$
+$$y_{it} = \beta_1 g_1(\sum_{s\in t} \sum_{p\in i} T_{ps}) + $$
+$$\beta_2 g_2(\sum_{s\in t} \sum_{p\in i} T_{ps}) + \cdots + $$
+$$\beta_k g_k(\sum_{s\in t} \sum_{p\in i} T_{ps})$$
 
 Where $T_{ps}$ is the gridded weather for cell $p$ in time step
 $s$, $g_k(\cdot)$ is the non-linear transformation (e.g., raising to
@@ -198,39 +198,42 @@ observed for region $i$ in reporting period $t$. Weather data products can have 
 Returning to the "true model" of your process, the decisions around how to
 generate a linear expression that you can estimate have important
 implications. Different functional forms serve different purposes and
-describe different underlying relationships. Some of the frequently used functional forms along with a good reference for understanding them in detail are listed below.
-
-
[…]
