### 2020-08-27 #81 @labarba
👋  @atrisovic — I am the single-point-of-failure here, as I'm heavily overcommitted. I am trying to start chipping away at neglected tasks, like JOSE editing.

I read your paper and I'm going to send it back to you for some more work, because it does not quite meet our expectations. In particular, see the section of our Documentation about the [What should my paper contain?](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain). We ask that you

> Include a “Statement of Need” section, explaining how the submitted artifacts contribute to computationally enabled teaching and learning, and describing how they might be adopted by others.
> For learning modules, describe the learning objectives, content, instructional design, and experience of use in teaching and learning situations.

The text does address the gap this fills, but we ask for a section heading for the Statement of Need. The second item is missing, and is important for another instructor to imagine how they might re-use your tutorial for their teaching needs.

Thank you for your patience! Tough times.
