+++ content/reduced-form-specification.md
@@ -259,7 +259,7 @@ high-resolution weather values should be raised to powers before
 aggregating up to the regions in the economic data. That is, the the
 predictor for the $k$-th power of the polynomial is 
     $$f(T_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} T_{p i t}^k$$ 
-    where $\psi_{p}$ is the weight assigned to the $p$ grid.  
+    where $\psi_{p}$ is the weight assigned to the $p$ gridcell.  
     
 The dose-response regression would then be applied as follows
     $$F(T_{it})=\sum_{k} \beta_k f(T_{it}^k)$$
@@ -269,39 +269,60 @@ dose-response relationship:
     $$F(T_{pit})=\sum_{k} \beta_k T_{pit}^k$$
 
 - **[Restricted cubic spline](https://support.sas.com/resources/papers/proceedings16/5621-2016.pdf)**
-    1. Fitting a piecewise polynomial function between pre-specified knots
-    2. More independence compared to poly in choosing function knots
-    3. Highly parametric due to freedom of choice of knots
-    
-    For transforming the temperature data into restricted cubic splines, we need to fix the location and the number of knots.     The reference above on cubic splines can be helpful in deciding the knot specifications. As before let the grid $p$       temperature be $T_{p i t}$. Let us do this exercise for $n$ knots, placed at $t_1<t_2<...<t_n$, then for $T_{p i     t}$, which is a continuous variable, we have a set of $(n-2)$ new variables. We have:  
-    $$f(T_{i t}^k)= \sum_{p \in \Theta} \psi_{p}*\{(T_{p i t}-t_k)^3_+ - (T_{p i t} - t_{n-                   1})^3_+*\frac{t_n-t_k}{t_n-t_{n-1}}+(T_{p i t} - t_{n})^3_+*\frac{t_{n-1}-t_k}{t_{n}-t_{n-1}}\}$$ $$\forall k \in \{1,2,...,n-2\}$$ 
+
+Restricted cubic splines produce smooth dose-response curves, like
+polynomials, but mitigate some of the problems that polynomials with
+many terms have. Whereas polynomials with high-order terms can produce
+very extreme results under extrapolation, RCS always produces a linear
+extrapolation. RCS also provides additional degrees of freedom through
+the placement of knots, and this can be used either to reflect
+features of the underlying process being modeled or to improve the
+fit.
+
+In the case where knots are choosen to maximize the fit of the curve,
+cross-validation is the preferred approach for both selecting the
+number of knots and their placement. The reference in this subsection
+title on cubic splines can be helpful in deciding the knot
+specifications.
+
+Once knot locations are determiend, the weather data needs to be
+translated into RCS terms. As before let the gridded weather be $T_{p
+i t}$ and let there be $n$ knots, placed at $T_1<T_2<...<T_n$. Then we
+have a set of $(n-2)$ terms, here indexed by $k$ and defined as:
+    $$f(T_{i t})_k= \sum_{p \in \Theta(i)} \psi_{p} \{(T_{p i
+	t}-T_k)^3_+ - (T_{p i t} - T_{n-1})^3_+
+	\frac{T_n-T_k}{T_n-T_{n-1}}+(T_{p i t} - T_{n})^3_+ \frac{T_{n-1}-T_k}{T_{n}-T_{n-1}}\}$$ $$\forall k \in \{1,2,...,n-2\}$$ 
     where, $\psi_{p}$ is the weight assigned to the $p$ grid.  
     
[…]
