### 2020-09-04 58ca0f4
Updated discussion on observations; changed recommendations
