+++ content/weather-and-climate-data.md
@@ -47,11 +47,15 @@ Typically, there will be variables that correspond to each of the dimensions, an
 
 The netCDF file structure is self-describing, meaning all the information you need to understand what data are within are contained within the file as well (in theory). *However*, the format doesn’t require any information be put there, so names of attributes, which attributes are included, etc., may vary between files, especially if they’re ‘unofficial’ files not created by a major modeling group as part of a larger project.
 
-But fear not: thankfully, most data you will be facing has standardized to something akin to the format used by CMIP5 (a ‘model intercomparison project’ in which different modeling groups agreed to run their models on identical climate ‘experiments’ and publish their data in a uniform format).
+But fear not: thankfully, most data you will be facing has standardized to something akin to the format used by CMIP output (a ‘model intercomparison project’ in which different modeling groups agreed to run their models on identical climate ‘experiments’ and publish their data in a uniform format; CMIP6 is the latest generation of the largest 'MIP', analysis of which makes up a substantial portion of IPCC reports).
 
-For example, most climate data you will encounter will also follow a CMIP5 filename format, of the form:
+For example, climate data you will encounter will generally follow a CMIP5 filename format, of the form:
 
 `[variable shorthand]_[frequency]_[model name]_[experiment]_[run]_[timeframe].nc`
+
+or a CMIP6 filename format, of the form:
+
+`[variable shorthand]_[frequency]_[model name]_[experiment]_[run]_[grid label]_[timeframe].nc`
 
 (This terminology will also be useful to recognize even when filenames are not of this format, as is likely for weather data products.)
 
@@ -61,7 +65,10 @@ For example, most climate data you will encounter will also follow a CMIP5 filen
 - **frequency**: *day* for daily, *ann* for annual. *mon* for monthly; is sometimes listed as *Amon* for atmospheric variables; feel free to ignore this distinction.
 - **experiment**: some descriptor of the forcing used (forcing = profile of greenhouse gases), e.g. *rcp85* for the RCP8.5 scenario frequently used in projections.
 - **run**: if the same model was run multiple times with the same forcing, but different physics or initial conditions, it will be noted here (e.g. *r1i1p1*). Don’t worry about this.
+- **grid label**: whether data was regridded from the model's native grid
 - **timeframe**: frequently in `yyyymmdd-yyyymmdd` format.
+
+(For more information on "CMIP5" and "CMIP6" terminology, please see: [CMIP6 Guidance for Data Users](https://pcmdi.llnl.gov/CMIP6/Guide/dataUsers.html) or the [CMIP5 Standard Output](https://pcmdi.llnl.gov/mips/cmip5/requirements.html))
 
 There are two common ways in which data is stored in netCDF files:
 
@@ -110,7 +117,7 @@ You’ll often only want or need a subset of a variable. In this case, make sure
 | ------------------------------------------ | ----------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | `ncread(fn,var,`<br>`[1 1 1],[5 5 365]);` | `ncfile <- nc_open(fn)`<br>`vardata <- ncvar_get(ncfile,var,             start=c(1,1,1), count=c(5,5,365))` | `ds = xr.open_dataset(fn))`<br>`ds.loc[0:5,0:5,0:365]` <br>(data is loaded lazily and only fully loaded when calculations are done, so you can slice data without loading it into memory. Slicing can be done by time, variable, etc.; check the docs) |
 
[…]
