+++ content/reduced-form-specification.md
@@ -38,7 +38,7 @@ weather variables are listed below, and why you might choose them:
        study. $T_{avg}$ is most appropriate when there is some natural
        inertia in the response, so that the dependent variable is
        responding to a kind of average over the last 24 hours. Note
-       that $T_{avg}$ is often just equalt to $(T_{min} + T_{max}) / 2$, unless calculated from sub-daily data.
+       that $T_{avg}$ is often just equal to $(T_{min} + T_{max}) / 2$, unless calculated from sub-daily data.
     3. [*HDD/CDD & GDD:*](https://www.degreedays.net/introduction)
        Degree days (DD) are a measure of ’how much’ and for ’how long’
        the outside air temperature was above or below a certain
@@ -128,11 +128,11 @@ functional form requires raising the temperatures to various powers. Importantly
 While doing the spatial aggregation, we need to decide whether we want
 to transform the data first and then aggregate it
 (transformation-before-aggregation) or aggregate it and then transform
-it (aggregation-before-transformation). This decision is based on the
+it (aggregation-before-transformation). This decision is based on
 whether the phenomenon in consideration is occurring at the local
 (grid or individual) scale or at the larger administrative units
 (country, state, county, etc.) scale. Also, it matters what variable
-is being consideration. For example, doing
+is under consideration. For example, doing
 aggregation-before-transformation for temperature will distort the
 signal less that doing it for precipitation. This is because
 precipitation is highly local both temporally and spatially; it could
@@ -158,7 +158,7 @@ We want to understand how local agents respond to weather shocks. Suppose that t
 However, we do not observe agent-level responses. Instead, we have region-wide sums, $y_{it}$ for region $i$ and reporting period $t$. For example, if $y_{js}$ is death risk for agent $j$ for a given day, we may only observe total deaths across a region in each year, $$y_{it} = \sum_{s \in t} \sum_{j \in i} y_{js}.$$  
 
 We can determine the agent-level response $f(T_{ps})$ if we assume
-linearity. First, let us represent this the way we would if we could run a regression with agent-level data, breaking up the dose-response
+linearity. First, let us represent this as if we could run a regression with agent-level data, breaking up the dose-response
 curve into a sum of terms:  
 $$f(T_{ps}) = \beta_1 g_1(T_{ps}) + \beta_2 g_2(T_{ps}) + \cdots + \beta_k g_k(T_{ps})$$  
 
@@ -188,7 +188,7 @@ service employment at the administrative office level, we need to take
 into account the fact that hiring/firing of public service employees
 happens at the office level only.  Estimating grid-level effects will
 lead to wrong estimation, as it should result in zero estimate for
-those (almost all) grid cells which do not have contain administrative
+those (almost all) grid cells which do not contain administrative
 offices, and extremely large values for those (very few) cells, which
 do.
 
@@ -222,7 +222,7 @@ for projections: only the estimate for the extreme bins will be used.
 The bin widths should generally be even to facilitate easy
 understanding, although the lowest and highest bins can go to negative
 and positive infinity (or the appropriate analog). You may want to
-have smaller sized bins around weather values where we expect most of
+have smaller size bins around weather values where we expect most of
 the response to occur and where there is a lot of data, but be
 prepared to also show evenly-spaced bins.
 
@@ -322,7 +322,7 @@ And, each spline term in the parentheses $(\nabla)_+$ e.g. $(T_{p i t} - T_2)_+$
     $\nabla_+=\nabla_+$ if $\nabla_+>0$  
     $\nabla_+=0$ if $\nabla_+<0$  
 
[…]
