### 2020-09-06 6428527
formula rendering fix
