### 2020-09-09 2031c19
Update paper.md
Revision after submission.
