### 2020-09-09 #81 @jrising
@labarba Thank you again. We have now extensively revised (if I can use that phrase with such a short paper) our `paper.md`. There is now a statement of need, and adding more text on the structure of the tutorial, its objectives, and our experience using it. If you are available, we would very much appreciate it if you took a look again.

In the meantime, the tutorial has also been improved. Thanks!
