### 2020-09-22 #81 @kyleniemeyer
Hi @labarba @jrising @atrisovic I can take over editing this submission. It does look like the authors have revised the paper to include the material that was missing originally, and I think we can now identify reviewers and move forward.
