### 2020-09-22 #81 @kyleniemeyer
Hello @jwagemann  @angela-li and @howardjp , this authors of this submission to JOSE (Journal of Open Source Education, https://jose.theoj.org) suggested you as potential reviewers for their submission "ClimateEstimate.net: A tutorial on climate econometrics".

Are you interested and available to review this for us? Thanks!
