### 2020-09-22 #90 @whedon
[REVIEW]: ClimateEstimate.net: A tutorial on climate econometrics
**Submitting author:** @jrising (<a href="http://orcid.org/0000-0001-8514-4748">James Rising</a>)
**Repository:** https://github.com/atrisovic/weather-panel.github.io
**Branch with paper.md** (empty if default branch):
**Version:** v2.0

- [x] **Version:** Does the release version given match the repository release (v1.2)?

- [x] **Version:** Does the release version given match the repository release (v1.2)?
