+++ requirements.txt
@@ -1,10 +1,51 @@
-# Requirements for the demo notebooks
-# Useful for MyBinder configuration
-pandas
-numpy
-datascience
-folium
-matplotlib
-nbinteract
-scipy
-jupyter-book
+appnope==0.1.0
+atomicwrites==1.3.0
+attrs==19.3.0
+backcall==0.1.0
+bleach==3.1.0
+decorator==4.4.1
+defusedxml==0.6.0
+entrypoints==0.3
+importlib-metadata==0.23
+ipykernel==5.1.3
+ipython==7.9.0
+ipython-genutils==0.2.0
+jedi==0.15.1
+Jinja2==2.10.3
+jsonschema==3.1.1
+jupyter-book==0.6.3
+jupyter-client==5.3.4
+jupyter-core==4.6.1
+jupytext==1.2.4
+libsass==0.19.4
+MarkupSafe==1.1.1
+mistune==0.8.4
+more-itertools==7.2.0
+nbconvert==5.6.1
+nbformat==4.4.0
+packaging==19.2
+pandocfilters==1.4.2
+parso==0.5.1
+pexpect==4.7.0
+pickleshare==0.7.5
+pluggy==0.13.0
+prompt-toolkit==2.0.10
+ptyprocess==0.6.0
+py==1.8.0
+Pygments==2.4.2
+pyparsing==2.4.2
+pyrsistent==0.15.5
+pytest==5.2.2
+python-dateutil==2.8.1
+PyYAML==5.1.2
+pyzmq==18.1.0
+ruamel.yaml==0.16.5
+ruamel.yaml.clib==0.2.0
+six==1.12.0
+testpath==0.4.4
+tornado==6.0.3
+tqdm==4.37.0
+traitlets==4.3.3
+wcwidth==0.1.7
+webencodings==0.5.1
+zipp==0.6.0
