### 2020-10-08 094af9e
Bump bleach from 3.1.0 to 3.1.4
Bumps [bleach](https://github.com/mozilla/bleach) from 3.1.0 to 3.1.4.
- [Release notes](https://github.com/mozilla/bleach/releases)
- [Changelog](https://github.com/mozilla/bleach/blob/master/CHANGES)
- [Commits](https://github.com/mozilla/bleach/compare/v3.1.0...v3.1.4)

Signed-off-by: dependabot[bot] <support@github.com>
