+++ requirements.txt
@@ -2,7 +2,7 @@ appnope==0.1.0
 atomicwrites==1.3.0
 attrs==19.3.0
 backcall==0.1.0
-bleach==3.1.0
+bleach==3.1.4
 decorator==4.4.1
 defusedxml==0.6.0
 entrypoints==0.3
