+++ paper.md
@@ -1,75 +1,0 @@
----
-title: 'A practical guide to climate econometrics: Navigating key decision points in weather and climate data analysis'
-tags:
-  - climate change
-  - econometrics
-  - tutorial
-authors:
-  - name: James A. Rising
-    orcid: 0000-0001-8514-4748
-    affiliation: 1
-  - name: Azhar Hussain
-    orcid: 0000-0002-6475-1052
-    affiliation: 2
-  - name: Kevin Schwarzwald
-    orcid: 0000-0001-8309-7124
-    affiliation: 3
-  - name: Ana Trisovic
-    orcid: 0000-0003-1991-0533
-    affiliation: 4
-affiliations:
- - name: Grantham Research Institute, London School of Economics
-   index: 1
- - name: Department of Economics, London School of Economics
-   index: 2
- - name: Department of Earth and Environmental Sciences, Columbia University
-   index: 3
- - name: The Institute for Quantitative Social Science, Harvard University
-   index: 4
-date: 1 March 2020
-bibliography: paper.bib
----
-
-# Summary
-
-The use of econometrics to study how social, economic, and biophysical systems respond to weather has started a torrent of new research [@carleton2016social]. It is allowing researchers to better understand the impacts of climate change, disaster risk and responses, resource management, human behavior, and sustainable development.
-
-The research often features weather regression modeling and analyzing socioeconomic outcomes as they vary over time. The researchers need to collect, integrate, and analyze many different datasets such as historical climate data, future climate models, GIS information, and administrative data, in order to obtain meaningful and comprehensive results. While the challenges of climate econometrics are similar across all problems, the appropriate use of data, its integration and aggregation is unique to each research question, thereby presenting a stumbling block to early-career researchers. No single sequence of steps can be widely applied for resolving these issues. Similarly, software aimed at facilitating steps in this process needs to be fully understood and operated with care. 
-
-In light of this, we present a new tutorial that offers step-by-step guidance on carrying out a climate econometric analysis. Moreover, it features reusable code snippets in multiple programming languages and points to several pre-existing packages while providing some introductory comments on their appropriate use. The tutorial is structured into the following sections:
-
-1. Introduction: explains the prerequisites, introduces definitions and conventions, and recommends relevant literature.
-2. Using weather and climate data: introduces the data and its attributes, the NetCDF file format, supported programming languages, and common limitations of the data.
-3. Developing a reduced-form specification: provides a number of considerations before starting an analysis, such as choosing weather variables, functions for creating a model, and caveats when working with spatial and temporal processes.
-4. Weighting schemes: explains the importance of weighting schemes and how to work with them based on their file formats and origins. 
-Generating geographical unit data: introduces geographic information systems, shapefiles, and how to work with them effectively.
-5. Suggestions for work organization: presents recommendations for organizing research, including naming conventions, version control, and automation, which can improve researchers' productivity and research reproducibility.
-6. Contributions: introduces the current tutorial contributors and invites authors to contribute more information.
-
-The audiences for this tutorial are researchers and students trained in econometrics and experienced in at least one scientific programming language, such as Stata, R, Matlab, Julia, or Python. In addition to its use in a teaching event or for self-study, this tutorial can also be used as a reference manual, as each of its topics and lessons can be comprehended independently.
-
-Over the past year, the tutorial has been actively used by Ph.D. and Master's students at the London School of Economics, and it has been highlighted at multiple workshops. Several independent research projects have been developed by following the steps of the tutorial, which provides additional evidence for the success of its learning objectives.
-
[…]
