+++ _config.yml
@@ -1,151 +1,0 @@
-# Welcome to Jekyll!
-#
-# This config file is meant for settings that affect your whole blog, values
-# which you are expected to set up once and rarely edit after that. If you find
-# yourself editing this file very often, consider using Jekyll's data files
-# feature for the data you need to update frequently.
-#
-# For technical reasons, this file is *NOT* reloaded automatically when you use
-# 'bundle exec jekyll serve'. If you change this file, please restart the server process.
-
-# Site settings
-# These are used to personalize your new site. If you look in the HTML files,
-# you will see them accessed via {{ site.title }}, {{ site.email }}, and so on.
-# You can create any custom variable you would like, and they will be accessible
-# in the templates via {{ site.myvariable }}.
-
-#######################################################################################
-# Jekyll site settings
-title: Weather Panel Tutorial 
-author: Azhar Hussain, James Rising, Kevin Schwarzwald, Ana Trisovic
-email: YOUR EMAIL
-description: >-
-  This is an example book built with Jupyter Books.
-
-baseurl: /  # the subpath of your site, e.g. /blog. If there is no subpath for your site, use an empty string ""
-url: YOUR URL                  # the base hostname & protocol for your site, e.g. http://example.com
-
-
-#######################################################################################
-# Jupyter Book settings
-
-# Main page settings
-# footer_text: This page was created by <a href="http://www.impactlab.org">Climate Impact Lab</a>
-
-# Sidebar settings
-show_sidebar: true                # Show the sidebar. Only set to false if your only wish to host a single page.
-collapse_inactive_chapters: true  # Whether to collapse the inactive chapters in the sidebar
-textbook_logo: images/climate_estimate_logo.png               # A logo to be displayed at the top of your textbook sidebar. Should be square
-textbook_logo_link: https://climateestimate.net                   # A link for the logo.
-sidebar_footer_text: Powered by <a href="https://jupyterbook.org">Jupyter Book</a>
-number_toc_chapters: true         # Whether to add numbers to chapterse in your Table of Contents. If true, you can control this at the Chapter level in _data/toc.yml
-
-# Search settings
-search_max_words_in_content: 100   # In the search function, use at most this many words (too many words will make search slow)
-
-#######################################################################################
-# Interact link settings
-
-# General interact settings
-use_jupyterlab: false                     # If 'true', interact links will use JupyterLab as the interface
-
-# Jupyterhub link settings
-use_jupyterhub_button: false              # If 'true', display a button that will direct users to a JupyterHub (that you provide)
-jupyterhub_url: ''                     # The URL for your JupyterHub. If no URL, use ""
-jupyterhub_interact_text: Interact             # The text that interact buttons will contain.
-
-# Binder link settings
-use_binder_button: false                 # If 'true', add a binder button for interactive links
-binderhub_url: https://mybinder.org                        # The URL for your BinderHub. If no URL, use ""
-binder_repo_base: https://github.com/                     # The site on which the textbook repository is hosted
-binder_repo_org: YOUR-ORG                     # The username or organization that owns this repository
-binder_repo_name: YOUR-REPO                        # The name of the repository on the web
-binder_repo_branch: gh-pages                   # The branch on which your textbook is hosted.
-binderhub_interact_text: Interact              # The text that interact buttons will contain.
-
-# Thebelab settings
-use_thebelab_button: true                # If 'true', display a button to allow in-page running code cells with Thebelab
-thebelab_button_text: Thebelab                 # The text to display inside the Thebelab initialization button
-codemirror_theme: abcdef                     # Theme for codemirror cells, for options see https://codemirror.net/doc/manual.html#config
-
-# nbinteract settings
[…]
