### 2020-10-28 #90 @atrisovic
Hi @jwagemann, how are you doing? 👋  I had the same problem when I was reviewing another submission, but I somehow solved it by making a change in the URL. I don't exactly remember what I changed 😄 but [maybe this thread ](https://stackoverflow.com/questions/34020327/cant-find-github-invitation-to-join-repo)could help.
