### 2020-10-28 #90 @atrisovic
I think for me it worked when I viewed invitations from my profile, ie, something along the lines: `https://github.com/<username>/<reponame>/invitations`  🤔
