### 2020-10-28 #90 @jwagemann
It would be something like this in my case

`https://github.com/jwagemann/openjournals/jose-reviews/invitations`? This brings me to a 404.
