### 2020-11-02 atrisovic/weather-panel.github.io#37 @jwagemann
Review JOSE submission - General comments
- adjust title to e.g. ClimateEstimate.net - Weather Panel is misleading
- have a native speaker and someone external read through the tutorial text - sometimes the sentences are a bit cryptic and hard to follow
- Make use of coloured boxes to introduce content such as ‘Nice to know’ , ‘Note’ or ‘Hint’
- since it is a Jupyterbook - it would be actually very nice to have the examples in for of Jupyter notebooks available where learners can try out the examples while following the tutorial
- add to each section an ‘Additional resources’ section and add there in bullet points all additional external content you are referring to
- headings are often quite mis-leading, e.g. when you introduce shapefiles, the header of the section is called ‘geographical information systems’
- from the beginning till the end, it is not clear to me what each section aims to achieve. Examples are not easy to follow and I don’t believe that anyone apart from the authors could pick up the tutorial and can teach with it. Make yourself clear, what aim each section has and only add content that helps to meet the aim of the section
