### 2020-11-02 atrisovic/weather-panel.github.io#38 @jwagemann
Review JOSE submission - Comments to Introduction section
- Adjust the Introduction of the tutorial more with the learning objectives described in the paper
- Differentiate sections clearer: e.g. introduce a section for audience, learning objectives, how to use this tutorial, etc.
- Also telling the story - at the moment, the image is introduced, but not explained at all. If you include the image, then also highlight some of the most important findings and integrate the challenges / difficulties associated with this type of research and the reason why one needs this tutorial.
