### 2020-11-02 atrisovic/weather-panel.github.io#39 @jwagemann
Review JOSE submission - Weather and Climate Data
- make the description of the section more explaining. You introduce a learner to different weather and climate data and also the netcdf format, but a learner without pre-knowledge will have a hard time to understand what you intend to do
- instead of saying almost all weather and climate data are released in netcdf, I would say that a common data format used with weather and climate data is netCDF.
- I don’t understand this sentence: ‘Through this section, when possible relevant commands for working with netCDF files are listed for:’
- include coding examples as ‘code blocks’ instead of integrating them into the sentence or in a table? This makes it easier for a learner to simply copy paste some examples
- instead of matrices, maybe call them arrays?
- The ‘NetCDF content’ description is very abstract for someone who never used netCDF data before. Maybe show the content based on an example data set?
- CMIP data is all of a sudden introduced which might confuse a reader, also don’t think that the naming convention is already appropriate in the place where it is introduced. I would add this when the data is introduced as a ‘Hint’ or ‘Note’ section
- Diagnostic Maps of Climate data
   - this section feels it comes out of the blue and I doubt learners can really follow, because the section before did not introduce how to read the data properly
- 1.2 Gridded data
  - rephrase ‘Weather stations are imperfect’ and explain more in detail and more accurately the challenges with data coming from weather stations
- add hyperlinks to data documentations and data organisations
- cross-reference and link content within the tutorial (not only say ‘see content below without linking to it’)
- too many aspects are touched - e.g. location information is mentioned, but then only global climate data is introduced
- What is the main learning outcome this module shall have?
- Proposal to restructure
  - Introduction to netCDF
  - Overview of weather and climate data
  - Example 1: BEST
    - Description
    - Where to find
    - How to download
  - Example 2: CHIRPS
    - Description
    - Where to find
    - How to download
  - Example 3: ERA-5
    - Description
    - Where to find
    - How to download
  - ...
