### 2020-11-02 atrisovic/weather-panel.github.io#40 @jwagemann
Review JOSE submission - Developing a reduced-form specification
- Define ‘weather-panel data’
- Also a short definition what a reduced-form specification is
- Under precipitation - what does non-normal mean?
- General section ‘Choosing weather variables’ is not about choosing the variables but more a list of possible variables
- how does this section relate to developing a reduced-form specification?
- In general - this section would be more descriptive, if you tell the story with accompanying images, e.g. what is the problem and why do we need transformations and what do the different solutions do?
- I further recommend to show actual coding examples of the different transformations based on the chosen programming language (e.g. Python or R)
- Instead of using an image from Carleton et al. (2019), you can show the different outcomes based on the example provided
- the same also for section 2.3, show an actual example how cross-validation is done
