### 2020-11-02 atrisovic/weather-panel.github.io#41 @jwagemann
Review JOSE submission - Weighting Schemes
- what do you refer to with ‘first case’ and ‘second case’?
- where do weighting data files come from? Which files are these?
- I don’t understand the ‘Kinds of weight schemes and data sources’ section
- The example comes somehow out of the blue and it is difficult to follow what you are actually trying to explain
- it would be helpful to define what the authors mean with weighting schemes
- the topic with projections is introduced with a comic and then not elaborated further which projection is best → hard to understand for a learner
