### 2020-11-02 atrisovic/weather-panel.github.io#42 @jwagemann
Review JOSE submission - Generate geographical unit data
- maybe use a table to describe administrative units
- the section does not tell me why it is important to generate geographical unit data
- here examples of different programming languages are shown again and I’d say it would be better to stick to one example in one programming language and to refer to packages in other programming languages
