### 2020-11-02 atrisovic/weather-panel.github.io#43 @jwagemann
Review JOSE submission - Suggestion for work organisation
- do not only provide the comic without explanation for reasons why Git is good, provide a list of bullet points yourself why you use git and how it benefits projects
- 5.4 workflow automation
    - if you make this as a part of your tutorial, use an actual example and show how the workflow automation can be implemented and also list advantages why one would want to do so
