### 2020-11-02 #90 @jwagemann
I thank the authors for their time to develop the tutorial **‘ClimateEstimate.net - A tutorial on climate econometrics’** and giving me the opportunity to review it. The tutorial claims to be a ‘step-by-step guidance on carrying out a climatic econometric analysis’. My recommendation is to `accept with major revisions` for the following reasons:

- at the current state, it fails to meet the aim to be a step-by-step guide on how to carry out climatic econometrics. Sections often seem arbitrary and for a learner or external person it is often hard to understand what the section aims to describe or to do
- Many parts are not easy to follow, as authors try to capture too many aspects. Also many headlines seem a bit out of place, e.g the headline ‘Geographic Information Systems’ is used when the shapefile format is explained
- at the current state, the tutorial cannot be used by anyone other than the authors in order to teach students on the important topic of climatic econometrics
- The smart use of images and explanatory media is missing. The tutorial is heavily based on text and hardly works with additional images that help to understand the complex content
- The general writing needs to be spell-checked and the general tone feels somehow a bit sloppy: e.g. _‘here are some suggestions for dealing with the mess that is political geography’_ . I would not use the word ‘_mess’_ in any type of tutorial.
- The tutorial does not follow a consistent instructional design. Some sections are accompanied with code examples, others explain the content only in a very abstract way.
- Coding examples are either in R or try to show examples in multiple programming languages, which is quite confusing for the learner. I recommend to focus only on one programming language and provide only links to similar packages in other programming languages
- The tutorial misses an accompanying coding script in one of the programming languages. Without a coding script, it is very hard that a learner will be able to implement the content explained

I opened a series of issues with recommendations / comments for each section in the tutorial:
- [General comments](https://github.com/atrisovic/weather-panel.github.io/issues/37)
- [Comments to _Introduction_ section](https://github.com/atrisovic/weather-panel.github.io/issues/38)
- [Comments to _Weather and Climate Data_ section](https://github.com/atrisovic/weather-panel.github.io/issues/39)
- [Comments to _Developing a reduced-form specification_ section](https://github.com/atrisovic/weather-panel.github.io/issues/40)
- [Comments to _Weighting Schemes_ section](https://github.com/atrisovic/weather-panel.github.io/issues/41)
- [Comments to _Geographical Unit Data_ section](https://github.com/atrisovic/weather-panel.github.io/issues/42)
- [Comments to _Work Organisation_ section](https://github.com/atrisovic/weather-panel.github.io/issues/43)

I am happy to review a new submission.
