+++ requirements.txt
@@ -1,51 +1,97 @@
+alabaster==0.7.12
 appnope==0.1.0
-atomicwrites==1.3.0
-attrs==19.3.0
-backcall==0.1.0
-bleach==3.1.4
-decorator==4.4.1
+argon2-cffi==20.1.0
+async-generator==1.10
+attrs==20.3.0
+Babel==2.8.1
+backcall==0.2.0
+beautifulsoup4==4.9.3
+bleach==3.2.1
+certifi==2020.11.8
+cffi==1.14.3
+chardet==3.0.4
+click==7.1.2
+colorama==0.4.4
+decorator==4.4.2
 defusedxml==0.6.0
+docutils==0.16
 entrypoints==0.3
-importlib-metadata==0.23
-ipykernel==5.1.3
-ipython==7.9.0
+gitdb==4.0.5
+GitPython==3.1.11
+idna==2.10
+imagesize==1.2.0
+importlib-metadata==2.0.0
+ipykernel==5.3.4
+ipython==7.19.0
 ipython-genutils==0.2.0
-jedi==0.15.1
-Jinja2==2.10.3
-jsonschema==3.1.1
-jupyter-book==0.6.3
-jupyter-client==5.3.4
-jupyter-core==4.6.1
-jupytext==1.2.4
-libsass==0.19.4
+ipywidgets==7.5.1
+jedi==0.17.2
+Jinja2==2.11.2
+jsonschema==3.2.0
+jupyter-book==0.8.3
+jupyter-cache==0.4.1
+jupyter-client==6.1.7
+jupyter-core==4.6.3
+jupyter-sphinx==0.3.2
+latexcodec==2.0.1
+markdown-it-py==0.5.6
 MarkupSafe==1.1.1
 mistune==0.8.4
-more-itertools==7.2.0
+myst-nb==0.10.1
+myst-parser==0.12.10
+nbclient==0.5.1
 nbconvert==5.6.1
-nbformat==4.4.0
-packaging==19.2
-pandocfilters==1.4.2
-parso==0.5.1
-pexpect==4.7.0
+nbdime==2.1.0
+nbformat==5.0.8
+nest-asyncio==1.4.3
+notebook==6.1.5
+oset==0.1.3
+packaging==20.4
+pandocfilters==1.4.3
+parso==0.7.1
+pexpect==4.8.0
 pickleshare==0.7.5
-pluggy==0.13.0
-prompt-toolkit==2.0.10
+prometheus-client==0.8.0
+prompt-toolkit==3.0.8
 ptyprocess==0.6.0
-py==1.8.0
-Pygments==2.4.2
-pyparsing==2.4.2
-pyrsistent==0.15.5
-pytest==5.2.2
+pybtex==0.23.0
+pybtex-docutils==0.2.2
+pycparser==2.20
+pydata-sphinx-theme==0.4.1
+Pygments==2.7.2
+pyparsing==2.4.7
+pyrsistent==0.17.3
 python-dateutil==2.8.1
-PyYAML==5.1.2
-pyzmq==18.1.0
-ruamel.yaml==0.16.5
-ruamel.yaml.clib==0.2.0
-six==1.12.0
+pytz==2020.4
+PyYAML==5.3.1
+pyzmq==19.0.2
+requests==2.25.0
+Send2Trash==1.5.0
+six==1.15.0
+smmap==3.0.4
+snowballstemmer==2.0.0
+soupsieve==2.0.1
+Sphinx==3.3.0
+sphinx-book-theme==0.0.39
+sphinx-comments==0.0.3
+sphinx-copybutton==0.3.1
+sphinx-panels==0.5.2
+sphinx-thebe==0.0.8
[…]
