+++ .github/workflows/pythonpackage.yml
@@ -1,4 +1,4 @@
-name: Python package
+name: deploy-book 
 
 on: 
   push:
@@ -6,45 +6,27 @@ on:
       - master
 
 jobs:
-  build:
-
+  deploy-book:
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.7]
-
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
+
+    - name: Set up Python 3.7 
       uses: actions/setup-python@v1
       with:
-        python-version: ${{ matrix.python-version }}
+        python-version: 3.7 
+
     - name: Install dependencies
       run: |
-        sudo apt-get install python3-pip
         pip install -r requirements.txt
+
     - name: Build Jupyter Book 
       run: |
-        #pip install 'jupyter-book==0.6.3'
-        PATH="${PATH}:${HOME}/.local/bin"
-        ls -ls
-        jupyter-book build .
-    - name: Install Ruby dependencies
-      run: |
-        sudo gem install bundler -v 1.17.2
-        bundle install
-        rm -rf _build
-        make site
-    - name: GitHub Pages Step-by-step
-      run: |
-        sudo chown -R $(whoami):$(whoami) .
-        git config --global user.email "$GITHUB_ACTOR@users.noreply.github.com"
-        git config --global user.name "$GITHUB_ACTOR"
-        cp -r _site /tmp
-        rm -rf *
-        mv /tmp/_site/* .
-        git checkout -b gh-pages
-        git add --all --force
-        git commit -a --allow-empty-message -m ''
-        git remote set-url origin "https://$GITHUB_ACTOR:${{ secrets.GITHUB_TOKEN }}@github.com/$GITHUB_REPOSITORY"
-        git push --force origin gh-pages 
+        jupyter-book build tutorial-content/ 
+
+    - name: Push to GitHub Pages
+      uses: peaceiris/actions-gh-pages@v3
+      with: 
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        publish_dir: ./tutorial-content/_build/html
+        cname: climateestimate.net
