+++ paper.bib
@@ -1,53 +1,0 @@
-@article{carleton2016social,
-  title={Social and economic impacts of climate},
-  author={Carleton, Tamma A and Hsiang, Solomon M},
-  journal={Science},
-  volume={353},
-  number={6304},
-  pages={aad9837},
-  year={2016},
-  publisher={American Association for the Advancement of Science},
-  doi={10.1126/science.aad9837}
-}
-@article{hsiang2018economist,
-  title={An economist's guide to climate change science},
-  author={Hsiang, Solomon and Kopp, Robert E},
-  journal={Journal of Economic Perspectives},
-  volume={32},
-  number={4},
-  pages={3--32},
-  year={2018},
-  doi={10.3386/w25189}
-}
-@article{hsiang2016climate,
-  title={Climate econometrics},
-  author={Hsiang, Solomon},
-  journal={Annual Review of Resource Economics},
-  volume={8},
-  pages={43--75},
-  year={2016},
-  publisher={Annual Reviews},
-  doi={10.3386/w22181}
-}
-@article{ciscar2019assessing,
-  title={Assessing future climate change impacts in the EU and the USA: insights and lessons from two continental-scale projects},
-  author={Ciscar, Juan-Carlos and Rising, James and Kopp, Robert E and Feyen, Luc},
-  journal={Environmental Research Letters},
-  volume={14},
-  number={8},
-  pages={084010},
-  year={2019},
-  publisher={IOP Publishing},
-  doi={10.1088/1748-9326/ab281e}
-}
-@article{nissan2019use,
-  title={On the use and misuse of climate change projections in international development},
-  author={Nissan, Hannah and Goddard, Lisa and de Perez, Erin Coughlan and Furlow, John and Baethgen, Walter and Thomson, Madeleine C and Mason, Simon J},
-  journal={Wiley Interdisciplinary Reviews: Climate Change},
-  volume={10},
-  number={3},
-  pages={e579},
-  year={2019},
-  publisher={Wiley Online Library},
-  doi={10.1002/wcc.579}
-}
