+++ tutorial-content/content/working-with-shapefiles.md
@@ -1,0 +1,169 @@
+# Working with shapefiles
+
+Your first step should be to view your shapefile in a software system
+like QGIS or ArcGIS.
+QGIS is a free and open-source desktop geographic information system application that supports viewing, editing and analysis of geospatial data. ArcGIS is proprietary software for working with maps and geographic data.
+
+These systems also allow you to perform many kinds of spatial
+analysis. Some of the most commonly useful are calculating Zonal
+Statistics, such as the average of a gridded dataset across each
+shapefile region, and Merging units, and translating points to
+polygons with Voronoi Polygons.
+
+ Shapefiles can be opened with Python packages like the following
+
+```python
+import fiona
+shape = fiona.open("my_shapefile.shp")
+print shape.schema
+{'geometry': 'LineString', \ 
+'properties': OrderedDict([(u'FID', 'float:11')])}
+
+# or
+
+import shapefile
+shape = shapefile.Reader("my_shapefile.shp")
+
+# or
+
+import geopandas as gpd
+shapefile = gpd.read_file("my_shapefile.shp")
+print(shapefile)
+```
+
+Data analysis software R also supports working with spatial data. To
+read shape files you could use a package like `maptools`,  `rgdal`,
+`sf`, or `PBSmapping`.
+
+```R
+library(maptools)
+shapefile <- readShapePoly("/my_shapefile.shp")
+
+# or
+
+library(PBSmapping)
+shapefile <- importShapefile("/my_shapefile.shp")
+```
+
+## Constructing averages within spatial units
+
+Now, you probably have a gridded spatiotemporal dataset of historical
+weather and economic output data specific to shapefile regions.  The
+next step is construct the weather measures that will correspond to
+each of your economic data observations.  To do this, you will need to
+construct a weighted average of the weather in each region for each
+timestep.
+
+In some cases, there are tools available that will help you do
+this. If you are using area weighting (i.e., no weighting grid) and
+your grid is fine enough that every region fully contains at least one
+cell, one tool you can use is
+[regionmask](http://www.matteodefelice.name/post/aggregating-gridded-data/).
+
+If your situation is more complicated, or if you just want to know how
+to do it yourself, it is important to set up the mathematical process
+efficiently since this can be a computationally expensive step.
+
+The regional averaging process is equivalent to a matrix
+transformation: $$w_\text{region} = A w_\text{gridded}$$
+
+where $w_\text{region}$ is a vector of weather values across
+regions, in a given timestep; and $w_\text{gridded}$ is a vector of
+weather values across grid cells.  Suppose there are $N$ regions and
+$R C$ grid cells, then the transformation matrix $A$ will be $N x
+R C$. The $A$ matrix does not change over time, so once you
+calculate it, the process for generating each time step is faster.
+
+Below, we sketch out two approaches to generating this matrix, but a
+few comments are common no matter how you generate it.
+
+1. The sum of entries across each row should be 1. Missing values can
+   cause reasonable-looking calculations to produce rows sums that are
+   less than one, so make sure you check.
+   
+2. This matrix is huge, but it is very sparse: most entries
+   are 0. Make sure to use a sparse matrix implementation (e.g.,
+   `sparseMatrix` in R, `scipy.sparse` in python, `sparse` in Matlab).
+   
+3. The $w_\text{gridded}$ data starts as a matrix, but here we use
+   it as a vector.  It is easy (in any language) to convert a matrix
+   to a vector with all of its values, but you need to be careful
+   about the order of the entries, and order the columns of $A$ the
+   same way.
+   
+   In R, `as.vector` will convert from a matrix to a vector, with each
+   *column* being listed in full before moving on to the next column.
+   
+   In python, `numpy.flatten` will convert a numpy matrix to a vector,
+   with each *row* being listed in full before moving on to the next
[…]
