+++ tutorial-content/_toc.yml
@@ -1,7 +1,7 @@
 - file: content/getting-started
-  title: Introduction to the Tutorial
+  title: Introduction 
 
-- part: Using Weather and Climate Data
+- part: Weather and Climate Data
   chapters:
   - file: content/weather-and-climate-data
 
@@ -13,9 +13,10 @@
   chapters:
   - file: content/weighting-schemes
 
-- part:  Generating geographical unit data
+- part:  Geographical unit data
   chapters:
   - file: content/geographical-unit-data
+  - file: content/working-with-shapefiles
 
 - part: Suggestions for work organization 
   chapters:
