+++ _sources/content/geographical-unit-data.md
@@ -1,6 +1,4 @@
-# 4. Generating geographical unit data
-
-## 4.1 Introduction
+# Geographical unit data
 
 A geographical unit, area or region, is a portion of a country or other region delineated for the purpose of administration, and as such, it is a common unit for recording economic outcome data.  These geographic regions may be defined in subtle, non-intuitive ways.  For example, a “[city](https://ec.europa.eu/eurostat/web/cities/spatial-units)” is a local administrative unit where the majority of the population lives in an urban center, while the “[greater city](https://ec.europa.eu/eurostat/web/cities/spatial-units)” is an approximation of the urban center beyond the administrative city boundaries.
 
@@ -19,10 +17,11 @@ territories with homogeneous features that are relevant to the
 study. For example, if the weather is relevant for the study, the
 administrative unit should be fairly homogeneous concerning mean temperature and precipitation.
 
-[Using Weather Data and Climate Model Output in Economic Analyses of Climate Change](https://academic.oup.com/reep/article/7/2/181/1522753) describes
-the common pitfalls in translating weather data into geographical unit data.
+```{seealso}
+[Using Weather Data and Climate Model Output in Economic Analyses of Climate Change](https://academic.oup.com/reep/article/7/2/181/1522753) describes the common pitfalls in translating weather data into geographical unit data.
+```
 
-## 4.2 Geographic information systems
+## Geographic information systems
 
 Much is made of geographic information system (GIS), but these have
 long since become unnotably common parts of spatial research. GIS
@@ -38,15 +37,19 @@ using GIS software like QGIS.
 
 A [shapefile](https://www.esri.com/library/whitepapers/pdfs/shapefile.pdf) also stores geometry and attribute information for the spatial features in a data set. The geometry for a feature is stored as a shape comprising a vector of coordinates. Area features are represented as a closed-loop, double-digitized polygons. The shapes together with data attributes linked to each shape create the representation of geographic data like countries, rivers and lakes.
 
+
+````{note}
 Despite its name indicating a singular file, a shapefile is actually a collection of at least three basic files that need to be stored in the same directory to be used. The three mandatory files have filename extensions `.shp`, `.shx` and `.dbf`. There may be additional files like `.prj` with the shape file’s projection information. All files must have the same name, for example:
 
-> `states.shp` 
+```
+   states.shp
+   states.shx
+   states.dbf
+```
+````
 
-> `states.shx` 
 
-> `states.dbf` 
-
-## 4.3 Finding shapefiles
+## Finding shapefiles
 
 It is usually possible to find a shapefile corresponding to a set of
 socioeconomic data by searching around online. Many governments will
@@ -58,7 +61,7 @@ and [Natural Earth](http://www.naturalearthdata.com/) which has
 clean physical shape information.
 
 
-## 4.4 Creating shapefiles
+## Creating shapefiles
 
 In some cases, you will have to make the shapefile from scratch. The
 most common ways of doing this is by clicking out the shape of each
@@ -73,182 +76,4 @@ polygon. This generally requires a few steps.
    [GDAL Georeferencer](https://www.qgistutorials.com/en/docs/georeferencing_basics.html) to
    make this a point-and-click task.
 3. Create new polygons by clicking around the edges of the polygon.
-
-
-## 4.5 Working with shapefiles
-
-Your first step should be to view your shapefile in a software system
-like QGIS or ArcGIS.
-QGIS is a free and open-source desktop geographic information system application that supports viewing, editing and analysis of geospatial data. ArcGIS is proprietary software for working with maps and geographic data.
-
-These systems also allow you to perform many kinds of spatial
[…]
