+++ tutorial-content/content/reduced-form-specification.md
@@ -1,19 +1,17 @@
-# 2. Developing a reduced-form specification
+# Developing a reduced-form specification
 
 This section describes some of the considerations that go into
 developing a reduced-form specification using weather panel data. Our
 discussion here will be very practical and limited to nonlinear panel
-regressions, so also
-read
-[Estimating the Economic Impacts of Climate Change Using Weather Observations](https://academic.oup.com/reep/article/14/1/1/5718007) for
-a review of the available econometric techniques and their strengths
-and weaknesses.  For an extensive review of the results available from
-the climate econometric literature and the empirical methods used to
-identify them, a good resource
-is
-[Social and Economic Impacts of Climate](http://science.sciencemag.org/content/353/6304/aad9837).
+regressions. 
 
-## 2.1. Choosing weather variables
+```{seealso}
+See also [Estimating the Economic Impacts of Climate Change Using Weather Observations](https://academic.oup.com/reep/article/14/1/1/5718007) for a review of the available econometric techniques and their strengths and weaknesses.  For an extensive review of the results available from
+the climate econometric literature and the empirical methods used to
+identify them, a good resource is [Social and Economic Impacts of Climate](http://science.sciencemag.org/content/353/6304/aad9837).
+```
+
+## Choosing weather variables
 
 The choice of weather variables depends on the question we are trying
 to answer, and there are many forms to represent any given
@@ -98,7 +96,7 @@ Some datasets (such as [HadEX2](https://climatedataguide.ucar.edu/climate-data/h
          throughout East Africa, South and Southeast Asia, and
          Oceania.
 
-## 2.2. Dealing with the spatial and temporal scales of economic processes
+## Spatial and temporal scales of economic processes
 
 The process of developing a reduced-form specification starts with a
 study of the "true model", or data-generating process, that relates
@@ -204,152 +202,3 @@ observed for region $i$ in reporting period $t$. Weather data products can have 
 
 ![Humor](images/cartoon_sec2.JPG)
 
-## 2.3. Common functional forms (pros, cons, and methods)
-
-Returning to the "true model" of your process, the decisions around how to
-generate a linear expression that you can estimate have important
-implications. Different functional forms serve different purposes and
-describe different underlying processes. Some of the frequently used functional forms along with a good reference for understanding them in detail are listed below.
-
-- **[Bins](https://pubs.aeaweb.org/doi/pdfplus/10.1257/app.3.4.152)**
-
-Bins offer a very flexible functional form, although the choice of bin
-edges means that they are not as "non-parametric" as often assumed. It
-is also highly sensitive to existence of outliers in data. This can
-particularly be an issue for extrapolation when the results are used
-for projections: only the estimate for the extreme bins will be used.
-
-The bin widths should generally be even to facilitate easy
-understanding, although the lowest and highest bins can go to negative
-and positive infinity (or the appropriate analog). You may want to
-have smaller size bins around weather values where we expect most of
-the response to occur and where there is a lot of data, but be
-prepared to also show evenly-spaced bins.
-
-The interpretation of a binned result is in terms of the time unit
-used. For example, if daily temperatures are used, then the marginal
-effect for a given bin is the additional effect of "one more day" for
-which the temperature falls into that bin.
-
-To calculate bin observations, count up the number of timesteps where
[…]
