+++ _sources/content/functional-forms.md
@@ -1,0 +1,149 @@
+# Functional forms: pros, cons, and methods
+
+Returning to the "true model" of your process, the decisions around how to
+generate a linear expression that you can estimate have important
+implications. Different functional forms serve different purposes and
+describe different underlying processes. Some of the frequently used functional forms along with a good reference for understanding them in detail are listed below.
+
+## [Bins](https://pubs.aeaweb.org/doi/pdfplus/10.1257/app.3.4.152)
+
+Bins offer a very flexible functional form, although the choice of bin
+edges means that they are not as "non-parametric" as often assumed. It
+is also highly sensitive to existence of outliers in data. This can
+particularly be an issue for extrapolation when the results are used
+for projections: only the estimate for the extreme bins will be used.
+
+The bin widths should generally be even to facilitate easy
+understanding, although the lowest and highest bins can go to negative
+and positive infinity (or the appropriate analog). You may want to
+have smaller size bins around weather values where we expect most of
+the response to occur and where there is a lot of data, but be
+prepared to also show evenly-spaced bins.
+
+The interpretation of a binned result is in terms of the time unit
+used. For example, if daily temperatures are used, then the marginal
+effect for a given bin is the additional effect of "one more day" for
+which the temperature falls into that bin.
+
+To calculate bin observations, count up the number of timesteps where
+the weather predictor falls into each bin:
+    $$T_{it} = \sum_{p \in i} \psi_{p} \sum \mathbf{1} \left \{  {T_{p i t} \in Bin_k} \right \}$$
+    where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
+    
+## [Polynomial](https://en.wikipedia.org/wiki/Polynomial_regression)
+
+Polynomial specifications balance smooth dose-response curves with the
+flexibility of increasing the effective resolution of the curve by
+increasing the degree of the polynomial. Using more degrees improves
+the ability of the curve to fit the data, but may lead to
+over-fitting. To choose the number of terms in the polynomial,
+consider cross-validation.
+
+Another benefit of polynomials for climate change estimates is that
+they provide smooth extrapolation to higher temperatures. Again, it is
+important to highlight the fact that the evaluation of the
+dose-response curve at temperatures outside the bounds of the observed
+data reflects assumptions, rather than evidence. Cross-validation that
+leaves out the latest periods or most extreme temperatures can improve
+confidence in these assumptions.
+
+In calculating the predictor values for a polynomial, consider the
+scale of the data-generating process. If it is a local process, the
+high-resolution weather values should be raised to powers before
+aggregating up to the regions in the economic data. That is, the the
+predictor for the $k$-th power of the polynomial is 
+    $$f(T_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} T_{p i t}^k$$ 
+    where $\psi_{p}$ is the weight assigned to the $p$ gridcell.  
+    
+The dose-response regression would then be applied as follows
+    $$F(T_{it})=\sum_{k} \beta_k f(T_{it}^k)$$
+	
+while the coefficients can be interepted as describing a local
+dose-response relationship:
+    $$F(T_{pit})=\sum_{k} \beta_k T_{pit}^k$$
+
+## [Restricted cubic spline](https://support.sas.com/resources/papers/proceedings16/5621-2016.pdf)
+
+Restricted cubic splines produce smooth dose-response curves, like
+polynomials, but mitigate some of the problems that polynomials with
[…]
