+++ _sources/content/reduced-form-specification.md
@@ -1,19 +1,17 @@
-# 2. Developing a reduced-form specification
+# Developing a reduced-form specification
 
 This section describes some of the considerations that go into
 developing a reduced-form specification using weather panel data. Our
 discussion here will be very practical and limited to nonlinear panel
-regressions, so also
-read
-[Estimating the Economic Impacts of Climate Change Using Weather Observations](https://academic.oup.com/reep/article/14/1/1/5718007) for
-a review of the available econometric techniques and their strengths
-and weaknesses.  For an extensive review of the results available from
-the climate econometric literature and the empirical methods used to
-identify them, a good resource
-is
-[Social and Economic Impacts of Climate](http://science.sciencemag.org/content/353/6304/aad9837).
+regressions. 
 
-## 2.1. Choosing weather variables
+```{seealso}
+See also [Estimating the Economic Impacts of Climate Change Using Weather Observations](https://academic.oup.com/reep/article/14/1/1/5718007) for a review of the available econometric techniques and their strengths and weaknesses.  For an extensive review of the results available from
+the climate econometric literature and the empirical methods used to
+identify them, a good resource is [Social and Economic Impacts of Climate](http://science.sciencemag.org/content/353/6304/aad9837).
+```
+
+# Choosing weather variables
 
 The choice of weather variables depends on the question we are trying
 to answer, and there are many forms to represent any given
@@ -98,7 +96,7 @@ Some datasets (such as [HadEX2](https://climatedataguide.ucar.edu/climate-data/h
          throughout East Africa, South and Southeast Asia, and
          Oceania.
 
-## 2.2. Dealing with the spatial and temporal scales of economic processes
+# Spatial and temporal scales of economic processes
 
 The process of developing a reduced-form specification starts with a
 study of the "true model", or data-generating process, that relates
@@ -141,7 +139,7 @@ signal less that doing it for precipitation. This is because
 precipitation is highly local both temporally and spatially; it could
 rain for <1 min in <1 km radius area.
 
-### Transformation-before-aggregation
+## Transformation-before-aggregation
 
 When an economic process is occurring at the local level (for example,
 for individuals or households), we need to first do our estimation at
@@ -179,7 +177,7 @@ $$\beta_k (\sum_{s\in t} \sum_{p\in i} g_k(T_{ps}))$$
 That is, the variables used in the regression should be the sum over
 weather data that has been transformed at the grid level.
 
-### Aggregation-before-transformation
+## Aggregation-before-transformation
 
 When an economic process is occurring at the regional level, we need
 to first aggregate weather variable to that level before transforming
@@ -204,152 +202,3 @@ observed for region $i$ in reporting period $t$. Weather data products can have 
 
 ![Humor](images/cartoon_sec2.JPG)
 
-## 2.3. Common functional forms (pros, cons, and methods)
-
-Returning to the "true model" of your process, the decisions around how to
-generate a linear expression that you can estimate have important
-implications. Different functional forms serve different purposes and
-describe different underlying processes. Some of the frequently used functional forms along with a good reference for understanding them in detail are listed below.
-
-- **[Bins](https://pubs.aeaweb.org/doi/pdfplus/10.1257/app.3.4.152)**
-
-Bins offer a very flexible functional form, although the choice of bin
-edges means that they are not as "non-parametric" as often assumed. It
-is also highly sensitive to existence of outliers in data. This can
[…]

[…]
