+++ tutorial-content/content/geographical-unit-data.md
@@ -1,79 +1,80 @@
 # Generating geographical unit data
+ 
+```{admonition} Key objectives
+:class: note
+- Understand what geographical unit and administrative unit data are
+- Understand sharefiles, how to obtain them and how to work with them
+```
 
-A geographical unit, area or region, is a portion of a country or other region delineated for the purpose of administration, and as such, it is a common unit for recording economic outcome data.  These geographic regions may be defined in subtle, non-intuitive ways.  For example, a “[city](https://ec.europa.eu/eurostat/web/cities/spatial-units)” is a local administrative unit where the majority of the population lives in an urban center, while the “[greater city](https://ec.europa.eu/eurostat/web/cities/spatial-units)” is an approximation of the urban center beyond the administrative city boundaries.
+Socioeconomic data is collected and corresponding to geographical units, such as states, countries, provinces, or municipalities, which is a portion of a country delineated for administration, and hence, often called  "administrative units". For that reason, administrative units (or politically-defined regions) are frequently used in economic analysis rather than regular grids. Communicating data analysis results in administrative units is also particularly effective since politically defined regions are relevant for policy-makers. 
 
-One kind of geographic region is simply called an "administrative
-unit", and refers to states and provinces, or counties and
-municipalities. The top-level of this hierarchy of administrative
-units is "ADM0", referring to countries; "ADM1" is the first level of
-political division, usually called states or provinces; "ADM2" is the
-second level of division, and has a wider range of names across the
-globe.
- Administrative units in economics analyses are typically politically-defined regions, rather than regular grids, because socioeconomic data is collected and corresponding to the political regions. They can also be more effective for communicating results, since politically defined regions are also more relevant for policy-makers.
+The top-level of this hierarchy of administrative units is "ADM0", referring to countries; "ADM1" is the first level of political division, usually called states or provinces; "ADM2" is the second level of division, and has a wider range of names across the globe (see example below).
+ 
+```{list-table} Example of administrative units table for [Philippines](http://www.eki.ee/knab/adm2.htm)
+:header-rows: 1
+ 
+* - ADM0
+  - ADM1
+  - ADM2
+* - Philippines (PH)
+  - Ilocos, Iloko
+  - Ilocos Norte, Hilagang Iloko (ILN)
+* -
+  -
+  - Ilocos Sur, Timog Iloko (ILS)
+* -
+  -
+  - ...
+* - 
+  - Cagayan Valley, Lambak ng Kagayan
+  - Batanes (BTN)
+* -
+  -
+  - Cagayan, Kagayan (CAG)
+* -
+  -
+  - Isabela (ISA)
+* -
+  - ...
+  - ...
+```
+ 
+Administrative unit data can capture existing administrative units (high granularity) or groups of those units (lesser granularity). For example, the administrative unit database, [Global Administrative Regions](https://gadm.org), offers a granularity of 386,735 administrative areas for the entire world, that can be grouped according to the needs of a study.
 
- Administrative unit data can capture existing administrative units (high granularity) or groups of those units (lesser granularity). For example, the administrative unit database, [Global Administrative Regions](https://gadm.org), offers a granularity of 386,735 administrative areas for the entire world, that can be grouped according to the needs of a study.
-When aggregating administrative units, it is important to capture
-territories with homogeneous features that are relevant to the
-study. For example, if the weather is relevant for the study, the
-administrative unit should be fairly homogeneous concerning mean temperature and precipitation.
-
+When aggregating administrative units, it is important to capture territories with homogeneous features that are relevant to the study. For example, if the weather is relevant for the study, the administrative units should be fairly homogeneous concerning mean temperature and precipitation.
+ 
 ```{seealso}
[…]
