+++ tutorial-content/content/working-with-shapefiles.md
@@ -1,127 +1,109 @@
 # Working with shapefiles
 
-Your first step should be to view your shapefile in a software system
-like QGIS or ArcGIS.
-QGIS is a free and open-source desktop geographic information system application that supports viewing, editing and analysis of geospatial data. ArcGIS is proprietary software for working with maps and geographic data.
+Once you obtain your shapefiles, you should first view them in a software system like QGIS or ArcGIS to make sure everything is in order. Both R and Python support working with shapefiles and spatial data. See the following examples:
 
-These systems also allow you to perform many kinds of spatial
-analysis. Some of the most commonly useful are calculating Zonal
-Statistics, such as the average of a gridded dataset across each
-shapefile region, and Merging units, and translating points to
-polygons with Voronoi Polygons.
-
-Both R and Python support working with spatial data. See the following examples:
- 
 ````{tabbed} R
-To read shape files you could use a package like `maptools`,  `rgdal`, `sf`, or `PBSmapping`.
-
+To read shape files you could use a package like `maptools`, `rgdal`, `sf`, or `PBSmapping`.
+ 
 ```{code-block} R
 library(maptools)
 shapefile <- readShapePoly("/my_shapefile.shp")
-
+ 
 # or
-
+ 
 library(PBSmapping)
 shapefile <- importShapefile("/my_shapefile.shp")
 ```
 ````
-
+ 
 ````{tabbed} Python
-
+ 
 Shapefiles can be opened with Python packages in a few different ways:
-
+ 
 ```{code-block} python
 import fiona
 shape = fiona.open("my_shapefile.shp")
 print shape.schema
-{'geometry': 'LineString', \ 
+{'geometry': 'LineString', \
 'properties': OrderedDict([(u'FID', 'float:11')])}
-
+ 
 # or
-
+ 
 import shapefile
 shape = shapefile.Reader("my_shapefile.shp")
-
+ 
 # or
-
+ 
 import geopandas as gpd
 shapefile = gpd.read_file("my_shapefile.shp")
 print(shapefile)
 ```
 ````
 
-
 ## Constructing averages within spatial units
-
-Now, you probably have a gridded spatiotemporal dataset of historical
-weather and economic output data specific to shapefile regions.  The
-next step is construct the weather measures that will correspond to
-each of your economic data observations.  To do this, you will need to
-construct a weighted average of the weather in each region for each
-timestep.
-
-In some cases, there are tools available that will help you do
-this. If you are using area weighting (i.e., no weighting grid) and
-your grid is fine enough that every region fully contains at least one
-cell, one tool you can use is
+ 
+At this point, you probably have a gridded spatiotemporal dataset of historical weather and economic output data specific to shapefile regions. The next step is construct the weather measures that will correspond to each of your economic data observations. To do this, you will need to construct a weighted average of the weather in each region for each timestep.
+ 
+In some cases, there are tools available that will help you do this. If you are using area weighting (i.e., no weighting grid) and your grid is fine enough that every region fully contains at least one cell, one tool you can use is
 [regionmask](http://www.matteodefelice.name/post/aggregating-gridded-data/).
-
-If your situation is more complicated, or if you just want to know how
-to do it yourself, it is important to set up the mathematical process
-efficiently since this can be a computationally expensive step.
-
-The regional averaging process is equivalent to a matrix
-transformation: $$w_\text{region} = A w_\text{gridded}$$
-
+ 
+If your situation is more complicated, or if you just want to know how to do it yourself, it is important to set up the mathematical process efficiently since this can be a computationally intensive step.
+ 
+The regional averaging process is equivalent to a matrix transformation: 
+$$w_\text{region} = A w_\text{gridded}$$
+ 
[…]
