### 2020-11-30 325d8c6
edits to weighting schemes
