+++ tutorial-content/content/weighting-schemes.md
@@ -1,19 +1,30 @@
-# 3. Weighting schemes
+# Weighting schemes
 
-This section describes how to use different weighting schemes when aggregating gridded data to data regions.
+This section describes how to use different weighting schemes when
+working with gridded and regional data. As we use the term, weighting
+schemes assign a weight to each grid cell or regional
+observation. They have two major uses: (1) when running regressions,
+to weight observations to accurately estimate processes, or (2) to
+perform a weighted aggregation of gridded data to data regions.
 
-## 3.1 Why spatial weighting schemes matter
+## Why spatial weighting schemes matter
 
 Taking the unweighted average of weather within a region can misrepresent what populations, firms, or other phenomena of interest are exposed to. For example, an unweighted annual average temperature for Canada is about -8°C, but most of the population and agricultural activity is in climate zones with mean temperatures over 6°C, and the urban heat island effect can raise temperatures by another 4°C. The time of year matters too, and you should consider a weighting scheme across days within a year, or even hours within a day.
 
-As described in section [Dealing with the spatial and temporal scales of economic processes](reduced-form-specification#Dealing-with-the-spatial-and-temporal-scales-of-economic-processes), the scale of a phenomenon matters. Many processes occur at a more local scale than that which data is collected. The motivation for weighting is different for aggregation that represents averaged phenomena vs. phenomena that respond to averaged weather, and the sequence of analysis changes.
+As described in section
+[Dealing with the spatial and temporal scales of economic processes](reduced-form-specification#Dealing-with-the-spatial-and-temporal-scales-of-economic-processes),
+the scale of a phenomenon matters. Many processes occur at a more
+local scale than that which data is collected. The motivation for
+weighting is different for aggregation that represents averaged
+phenomena vs. phenomena that respond to averaged weather, and the
+sequence of analysis changes accordingly.
 
-In the first case, the phenomenon occurs locally, in response to local weather. In this case, we perform weighted aggregations to reflect the amount of the phenomenon in each location. For example, we would use population weighting to model the effects of heat on people. In this case, the order of operations is:
+When the phenomenon occurs locally, in response to local weather, we perform weighted aggregations to reflect the amount of the phenomenon in each location. For example, we would use population weighting to model the effects of heat on people. In this case, the order of operations is:
 
 1. Transform weather into the terms of the model specification.
 2. Average these transformed terms across space using a weighting scheme.
 
-In the second case, the phenomenon occurs at a data region level, in response to averaged weather. In this case, the weighting scheme reflects the relative importance of weather in different regions to the whole. For example, weighting rainfall by the distance from a shore could be important to predict the declaration of states of emergency. The order of operations is:
+When the phenomenon occurs at a data region level, in response to averaged weather, the weighting scheme reflects the relative importance of weather in different regions to the whole. For example, weighting rainfall by the distance from a shore could be important to predict the declaration of states of emergency. The order of operations is:
 
 1. Average the weather across space using a weighting scheme.
 2. Transform the averaged weather to the model specification.
@@ -24,14 +35,39 @@ $$T_{it} = \sum_{p \in P(i)} w_p T_{pt} \text{ such that } \sum_p w_{p \in P(i)}
 
 where $w_p$ is the weight for pixel $p$, and $P(i)$ is the set of pixels in data region $i$.
 
[…]
