+++ tutorial-content/content/geographical-unit-data.md
@@ -1,9 +1,15 @@
 # Generating geographical unit data
  
-```{admonition} Key objectives
+```{admonition} Key objectives and decision points
 :class: note
+
+Objectives:
 - Understand what geographical unit and administrative unit data are
 - Understand sharefiles, how to obtain them and how to work with them
+
+Decision points:
+- How to incorporate geographical and administrative units in your analysis?
+- How to generate weather measures that will correspond to your economic data and geographical regions?
 ```
 
 Socioeconomic data is collected and corresponding to geographical units, such as states, countries, provinces, or municipalities, which is a portion of a country delineated for administration, and hence, often called  "administrative units". For that reason, administrative units (or politically-defined regions) are frequently used in economic analysis rather than regular grids. Communicating data analysis results in administrative units is also particularly effective since politically defined regions are relevant for policy-makers. 
@@ -53,10 +59,9 @@ Geographic information system (GIS) is a computer system for capturing, storing,
 
 QGIS is a free and open-source desktop geographic information system application that supports viewing, editing and analysis of geospatial data. ArcGIS is proprietary software for working with maps and geographic data. These systems also allow you to perform many kinds of spatial analysis. Some of the most commonly useful are calculating Zonal Statistics, such as the average of a gridded dataset across each shapefile region, and Merging units, and translating points to polygons with Voronoi Polygons.
 
-Administrative units are often defined and described as shapefiles. A [shapefile](https://www.esri.com/library/whitepapers/pdfs/shapefile.pdf) stores geometry and attribute information for the spatial features in a data set. The geometry for a feature is stored as a shape comprising a vector of coordinates. Area features are represented as closed-loop, double-digitized polygons. The shapes together with data attributes linked to each shape create the representation of geographic data like countries, rivers and lakes.
- 
-Describing shapes requires specialized data formats. The most common data format is an ESRI Shapefile, and we will focus on these for this tutorial. You may also encounter GDB (common in hydrology), GeoJSON (used with web interfaces), KXML (used by Google Earth), and OSM (from OpenStreetMap) files. All of these can be converted to Shapefiles,
-using GIS software like QGIS.
+The boundaries of administrative units are described in specialized files that store geometry and attribute information for spatial features. The geometry for a feature is represented by a vector of coordinates, and area features are represented as closed-loop, double-digitized polygons. These shapes and their data attributes create the representation of geographic features like countries, rivers, and lakes.
+
+Describing shapes requires specialized data formats. The most common data format is an [ESRI Shapefile](https://www.esri.com/library/whitepapers/pdfs/shapefile.pdf), and we will focus on these in the tutorial. You may also encounter GDB (common in hydrology), GeoJSON (used with web interfaces), KXML (used by Google Earth), and OSM (from OpenStreetMap) files. All of these can be converted to Shapefiles, using GIS software like QGIS.
   
  
 ## Obtaining shapefiles
@@ -68,6 +73,49 @@ In some cases, you will need to create a shapefile from scratch. The most common
 1. Find an image that shows the regions that you want to digitize and import it into QGIS or ArcGIS.
 2. Typically, published images will not report the projection that was used, but you will need to find a mapping between points in the image and latitude-longitude coordinates. Use the [GDAL Georeferencer](https://www.qgistutorials.com/en/docs/georeferencing_basics.html) to make this a point-and-click task.
 3. Create new polygons by clicking around the edges of the polygon.
+
+
+## Working with shapefiles
+
[…]
