### 2020-12-03 b6dfe18
Update reduced-form-specification.md
