### 2020-12-04 896473d
edits to intro
