+++ tutorial-content/content/weather-and-climate-data.md
@@ -433,3 +433,216 @@ However, station data can’t be seen as the ‘true’ weather either; assumpti
 :class: warning
 Under normal circumstances, do not try to interpolate data yourself. Interpolated and reanalysis data products covered above were specifically designed for this purpose and have vetted methodologies and publicly available citable diagnostics and uncertainties.
 ```
+
+## Weather Mortality in the United States, Step 1: Preparing the Weather Data
+
+### Downloading the data 
+
+In our example, we will use Berkeley Earth (BEST) data. BEST data is
+available as national timeseries (area-weighted, so not usually useful
+even for studying national-level data) and as a 1-degree grid.
+
+1. Go to the BEST archive, http://berkeleyearth.org/archive/data/
+2. Scroll down to the Gridded Data (ignore the Time Series Data which
+   is listed first).
+3. Under Gridded Data, find "Daily Land (Experimental; 1880 –
+   Recent)".
+4. Download the Average Temperature (TAVG) data in 1-degree gridded
+   form for 1980-1989, 1990-1999, and 2000-2009.
+
+Place all three files in a `data/climate_data` folder.
+
+As a first pre-processing step, we will clip these files to the United
+States and combine them into a single file for easier use.
+
+### Loading the data
+
+First, let's load all of the data into memory. This code assumes that
+it is stored in a file (call it `preprocess_best.m` or
+`preprocess_best.py`) in a directory `code/`, a sister to the `data/`
+directory.
+
+````{tabbed} Python
+```{code-block} python
+import xarray as xr
+import datetime as dt
+
+ds = xr.open_mfdataset('../data/climate_data/Complete*.nc',combine='nested',concat_dim='time')
+
+# Through the concat_dim, some variables were broadcast over dimensions they shouldn't have been
+ds['climatology'] = ds.climatology.isel(time=0)
+ds['land_mask'] = ds.land_mask.isel(time=0)
+
+# Create time variable, which wasn't auto-generated from the netcdf due to BEST's ambiguous timing
+ds['time'] = (('time'),dt.datetime(1980,1,1)+np.arange(0,ds.dims['time'])*dt.timedelta(days=1))
+```
+````
+
+````{tabbed} Matlab
+```{code-block} matlab
+data_dir = '../data/climate_data/';
+fns = {'Complete_TAVG_Daily_LatLong1_1980.nc';...
+       'Complete_TAVG_Daily_LatLong1_1990.nc';...
+       'Complete_TAVG_Daily_LatLong1_2000.nc'};
+
+tas = cell(length(fns),1); months = cell(length(fns),1);
+for file_idx = 1:length(fns)
+   [Here is where we will combine the anomaly + climatology (see below).]
+end
+```
+````
+
+### Constructing temperature levels
+
+The variable `temperature` in BEST is actually the temperature
+_anomaly_; the actual temperature is formed by adding it to the
+`climatology` variable. Unfortunately, the `climatology` variable only
+accounts for days 1:365 of the year, and ignores leap days (which the
+`temperature` variable does not). This section doubles the climatology
+for Feb 28th to also work on Feb 29th, and creates a `tas` variable
+that's the `climatology` + `temperature`.
+
+````{tabbed} Python
+```{code-block} python
+import numpy as np
+import calendar
+
+# Expand climatology to span all days
[…]
