+++ example/code/pixel_overlaps.m
@@ -1,0 +1,991 @@
+% PIXEL_OVERLAPS    calculate the area overlaps between geographic vector
+%                   structs and a lat-lon pixel grid
+%   
+%   [geo_struct,issues,simps] = PIXEL_OVERLAPS(geo_struct,lat,lon) takes
+%   the struct of geographic vector data given by [geo_struct] in standard
+%   MATLAB geographic vector polygon format ([geo_struct].Geometry =
+%   'Polygon', can use either .X/.Y or .lon/.lat) and, for each element,
+%   finds all pixels in a grid given by lat/lon that overlap with the
+%   vector, and calculates for each pixel the overlapping area in km^2 with
+%   the vector. [issues] and [simps] give information about calculation
+%   issues (see the section on LOGGING AREAS... below). The following
+%   fields area added to [geo_struct] before output (both are [num_pix x 1]
+%   vectors):
+%       - [tested_pix]: the linear indices (in lon x lat space) of the
+%                       pixels that have a non-zero overlap with the vector 
+%                       polygon
+%       - [pix_areas]: the area (in km^2, assuming WGS84 ellipsoid) of
+%                      overlap between the pixel given by each element in
+%                      [tested_pix] and the vector polygon
+%   
+%   [issues,simps] = PIXEL_OVERLAPS(geo_struct,lat,lon,'output_struct',0) 
+%   doesn't return the modified geo_struct. 
+%
+%   [___] = PIXEL_OVERLAPS(geo_struct,lat,lon,'bounds',lat_bnds,lon_bnds)
+%   allow for the explicit input of latitude and longitude bounds
+%   (especially useful if a non-rectangular grid is used to ensure
+%   calculations are as accurate as possible). [lat_bnds] and [lon_bnds]
+%   should be [nlon/nlat x 2] vectors with the minimum and maximum extend
+%   of each (square) pixel given by each column, respectively.
+%
+%   [___] = PIXEL_OVERLAPS(geo_struct,lat,lon,'centroids',centroids) allows
+%   for the explicit input of centroids of each polygon (used to set a
+%   starting guess for pixels to test for overlaps). If no centroids are
+%   inputted, then the program calculates them (2017a+) or calculates the
+%   center of the bounding box (before 2017a). It is suggested to input
+%   centroids manually, since they can be calculated much faster in other
+%   programs (GIS, R's spatial data tools, etc.). If inputted, centroids
+%   should be a [n_polygon x 2] matrix with columns giving the [lon] and
+%   [lat] coordinates of each centroid, respectively.
+%
+%   SAVING OPTIONS (see flags below for full function options): 
+%   If 'save_struct' is set to true (see flags), then the output
+%   [geo_struct] from above in addition to [issues], [simps], [lat], and
+%   [lon] is saved to a matfile given by [filenameS] (see flags).
+%
+%   PIXEL_OVERLAPS(...,'[flag]',[params],...) modify program run as below:
+%       Input options
+%           - 'bounds',[lat_bnds],[lon_bnds] - explicitly set the latitude
+%                                              and longitude bounds of each
+%                                              pixel (they are otherwise
+%                                              derived as being equidistant
+%                                              between pixel vertices).
+%                                              [lat_bnds] and [lon_bnds]
+%                                              are [nlat/nlon x 2]
+%                                              matrices.
+%           - 'centroids',centroids          - input already calculated
+%                                              centroids (they are
+%                                              otherwise derived using
+%                                              CENTROID if > 2017a or just
+%                                              as the center of the
+%                                              Bounding Box if not).
+%                                              Centroids is a [n_polygon x
+%                                              2] matrix with columns
+%                                              giving the [lon] and [lat]
+%                                              coordinates, respectively).
+%
+%       Data processing and efficiency options
[…]

[…]
