### 2021-01-07 40596e6
some updates to weather/climate data section
