+++ tutorial-content/content/weather-and-climate-data.md
@@ -5,51 +5,58 @@ When using weather data as independent variables in an economic model, or climat
 - There is no universally *right* or *correct* weather or climate data product
 - Every weather or climate data product has its use cases, limitations, uncertainties, and quirks
 
-This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research.
+This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commoly used weather and climate data format in multiple languages, understanding the differences between gridded weather data products, finding evaluations of weather data products, examples on how to do download several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
+
+After working through this section, you will be familiar with:
+- working with data in the NetCDF format in your preferred language
+- understanding the differences between reanalysis and interpolated data products
+- knowing how to find information evaluating these data products for the variables and regions that matter for your work
+- identifying major uncertainties inherent to different types of weather data products
 
 ## 1.1 The NetCDF Data Format
 
-Almost all climate and weather datasets are released in [netCDF](https://climatedataguide.ucar.edu/climate-data-tools-and-analysis/netcdf-overview) format. It's efficient, self-describing, and supported by any major programming language, though you’ll have to pre-process data into another format (.csv, etc.) before you can use it in STATA. If you get familiar with the commands to read the header and access data in the language you’re most comfortable with, you will be able to work with almost any climate or weather dataset published in the world.
+The [NetCDF](https://climatedataguide.ucar.edu/climate-data-tools-and-analysis/NetCDF-overview) format is a common data format used for weather and climate data. Most weather and climate datasets will be published primarily or additionally in the NetCDF format. It's efficient, self-describing, and supported by any major programming language, though you’ll have to pre-process data into another format (.csv, etc.) before you can use it in STATA. If you get familiar with the commands to read the header and access data in the language you’re most comfortable with, you will be able to work with most climate or weather datasets published in the world.
 
 ### Supported Languages
 
-Through this section, when possible relevant commands for working with netCDF files are listed for:
+Through this section, we introduce relevant commands whenever possible for the following languages and packages:
 
-- Matlab (native support)
 - Python
-    - [xarray](http://xarray.pydata.org/en/stable/) (recommended) - a package for dealing with N-dimensional data that natively supports netCDF files
-    - [netCDF4-python](https://unidata.github.io/netcdf4-python/netCDF4/index.html) module
+    - [xarray](http://xarray.pydata.org/en/stable/) (recommended) - a package for dealing with N-dimensional data that natively supports NetCDF files
+    - [netcdf4-python](https://unidata.github.io/netcdf4-python/netCDF4/index.html) module
+- Matlab (native support)
 - R ([ncdf4](https://cran.r-project.org/web/packages/ncdf4/index.html) package)
-- [nco](http://nco.sourceforge.net) (“netCDF operators”) - command line tools
-    - documentation may look overwhelming, but at its core, it’s an easy way to check the contents of a file, collate different netCDF files, and extract individual variables without having to go through a full language.
+- [nco](http://nco.sourceforge.net) (“NetCDF operators”) - command line tools
+    - a series of tools to check the contents of a file, collate different NetCDF files, and extract individual variables without having to go through a full language.
     - important commands: `ncview` (to display spatial data), `ncks` (“nc kitchen sink” - to split or
       concatenate files command line), and `ncdump` (to print contents of the file)
 
[…]
