### 2021-01-08 c06ebd2
outline remaining tutorial
