+++ tutorial-content/_toc.yml
@@ -5,14 +5,14 @@
   chapters:
   - file: content/weather-and-climate-data
 
+- part: Weighting schemes
+  chapters:
+  - file: content/weighting-schemes
+
 - part: Developing a reduced-form specification
   chapters:
   - file: content/reduced-form-specification
   - file: content/functional-forms
-
-- part: Weighting schemes
-  chapters:
-  - file: content/weighting-schemes
 
 - part:  Geographical unit data
   chapters:
