+++ _sources/content/weather-and-climate-data.md
@@ -1,19 +1,30 @@
-# 1. Using Weather and Climate Data
+# Using Weather and Climate Data
 
+```{admonition} Key objectives and decision points
+:class: note
+Objectives:
+
+- Understand how to work with data in the NetCDF format in your preferred language
+- Understand the differences between reanalysis and interpolated data products
+- Learn how to find information evaluating these data products for the variables and regions that are relevant for your work
+- Learn how to identify major uncertainties inherent to different types of weather data products
+
+Decision points:
+
+ - How to choose climate or weather data to use in your research project?
+```
+
+```{note}
 When using weather data as independent variables in an economic model, or climate data to project your research results into the future, please note:
 
 - There is no universally *right* or *correct* weather or climate data product
 - Every weather or climate data product has its use cases, limitations, uncertainties, and quirks
+```
 
 This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commoly used weather and climate data format in multiple languages, understanding the differences between gridded weather data products, finding evaluations of weather data products, examples on how to do download several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
 
-After working through this section, you will be familiar with:
-- working with data in the NetCDF format in your preferred language
-- understanding the differences between reanalysis and interpolated data products
-- knowing how to find information evaluating these data products for the variables and regions that matter for your work
-- identifying major uncertainties inherent to different types of weather data products
 
-## 1.1 The NetCDF Data Format
+## The NetCDF Data Format
 
 The [NetCDF](https://climatedataguide.ucar.edu/climate-data-tools-and-analysis/NetCDF-overview) format is a common data format used for weather and climate data. Most weather and climate datasets will be published primarily or additionally in the NetCDF format. It's efficient, self-describing, and supported by any major programming language, though you’ll have to pre-process data into another format (.csv, etc.) before you can use it in STATA. If you get familiar with the commands to read the header and access data in the language you’re most comfortable with, you will be able to work with most climate or weather datasets published in the world.
 
@@ -33,8 +44,8 @@ Through this section, we introduce relevant commands whenever possible for the f
 
 For any python code chunks, it’s assumed that the xarray package is loaded as `xr` (`import xarray as xr`) or the NetCDF4-python module is loaded as `nc` (`import netCDF4 as nc`). For any R code chunks, it’s assumed the ncdf4 package is loaded (`library(ncdf4)`). 
 
-```{note}
-If you know several of the languages referred to in this tutorial and just want the author's opinion on which one to use, we suggest:
+```{tip}
+If you know several of the languages referred to in this tutorial and just want our opinion on which one to use, we suggest:
 - Python (`xarray`): if you want tools specifically designed for modern uses of weather/climate data that do much of the annoying background work (dealing with different file structures, variable names, date formats, etc.) for you, at the expense of less flexbility for uncommon needs
 - Matlab: if you like a simple, bare-bones treatment of data where you are in explicit control of everything that happens, at the expense of having to be more careful with pre-processing and backend management
 ```
@@ -211,7 +222,7 @@ vardata <- ncvar_get(ncfile, var, start=c(1,1,1), count=c(5,5,365))
 ```
 ````
 
[…]
