### 2021-01-13 acdf835
add info on VCS (#43) and CodeOcean
