### 2021-01-13 885e2c6
quick typo scan and fix formulas
