+++ tutorial-content/content/weather-and-climate-data.md
@@ -21,7 +21,7 @@ When using weather data as independent variables in an economic model, or climat
 - Every weather or climate data product has its use cases, limitations, uncertainties, and quirks
 ```
 
-This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commoly used weather and climate data format in multiple languages, understanding the differences between gridded weather data products, finding evaluations of weather data products, examples on how to do download several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
+This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commonly used weather and climate data format in multiple languages, understanding the differences between gridded weather data products, finding evaluations of weather data products, examples on how to do download several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
 
 
 ## The NetCDF Data Format
@@ -37,7 +37,7 @@ Through this section, we introduce relevant commands whenever possible for the f
     - [netcdf4-python](https://unidata.github.io/netcdf4-python/netCDF4/index.html) module
 - MATLAB (native support)
 - R ([ncdf4](https://cran.r-project.org/web/packages/ncdf4/index.html) package)
-- [nco](http://nco.sourceforge.net) (“NetCDF operators”) - command line tools
+- [nco](http://nco.sourceforge.net) (“NetCDF operators”) - command-line tools
     - a series of tools to check the contents of a file, collate different NetCDF files, and extract individual variables without having to go through a full language.
     - important commands: `ncview` (to display spatial data), `ncks` (“nc kitchen sink” - to split or
       concatenate files command line), and `ncdump` (to print contents of the file)
@@ -46,14 +46,8 @@ For any python code chunks, it’s assumed that the xarray package is loaded as 
 
 ```{tip}
 If you know several of the languages referred to in this tutorial and just want our opinion on which one to use, we suggest:
-- Python (`xarray`): if you want tools specifically designed for modern uses of weather/climate data that do much of the annoying background work (dealing with different file structures, variable names, date formats, etc.) for you, at the expense of less flexbility for uncommon needs
+- Python (`xarray`): if you want tools specifically designed for modern uses of weather/climate data that do much of the annoying background work (dealing with different file structures, variable names, date formats, etc.) for you, at the expense of less flexibility for uncommon needs
 - MATLAB: if you like a simple, bare-bones treatment of data where you are in explicit control of everything that happens, at the expense of having to be more careful with pre-processing and backend management
-```
-
-In general, we recommend using free and open-source programming languages such as Python and R, as using proprietary software may hinder code sharing and reuse. However, it is possible to run MATLAB and Stata code online and free of charge. A cloud-based reproducibility tool called [Code Ocean](https://codeocean.com) facilitates code sharing in the programming languages such as MATLAB, Stata, R, Python, C++, and others. Through a web browser, Code Ocean enables its users to run and share their code as a "research capsule".
-
-```{seealso}
-See a [demonstration of shared research material in CodeOcean](https://codeocean.com/capsule/8792614). This data and code were originally prepared and published with the paper "Climate Change, Mortality, and Adaptation: Evidence from Annual Fluctuations in Weather in the US'' by Olivier Deschênes and Michael Greenstone. 
 ```
 
 ### NetCDF Contents
@@ -66,7 +60,7 @@ NetCDF files have three kinds of information:
 - *Variables*: The arrays themselves, containing the data you want. Each array has an order of dimensions. For a two-dimensional array, the first dimension corresponds to the rows and the second dimension to the columns.
[…]
