+++ _sources/content/functional-forms.md
@@ -27,8 +27,10 @@ which the temperature falls into that bin.
 
 To calculate bin observations, count up the number of timesteps where
 the weather predictor falls into each bin:
-    $$T_{it} = \sum_{p \in i} \psi_{p} \sum \mathbf{1} \left \{  {T_{p i t} \in Bin_k} \right \}$$
-    where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
+
+$$T_{it} = \sum_{p \in i} \psi_{p} \sum \mathbf{1} \left \{  {T_{p i t} \in Bin_k} \right \}$$
+
+where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
     
 ## [Polynomial](https://en.wikipedia.org/wiki/Polynomial_regression)
 
@@ -50,17 +52,21 @@ confidence in these assumptions.
 In calculating the predictor values for a polynomial, consider the
 scale of the data-generating process. If it is a local process, the
 high-resolution weather values should be raised to powers before
-aggregating up to the regions in the economic data. That is, the the
+aggregating up to the regions in the economic data. That is, the 
 predictor for the $k$-th power of the polynomial is 
-    $$f(T_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} T_{p i t}^k$$ 
-    where $\psi_{p}$ is the weight assigned to the $p$ gridcell.  
+
+$$f(T_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} T_{p i t}^k$$ 
+
+where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
     
 The dose-response regression would then be applied as follows
-    $$F(T_{it})=\sum_{k} \beta_k f(T_{it}^k)$$
+
+$$F(T_{it})=\sum_{k} \beta_k f(T_{it}^k)$$
 	
 while the coefficients can be interepted as describing a local
 dose-response relationship:
-    $$F(T_{pit})=\sum_{k} \beta_k T_{pit}^k$$
+
+$$F(T_{pit})=\sum_{k} \beta_k T_{pit}^k$$
 
 ## [Restricted cubic spline](https://support.sas.com/resources/papers/proceedings16/5621-2016.pdf)
 
@@ -73,7 +79,7 @@ the placement of knots, and this can be used either to reflect
 features of the underlying process being modeled or to improve the
 fit.
 
-In the case where knots are choosen to maximize the fit of the curve,
+In the case where knots are chosen to maximize the fit of the curve,
 cross-validation is the preferred approach for both selecting the
 number of knots and their placement. The reference in this subsection
 title on cubic splines can be helpful in deciding the knot
@@ -83,17 +89,24 @@ Once knot locations are determiend, the weather data needs to be
 translated into RCS terms. As before let the gridded weather be $T_{p
 i t}$ and let there be $n$ knots, placed at $T_1<T_2<...<T_n$. Then we
 have a set of $(n-2)$ terms, here indexed by $k$ and defined as:
-    $$f(T_{i t})_k= \sum_{p \in \Theta(i)} \psi_{p} \{(T_{p i
+
+$$f(T_{i t})_k= \sum_{p \in \Theta(i)} \psi_{p} \{(T_{p i
 	t}-T_k)^3_+ - (T_{p i t} - T_{n-1})^3_+
[…]
