+++ _sources/content/reduced-form-specification.md
@@ -12,11 +12,11 @@ Decision points:
 - Settle on the temporal scale (e.g. days, months, years, etc.) for the analysis. It is generally decided based on how frequently we expect the responses on our dependent variable to change and the temporal scale of other datasets we intend to use.
 - Choose what scheme to use for the spatial aggregation. We can choose either *aggregation-before-transformation* or *transformation-before-aggregation* depending on what scale the process we are interested in, is actually occurring.
 - Decide on what *functional forms* to use in the analysis. It is always good to experiment with multiple functional forms and then decide on the main specification depending on which of them provides the best linear approximation of the true process.
-- Use cross-validation to converge on the most-preferred specification. It is always a good strategy to show results for other specification for attesting the robustness of results.
+- Use cross-validation to converge on the most-preferred specification. It is always a good strategy to show results for other specifications for attesting the robustness of results.
 ```
 
 This section describes some of the considerations that go into
-developing a [reduced-form specification](https://stats.stackexchange.com/questions/86004/what-is-simply-meant-by-reduced-form) using weather [panel data](https://en.wikipedia.org/wiki/Panel_data). Our discussion here will be very practical and limited to nonlinear panel regressions. 
+developing a [reduced-form specification](https://stats.stackexchange.com/questions/86004/what-is-simply-meant-by-reduced-form) using weather [panel data](https://en.wikipedia.org/wiki/Panel_data). Our discussion here will be very practical and limited to non-linear panel regressions. 
 
 ```{seealso}
 See also [Estimating the Economic Impacts of Climate Change Using Weather Observations](https://academic.oup.com/reep/article/14/1/1/5718007) for a review of the available econometric techniques and their strengths and weaknesses.  For an extensive review of the results available from
@@ -31,84 +31,95 @@ The choice of weather variables depends on the question we are trying
 to answer, and there are many forms to represent any given
 variable. For example, in the case of temperature, we can use
 $T_{avg}$, $T_{min}$, $T_{max}$, days above 30 C, heating and cooling
-degree-days, or growing degree-days. Morevoer, it is very important to
+degree-days, or growing degree-days. Moreover, it is very important to
 first think about possible *mechanism(s)* behind a change witnessed in
 the environment, and then only make choices for variables that can
 explain that mechanism. A few of the important and frequently-used
 weather variables are listed below, and why we might choose them:
 
-- **Temperature:** Temperature relationships are often preferred in
-  climate impacts research, because temperature is more predictable
-  than many other weather variables. There are various measures of temperature that can be used. Some of them are listed below:
-    1. *$T_{min}$, $T_{max}$:*  Many socioeconomic processes are more
-       sensitive to extreme temperatures than to variation in the
-       average. This is also useful when temperature variation is
-       large, leading to significant differences in the cold end and hot end responses. These are important metrics when heterogeneity within time units matters, and may better capture heat waves and cold spells. Also, note that $T_{min}$ better reflects nighttime temperatures while $T_{max}$ better reflects daytime temperatures. Not all datasets include $T_{min}$ or $T_{max}$. 
-    2. *$T_{avg}$:*  A good mean metric for seeing average response
-       over the temperature support, when there is not much variation
-       in temperature within each time unit considered in the
-       study. $T_{avg}$ is most appropriate when there is some natural
-       inertia in the response, so that the dependent variable is
-       responding to a kind of average over the last 24 hours. Note
-       that $T_{avg}$ is often just equal to $(T_{min} + T_{max}) / 2$, unless calculated from sub-daily data.
-    3. [*HDD/CDD & GDD:*](https://www.degreedays.net/introduction)
[…]
