+++ tutorial-content/content/weather-and-climate-data.md
@@ -4,25 +4,22 @@
 :class: note
 Objectives:
 
-- Understand how to work with data in the NetCDF format in your preferred language
-- Understand the differences between reanalysis and interpolated data products
-- Learn how to find information evaluating these data products for the variables and regions that are relevant for your work
-- Learn how to identify major uncertainties inherent to different types of weather data products
+- Understand how to work with data in the NetCDF format in your preferred language.
+- Understand the differences between reanalysis and interpolated data products.
+- Learn how to find information evaluating these data products for the variables and regions that are relevant for your work.
+- Learn how to identify major uncertainties inherent to different types of weather data products.
 
 Decision points:
 
  - How to choose climate or weather data to use in your research project?
 ```
 
-```{note}
-When using weather data as independent variables in an economic model, or climate data to project your research results into the future, please note:
-
-- There is no universally *right* or *correct* weather or climate data product
-- Every weather or climate data product has its use cases, limitations, uncertainties, and quirks
-```
-
 This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commonly used weather and climate data format in multiple languages, understanding the differences between gridded weather data products, finding evaluations of weather data products, examples on how to do download several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
 
+When using weather data as independent variables in an economic model, or climate data to project your research results into the future, please keep in mind that:
+
+- There is no universally *right* or *correct* weather or climate data product.
+- Every weather or climate data product has its use cases, limitations, uncertainties, and quirks.
 
 ## The NetCDF Data Format
 
@@ -154,7 +151,7 @@ Here are some important common “attributes” of NetCDF files or variables:
 - **Units** - generally attached to the variable in question. Common variable units:
     - *Temperature* - almost always in Kelvin
     - *Precipitation* - often in *kg/m^2s*, which is the SI unit for precipitation rate (volume per time). Multiply by 3600 to get mm/hour, or 141.7323 to get in/hour (the density of water is 1000 kg/m^3, multiplying by the density calculates the rate in m/s, or depth/time - the rest is just accounting to your desired units of depth and time).
-- **Missing / Fill Value** - if there are some crazy high or low values in your data, you may want to check if those just represent the missing / fill value, a common sub-attribute of variables. Data may also be stored in ["masked arrays"](https://currents.soest.hawaii.edu/ocn_data_analysis/_static/masked_arrays.html).
+- **Missing / Fill Value** - if there are some absurdly high or low values in your data, you may want to check if those just represent the missing / fill value, a common sub-attribute of variables. Data may also be stored in ["masked arrays"](https://currents.soest.hawaii.edu/ocn_data_analysis/_static/masked_arrays.html).
 
 ### Reading NetCDF data
 
@@ -230,9 +227,9 @@ The `time` variable can also be listed in a few different formats. An integer re
 
 You may want to visualize your weather or climate data, either for internal diagnostics or for production figures showing your data. It's generally good practice to double-check that your data downloaded and processed correctly, by making sure there aren't suspiciously many NAs/NaNs, that the lat/lon grid matches up with where the data should go (first-order check: does your temperature/precipitation field trace out major land/ocean boundaries, etc.), and that the data is consistent. Here are easy ways to plot the first timestep of a gridded dataset: 
 
-````{tabbed} Python
+````{tabbed} Python (xarray)
 
-```{code-block} python (xarray)
+```{code-block} python 
 # Assuming your variable is a 3-D (lat,lon,time, in any order) 
 # variable called "tas", in a dataset loaded using 
[…]
