+++ _sources/content/working-with-data-product.md
@@ -1,0 +1,137 @@
+# How to Start Working with a Data Product
+
+The following are incredibly useful resources to keep in mind while working with weather data:
+
+📚  [UCAR Climate Data Guide](https://climatedataguide.ucar.edu)
+: an encyclopedia for weather and climate data products with expert guidance on strengths and weaknesses for most commonly-used datasets
+
+📚  [Reanalysis.org](https://reanalyses.org)
+: a forum and wiki for makers and users of reanalyses with a focus on evaluating data products and comparing them with observational data
+
+📚  [Reanalysis and Observational Datasets and Variables](https://www.esrl.noaa.gov/psd/data/writ/moncomp/datasets/)
+: a "who's who" of historical weather products with basic facts about each
+
+These resources will help you determine which data product is right for you. They will also help you better interpret results from existing studies. (For example, NCEP2, which was commonly used in economics and policy studies, has known issues including larger biases in the Southern Hemisphere).
+
+Additionally, think about if you want climatological ("what you expect") data, rather than weather ("what you get") data. Climatology is generally known with more precision and available at higher resolution, but will only represent average patterns (e.g., average temperature by month) rather than any particular year.
+
+## Getting Started with a Data Product: BEST and CHIRPS
+
+Say you're looking at agriculture in Ethiopia. You would like both temperature and precipitation data (see the warning on hydrological variables below), and would like to use observational datasets. You consider BEST for temperature due to their daily output and CHIRPS, a hybrid station-satellite data product, for precipitation because you found literature specifically examining its biases in your region of interest. 
+
+```{list-table}
+:header-rows: 1
+:widths: 10 20 20
+
+* - 
+  - CHIRPS
+  - BEST
+
+* - *1. Understand the Data Product*
+  - CHIRPS is unfortunately not covered on the UCAR Climate Data Guide. _However_, you find several articles specifically validating it in Ethiopia (e.g. [Dinku et al. 2018](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3244) or [Gebrechorkos et al. 2018](https://eprints.soton.ac.uk/435188/)). You see that satellite data products are more biased South of the Rift Valley than North. You also see that CHIRPS tends to overestimate rainfall. You consider how these biases may affect your results.
+  - BEST _is_ [covered in the Climate Data Guide](https://climatedataguide.ucar.edu/climate-data/global-surface-temperatures-best-berkeley-earth-surface-temperatures). You see that it is able to provide high-resolution data because it includes incomplete and partial station records than other global products. However, you also see that the data is highly smoothed, meaning that it will likely be more biased in areas with large heterogeneity in temperature - for example in the mountainous highlands of Ethiopia. You resolve to use different sources to check for robustness.
+
+* - *2. Prepare to Download the Data*
+  - CHIRPS data is stored in a [publicly accessible directory](https://data.chc.ucsb.edu/products/CHIRPS-2.0/) (that is a simpler setup than most). You navigate to the `africa_daily/bils/` directory, and choose between 0.5 degree resolution and 2.5 degree resolution. However, you realize that you may have to write a shell script to download this data, to avoid clicking every file separately (using `ftplib` in python and similar packages is also an option).  
+  - Click on 'Get Data (external)' on the Climate Data Guide website, taking you to Berkley Earth's data overview page. You navigate down to the section on 'Gridded Data'. You'll have to click on every decade separately, but without further ado, clean NetCDF files are being downloaded to your machine.
+
+* - *3. Accessing the Data*
[…]
