+++ _sources/content/weighting-schemes.md
@@ -25,7 +25,7 @@ perform a weighted aggregation of gridded data to data regions.
 Taking the unweighted average of weather within a region can misrepresent what populations, firms, or other phenomena of interest are exposed to. For example, an unweighted annual average temperature for Canada is about -8°C, but most of the population and agricultural activity is in climate zones with mean temperatures over 6°C, and the urban heat island effect can raise temperatures by another 4°C. The time of year matters too, and you should consider a weighting scheme across days within a year, or even hours within a day.
 
 As described in section
-[Dealing with the spatial and temporal scales of economic processes](reduced-form-specification#Dealing-with-the-spatial-and-temporal-scales-of-economic-processes),
+[spatial and temporal scales of economic processes](content:spatial-and-temporal-scales),
 the scale of a phenomenon matters. Many processes occur at a more
 local scale than that which data is collected. The motivation for
 weighting is different for aggregation that represents averaged
@@ -62,6 +62,7 @@ Below are some common data sources for various weighting schemes.
     - Also consider gridded land use datasets: https://www.atmos.illinois.edu/~meiyapp2/datasets.htm
 - Look at the IRI Data Library for a large variety of datasets, available in any format: https://iridl.ldeo.columbia.edu/
 
+(content:working-with-gridded-data)=
 ## Working with gridded weighting data
 
 Weighting data files come in a wide range of file formats, since any
@@ -147,7 +148,7 @@ resolution by some factor. You want to do this so that two conditions to be met 
 Here we
 assume that both are regular latitude-longitude
 grids. See
-[Working with gridded weighting data](#Working-with-gridded-weighting-data) to
+[working with gridded weighting data](content:working-with-gridded-data) to
 understand the grid scheme for your weighting file; note that gridded
 weather data often reports the center of each grid cell, rather than
 the corner.
