+++ tutorial-content/content/suggestions.md
@@ -2,11 +2,11 @@
 
 ```{admonition} Key objectives
 :class: note
-- Understand how to organize your project materials 
-- Learn about useful tools that will help you work productively   
+- Understand how to organize your project data and code.
+- Learn about useful tools that will help you work productively.
 ```
 
-Following these suggestions will help you organize your research, which could improve [reproducibility](https://the-turing-way.netlify.com) (replicability) and reusability of your code and results. These could be particularly helpful when collaborating with other researchers (including a future self). 
+Following these suggestions will help you organize your research, which could improve [reproducibility](https://the-turing-way.netlify.com) (replicability) and reusability of your code and results. In computational research, reproducibility is the ability to obtain consistent results using the same input data, code, and methods. Considering these early on is particularly helpful when collaborating with other researchers (including a future self). 
 
 ## Organization of data and code
 
@@ -25,34 +25,64 @@ create your dataset with a lot of possible predictors and decide
 later.  Often merging together your panel dataset is laborious, and
 you do not want to do it more times than necessary.
 
-Even if you have organized your working directory perfectly, it is still good to include 
-some additional documentation in readme files (readme.txt, readme.md).
-Describe the files and process in these files, and try to keep them up-to-date as things are added or changed. 
+Good naming practices should be applied to files and folders to make clear the contents of your project. Informative naming will help you understand the purpose of each file, and it will improve its findability. Here are a few recommendations:
+
+```{panels}
+Do's
+^^^
+✅ Use delimiters (such as underscores "_" or dashes "-") to separate information contained in the file name.
+
+✅ Make sure file names are informative of their contents, and be consistent in using a naming scheme. 
+
+✅ If you want to indicate sequence, start your file or folder names with numbers (e.g., `01_clean_data`, `02_analyze`, `03_results`)
+
+✅ Use ASCII encoding if possible, with UTF-8 or UTF-16 as secondary options.
+---
+Don'ts
+^^^
+❎ Don't use spaces, punctuation, accented characters or case sensitivity. Use periods for file type only (e.g., `.csv`)
+
+❎ Don't use extra-long file and folder names, they should be as concise as possible while still being descriptive. 
+
+❎ Don't use proprietary file formats for storing image data. Images, pictures, or figures should be saved as JPEG or GIF files.
+```
 
 ```{admonition} Data storage 
 :class: tip
-Most universities typically have a data storage product available for their students and affiliates. 
+Most universities have a data storage product available for their students and affiliates. 
 We recommend you inquire at your university about what can be the best place to store data.
 ```
 
-```{admonition} Immutable data
-:class: warning
-Original data (or source or raw data) should be immutable, meaning that it should never be modified by your code. 
-Instead of making changes to the original data, you should create derived (new) datasets.
-This is important because the pre-processing performed on source data is as important as the final analysis steps.
-In addition, it will allow you to reuse the original data multiple times.
+## Documentation
+
+Even if you have organized your working directory perfectly, it is still good to include additional documentation in readme files (`readme.txt` or `README.md`). A README file is usually a plain text file stored in a top-level directory. Describe the files and process in these files, and try to keep them up-to-date as things are added or changed. Project documentation will be primarily helpful to you later on (for example, if you need to reuse the code with some modifications), but also to your peers and collaborators in case you share your data and code. Here are a few tips on creating a good README file:
+
+
+```{panels}
+Do's
+^^^
+✅ Include a project description, explain how data was obtained and analyzed.
+
[…]
