+++ tutorial-content/content/example-step2.md
@@ -27,8 +27,8 @@ death data, which is what we have.
 Then, we want to transform the weather data by creating two predictor
 variables, for the linear and quadratic terms relative to 20 C.
 
-$$X_{1, i, y} = \sum_{t in \text{Year}(y)} \sum_{p \in \Theta(i)} \psi_{p} (T_{p i t} - 20)$$
-$$X_{2, i, y} = \sum_{t in \text{Year}(y)} \sum_{p \in \Theta(i)} \psi_{p} (T_{p i t}^2 - 20^2)$$
+$$X_{1, i, y} = \sum_{t \in \text{Year}(y)} \sum_{p \in \Theta(i)} \psi_{p} (T_{p i t} - 20)$$
+$$X_{2, i, y} = \sum_{t \in \text{Year}(y)} \sum_{p \in \Theta(i)} \psi_{p} (T_{p i t}^2 - 20^2)$$
 
 where $\psi_{p}$ is the population in gridcell $p$.
 
@@ -50,7 +50,103 @@ writeRaster(rr3, "~/groups/weatherpanels/weather-panel.github.io/example/data/pc
 
 ### Aggregating the weather data
 
-@Kevin
+One of the tutorial authors has produced a new tool for aggregating
+gridded data according to a shapefile: `xagg`. Since it handles
+partial grid-cell overlapping and arbitrary weighting grid, we
+recommend that it be used irrespective of the language being used
+elsewhere. Install it according to the instructions here:
+https://github.com/ks905383/xagg
+
+You will need to download a shapefile for the US counties. We use one
+from the ESRI community website:
+https://community.esri.com/ccqpr47374/attachments/ccqpr47374/enterprise-portal/4124/1/UScounties.zip
+
+Save its contents to a folder `geo_data` in the `data` directory. The
+following code should be run from the `data` directory.
+
+## Setting up the environment
+
+The approach uses `xarray` for gridded data, `geopandas` to work with shapefiles, and `xagg` to aggregate gridded data onto shapefiles. 
+
+```python
+import xarray as xr
+import geopandas as gpd
+import xagg as xa
+import numpy as np
+```
+
+First, load the data from the previous steps:
+
+```python
+# Load temperature data using xarray
+ds_tas = xr.open_dataset('climate_data/tas_day_BEST_historical_station_19800101-19891231.nc')
+
+# Load population data using xarray 
+ds_pop = xr.open_dataset('pcount/usap90ag.nc')
+
+# Load county shapefiles using geopandas
+gdf_counties = gpd.read_file('geo_data/UScounties.shp')
+```
+
+### Traansforming the data
+
+Next, we need to construct any nonlinear transformations of the data.
+
+For our econometric model, we want temperature in both linear and quadratic form, centered around $20^\circ$ C: 
+
+$T-20^\circ C$ and $T^2 - (20^\circ C)^2$.
+
+```python
+ds_tas['tas_adj'] = ds_tas.tas-20
+ds_tas['tas_sq'] = ds_tas.tas**2 - 20**2
+
+# xagg aggregates every gridded variable in ds_tas - however, we don't need
+# every variable currently in tas. Let'ss drop "tas" (the un-adjusted temperature)
+# and "land_mask" which is included, but not necessary for our further analysis.
+ds_tas = ds_tas.drop('tas')
+ds_tas = ds_tas.drop('land_mask')
+```
+
+### Create map of pixels onto polygons
+
+We need to create a weight map of pixels to polygons. For each
+polygon, we need to know which pixels overlap it and by how much.
+
+`xagg` does this by creating polygons for each pixel in the gridded
+dataset, taking the intersect between each county polygon and all
+pixel polygons, and calculating the average area of overlap between
+the pixels that touch the polygon and the polygon.
+
+```python
+weightmap =
[…]
