+++ tutorial-content/content/example-step1.md
@@ -1,0 +1,196 @@
+# Weather Mortality in the United States, Step 1: Preparing the Weather Data
+
+## Downloading the data 
+
+In our example, we will use Berkeley Earth (BEST) data. BEST data is
+available as national timeseries (area-weighted, so not usually useful
+even for studying national-level data) and as a 1-degree grid.
+
+1. Go to the BEST archive, http://berkeleyearth.org/archive/data/
+2. Scroll down to the Gridded Data (ignore the Time Series Data which
+   is listed first).
+3. Under Gridded Data, find "Daily Land (Experimental; 1880 –
+   Recent)".
+4. Download the Average Temperature (TAVG) data in 1-degree gridded
+   form for 1980 - 1989.
+5. Place this file (`Complete_TAVG_Daily_LatLong1_1980.nc`) in the
+   `data/climate_data` folder.
+
+As a first pre-processing step, we will clip this file to the United
+States.
+
+## Loading the data
+
+First, let's load all of the data into memory. This code assumes that
+it (the code) is stored in a file (call it `preprocess_best.m` or
+`preprocess_best.py`) in a directory `code/`, a sister to the `data/`
+directory.
+
+````{tabbed} Python
+```{code-block} python
+import numpy as np
+import xarray as xr
+import datetime as dt
+
+ds = xr.open_dataset('../data/climate_data/Complete_TAVG_Daily_LatLong1_1980.nc')
+
+# Create time variable, which wasn't auto-generated from the netcdf due to BEST's ambiguous timing
+ds['time'] = (('time'),dt.datetime(1980,1,1)+np.arange(0,ds.dims['time'])*dt.timedelta(days=1))
+```
+````
+
+````{tabbed} Matlab
+```{code-block} matlab
+data_dir = '../data/climate_data/';
+filename = 'Complete_TAVG_Daily_LatLong1_1980.nc';
+	   
+clim_tmp = ncread([data_dir,filename],'climatology');
+anom_tmp = ncread([data_dir,filename],'temperature');
+doy_tmp = ncread([data_dir,filename],'day_of_year');
+
+% Also loading the "months" variable - most datasets don't have it, but
+% it will make more sophisticated projecting methods easier. This is
+% particularly useful because of the gregorian calendar (which includes
+% leap days, and therefore would otherwise add an extra step to
+% determining which month each datastep is)
+months = ncread([data_dir,filename],'month');
+```
+````
+
+## Constructing temperature levels
+
+The variable `temperature` in BEST is actually the temperature
+_anomaly_; the actual temperature is formed by adding it to the
+`climatology` variable. Unfortunately, the `climatology` variable only
+accounts for days 1:365 of the year, and ignores leap days (which the
+`temperature` variable does not). This section doubles the climatology
+for Feb 28th to also work on Feb 29th, and creates a `tas` variable
+that's the `climatology` + `temperature`.
+
+````{tabbed} Python
+```{code-block} python
+import calendar
+
+# Expand climatology to span all days
+clim_tmp = xr.DataArray(dims=('time','latitude','longitude'),
+                        coords={'time':ds.time,'latitude':ds.latitude,'longitude':ds.longitude},
+                        data=np.zeros((ds.dims['time'],ds.dims['latitude'],ds.dims['longitude']))*np.nan)
+
[…]
