+++ tutorial-content/content/example-step4.md
@@ -1,39 +1,98 @@
 # Weather Mortality in the United States, Step 4: Producing results
 
-## Loading the data
+## Constructing annual polynomials summed over days
 
-Here we sum across all races and ages and merge the mortality and
-population data.
+We have county aggregated data from the previous step, but we also
+need to sum it to the annual level to match the mortality data. Again,
+this code should be run from a sister directory to the `data`
+directory.
+
+Let's first turn the data from wide form into long form and label the
+years.
+
+```R
+library(reshape2)
+
+clim <- read.csv("../data/climate_data/agg_vars.csv")
+
+clim2 <- melt(clim[, c(-1:-5)], id.vars='FIPS')
+clim2$date <- as.Date("1980-01-01") + as.numeric(gsub("tas_adj|tas_sq", "", clim2$variable))
+clim2$year <- as.numeric(substring(clim2$date, 1, 4))
+```
+
+Now we can sum over years. To do this, we will need to base this on
+the first several characters of the column names (not variable row
+values).
 
 ```R
 library(dplyr)
 
-df.mort <- read.fwf("~/groups/weatherpanels/weather-panel.github.io/example/data/cmf/Mort7988.txt",
-   c(5, 4, 10, 4), col.names=c('fips', 'year', 'ignore', 'deaths'))
-
-df2.mort <- df.mort %>% group_by(fips, year) %>% summarize(deaths=sum(deaths))
-
-df.pop <- read.fwf("~/groups/weatherpanels/weather-panel.github.io/example/data/cmf/Pop7988.txt",
-   c(5, 4, 9, rep(8, 12), 25, 1), col.names=c('fips', 'year',
-   'ignore', paste0('pop', 1:12), 'county', 'type'))
-
-df2.pop <- df.pop %>% group_by(fips, year) %>% summarize(pop=sum(pop1 +
-    pop2 + pop3 + pop4 + pop5 + pop6 + pop7 + pop8 + pop9 + pop10 +
-    pop11 + pop12), type=type[1])
-
-df3 <- df2.pop %>% left_join(df2.mort, by=c('fips', 'year'))
-df3$deaths[is.na(df3$deaths)] <- 0
-
-df4 <- subset(df3, type == 3)
-
-write.csv(df4[, -which(names(df4) == 'type')], "~/groups/weatherpanels/weather-panel.github.io/example/data/cmf/merged.csv", row.names=F)
+clim3 <- clim2 %>% group_by(FIPS, year) %>% summarize(tas_adj=sum(value[substring(variable, 1, 7) == 'tas_adj']), tas_sq=sum(value[substring(variable, 1, 6) == 'tas_sq']))
 ```
-
-## Constructing annual polynomials summed over days
-
-
 
 ## Merging weather and outcome data
 
+And now we can merge in the mortality data! This is by county (FIPS
+code) and year. We also construct the death rate, as deaths per
+100,000 people in the population.
+
+```R
+df <- read.csv("../data/cmf/merged.csv")
+
+df2 <- df %>% left_join(clim3, by=c('fips'='FIPS', 'year'))
+
+df2$deathrate <- 100000 * df2$deaths / df2$pop
+df2$deathrate[df2$deathrate == Inf] <- NA
+```
+
 ## Running the regression
 
+Let's run our central regression, relating death rate to
+temperature. Note that since we do not control for precipitation,
+temperature in this case includes correlated effects of rainfall. As a
+result, it is not ideal for future projections.
+
+For fixed effects, we use county fixed effects and state trends. More
+saturated fixed effects should be explored.
+
+```R
+library(lfe)
+
[…]
