+++ tutorial-content/_toc.yml
@@ -13,15 +13,16 @@
   - file: content/functional-forms
   - file: content/example-step2
 
-- part:  Geographical unit data
+- part: Geographical unit data
   chapters:
   - file: content/geographical-unit-data
   - file: content/working-with-shapefiles
   - file: content/weighting-schemes
   - file: content/example-step3
 
-- part: Suggestions for work organization 
+- part: Bringing it all together
   chapters:
+  - file: content/producing-results
   - file: content/suggestions
   - file: content/example-step4
 
