+++ _sources/content/working-with-data-product.md
@@ -1,3 +1,4 @@
+(content:working-with-data)=
 # How to Start Working with a Data Product
 
 The following are incredibly useful resources to keep in mind while working with weather data:
@@ -13,11 +14,18 @@ The following are incredibly useful resources to keep in mind while working with
 
 These resources will help you determine which data product is right for you. They will also help you better interpret results from existing studies. (For example, NCEP2, which was commonly used in economics and policy studies, has known issues including larger biases in the Southern Hemisphere).
 
-Additionally, think about if you want climatological ("what you expect") data, rather than weather ("what you get") data. Climatology is generally known with more precision and available at higher resolution, but will only represent average patterns (e.g., average temperature by month) rather than any particular year.
+The following section shows several examples of choosing and downloading weather data given a region and variable of interest. Though there are many more products than the three introduced here, each with their own download procedures and quirks, these examples show a few common setups that you may encounter (CHIRPS: ftp directory; BEST: website browser; ERA5: data storage system). 
+
+Generally, it's good practice to first research which data products are appropriate for your variable and area of interest. Questions you should be thinking of include:
+- does the data product have the variable I need? 
+- are the data available at the resolution I need? 
+- are biases reasonable for the variable and region of interest?
+
+The answers to the first two questions above are easily found on the website of each dataset. The third question is more complex - the [UCAR Climate Data Guide](https://climatedataguide.ucar.edu) introduced above is a good first place to look. A google scholar search of the form "[data product name] validation OR evaluation OR bias OR uncertainty" may be useful as well. 
 
 ## Getting Started with a Data Product: BEST and CHIRPS
 
-Say you're looking at agriculture in Ethiopia. You would like both temperature and precipitation data (see the warning on hydrological variables below), and would like to use observational datasets. You consider BEST for temperature due to their daily output and CHIRPS, a hybrid station-satellite data product, for precipitation because you found literature specifically examining its biases in your region of interest. 
+Say you're looking at agriculture in Ethiopia. You would like both temperature and precipitation data (remember the [warning on hydrological variables](content:warning-on-hydrological)), and would like to use observational datasets. You consider BEST for temperature due to their daily output and CHIRPS, a hybrid station-satellite data product, for precipitation because you found [literature](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3244) specifically examining its biases in your region of interest. 
 
 ```{list-table}
 :header-rows: 1
@@ -29,15 +37,15 @@ Say you're looking at agriculture in Ethiopia. You would like both temperature a
 
 * - *1. Understand the Data Product*
   - CHIRPS is unfortunately not covered on the UCAR Climate Data Guide. _However_, you find several articles specifically validating it in Ethiopia (e.g. [Dinku et al. 2018](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3244) or [Gebrechorkos et al. 2018](https://eprints.soton.ac.uk/435188/)). You see that satellite data products are more biased South of the Rift Valley than North. You also see that CHIRPS tends to overestimate rainfall. You consider how these biases may affect your results.
-  - BEST _is_ [covered in the Climate Data Guide](https://climatedataguide.ucar.edu/climate-data/global-surface-temperatures-best-berkeley-earth-surface-temperatures). You see that it is able to provide high-resolution data because it includes incomplete and partial station records than other global products. However, you also see that the data is highly smoothed, meaning that it will likely be more biased in areas with large heterogeneity in temperature - for example in the mountainous highlands of Ethiopia. You resolve to use different sources to check for robustness.
[…]
