+++ requirements.txt
@@ -6,7 +6,7 @@ attrs==20.3.0
 Babel==2.8.1
 backcall==0.2.0
 beautifulsoup4==4.9.3
-bleach==3.2.1
+bleach==3.3.0
 certifi==2020.11.8
 cffi==1.14.3
 chardet==3.0.4
