### 2021-02-03 5400eb0
python version of step 2
