+++ tutorial-content/content/example-step2.md
@@ -63,20 +63,23 @@ writeRaster(rr3, "../data/pcount/usap90ag.nc4",
 #!pip install rasterio # will work in Jupyter Notebook
 import xarray as xr
 rr = xr.open_rasterio("../data/pcount/usap90ag.bil")
+rr = rr.sel(band=1, drop=True)
 
-#TODO rr2 <- aggregate(rr, fact=24, fun=sum)
-
+# Crop to US bounding box
 xmin = -126
 xmax = -65
 ymin = 23
 ymax = 51
 
-sel_lon = (rr2.x >= xmin) & (rr2.x <= xmax)
-sel_lat = (rr2.y >= ymin) & (rr2.x >= ymax)
-rr3 = rr2.where(sel_lon & sel_lat, drop=True)
+sel_lon = (rr.x >= xmin) & (rr.x <= xmax)
+sel_lat = (rr.y >= ymin) & (rr.y <= ymax)
+rr2 = rr[sel_lat, sel_lon]
 
-rr3.rename({'x': 'lon','y': 'lat'})
-rr3.to_netcdf("../data/pcount/usap90ag.nc4")
+# Aggregate to 1-degree cells
+rr3 = rr2.coarsen(x=24, y=24).sum()
+
+rr3 = rr3.rename(x='lon', y='lat')
+rr3.to_dataset(name="Population").to_netcdf("../data/pcount/usap90ag.nc4")
 ```
 ````
 
@@ -149,17 +152,15 @@ import pandas as pd
 df_mort = pd.read_csv("../data/cmf/Mort7988.txt", names = ['input'])
 
 # parse input
-df_mort[['fips', 'year', 'ignore', 'deaths']] = df_mort.input.apply(
+df_mort2 = pd.DataFrame(df_mort.input.apply(
     lambda x: [
-        x[slice(*slc)] for slc in [(0,5), (5,9), (9,20), (20,len(x))]]).tolist()
-df_mort = df_mort.drop(columns=['input'])
+        x[slice(*slc)] for slc in [(0,5), (5,9), (20,len(x))]]).tolist(),
+	columns=['fips', 'year', 'deaths'])
 
-cols = ['fips', 'year', 'deaths']
-df_mort = df_mort[cols].apply(pd.to_numeric, errors='coerce')
+df_mort3 = df_mort2.apply(pd.to_numeric, errors='coerce')
 
-df_mort.deaths = pd.to_numeric(df_mort.deaths)
-df2_mort = df_mort.groupby(['fips', 'year']).sum()
-df2_mort.head()
+df_mort4 = df_mort3.groupby(['fips', 'year']).sum()
+df_mort4.head()
 ```
 | fips, year   |   deaths |
 |:-------------|---------:|
@@ -170,50 +171,50 @@ df2_mort.head()
 | (1001, 1983) |      267 |
 
 ```{code-block} python
-df_pop = pd.read_csv("cmf/Pop7988.txt", names = ['input'])
+df_pop = pd.read_csv("../data/cmf/Pop7988.txt", names = ['input'])
 
 slices = [(0, 5), (5, 9), (9,18)] + \
          [(n, n+8) for n in range(18, 114, 8)] + \
          [(114, 139), (139, 140)]
 
 # parse input
-df_pop[['fips', 'year', 'ignore'] + ["pop" + str(i) for i in range(1,13)] + \
-   ['county', 'type']] = df_pop.input.apply(
+df_pop2 = pd.DataFrame(df_pop.input.apply(
     lambda x: [
-        x[slice(*slc)] for slc in slices]).tolist()
-df_pop = df_pop.drop(columns=['input'])
+        x[slice(*slc)] for slc in slices]).tolist(),
[…]
