+++ _sources/content/example-step2.md
@@ -44,7 +44,8 @@ aggregate it to the scale of the weather. We also need it in NetCDF
 format, for the aggregation step. Again, this code assumes that it is
 being run from a directory `code`, sister to the `data` directory.
 
-```R
+````{tabbed} R
+```{code-block} R
 library(raster)
 rr <- raster("../data/pcount/usap90ag.bil")
 
@@ -55,6 +56,32 @@ writeRaster(rr3, "../data/pcount/usap90ag.nc4",
  overwrite=TRUE, format="CDF", varname="Population", varunit="people",
  xname="lon", yname="lat")
 ```
+````
+ 
+````{tabbed} Python
+```{code-block} python
+#!pip install rasterio # will work in Jupyter Notebook
+import xarray as xr
+rr = xr.open_rasterio("../data/pcount/usap90ag.bil")
+rr = rr.sel(band=1, drop=True)
+
+# Crop to US bounding box
+xmin = -126
+xmax = -65
+ymin = 23
+ymax = 51
+
+sel_lon = (rr.x >= xmin) & (rr.x <= xmax)
+sel_lat = (rr.y >= ymin) & (rr.y <= ymax)
+rr2 = rr[sel_lat, sel_lon]
+
+# Aggregate to 1-degree cells
+rr3 = rr2.coarsen(x=24, y=24).sum()
+
+rr3 = rr3.rename(x='lon', y='lat')
+rr3.to_dataset(name="Population").to_netcdf("../data/pcount/usap90ag.nc4")
+```
+````
 
 ## Downloading the mortality data
 
@@ -93,7 +120,8 @@ on).
 Here we sum across all races and ages and merge the mortality and
 population data.
 
-```R
+````{tabbed} R
+```{code-block} R
 library(dplyr)
 
 df.mort <- read.fwf("../data/cmf/Mort7988.txt",
@@ -116,3 +144,77 @@ df4 <- subset(df3, type == 3)
 
 write.csv(df4[, -which(names(df4) == 'type')], "../data/cmf/merged.csv", row.names=F)
 ```
+````
+ 
+````{tabbed} Python
+```{code-block} python
+import pandas as pd
+df_mort = pd.read_csv("../data/cmf/Mort7988.txt", names = ['input'])
+
+# parse input
+df_mort2 = pd.DataFrame(df_mort.input.apply(
+    lambda x: [
+        x[slice(*slc)] for slc in [(0,5), (5,9), (20,len(x))]]).tolist(),
+	columns=['fips', 'year', 'deaths'])
+
+df_mort3 = df_mort2.apply(pd.to_numeric, errors='coerce')
+
+df_mort4 = df_mort3.groupby(['fips', 'year']).sum()
+df_mort4.head()
+```
+| fips, year   |   deaths |
+|:-------------|---------:|
+| (1001, 1979) |      225 |
+| (1001, 1980) |      221 |
+| (1001, 1981) |      221 |
+| (1001, 1982) |      223 |
+| (1001, 1983) |      267 |
+
+```{code-block} python
+df_pop = pd.read_csv("../data/cmf/Pop7988.txt", names = ['input'])
+
+slices = [(0, 5), (5, 9), (9,18)] + \
+         [(n, n+8) for n in range(18, 114, 8)] + \
+         [(114, 139), (139, 140)]
+
+# parse input
+df_pop2 = pd.DataFrame(df_pop.input.apply(
+    lambda x: [
+        x[slice(*slc)] for slc in slices]).tolist(),
+	columns=['fips', 'year', 'ignore'] + ["pop" + str(i) for i in range(1,13)] + ['county', 'type'])
+
[…]
