### 2021-02-16 8298b2a
text edits, style and cosmetics
