+++ tutorial-content/content/weather-and-climate-data.md
@@ -122,14 +122,15 @@ The following terminology will be useful to recognize even when filenames are no
 ```{seealso}
 For more information on "CMIP5" and "CMIP6" terminology, see: [CMIP6 Guidance for Data Users](https://pcmdi.llnl.gov/CMIP6/Guide/dataUsers.html) or the [CMIP5 Standard Output](https://pcmdi.llnl.gov/mips/cmip5/requirements.html).
 ```
-
+```{caution}
 Weather and climate variables in NetCDF files may be organized in a few different ways:
 
-- one variables per file: each file contains a single variable over the whole (or some large subset) of the time domain
+- one variable per file: each file contains a single variable over a time domain
 - one timestep per file: each file contains a single timestep with a suite of variables
-- combination: file contains many variables over a large time domain (rare due to size constraints)
+- combination: file contains many variables over a time domain (rare due to size constraints)
 
-To figure out which file saving convention your NetCDF file uses, and what is contained, check the header of the file.
+To figure out which file saving convention your NetCDF file uses and what is contained, you'll need to check the header of the file.
+```
 
 ### The NetCDF Header
 
@@ -265,7 +266,7 @@ These files also include variables that give values for indices along each dimen
 Longitude can be either of the form `-180:180` or `0:360`. In the latter form, `359` is 1$^\circ$W and so forth. 
 ```
 
-The `time` variable can also be listed in a few different formats. An integer representation of “days since [some date, often 1850-01-01]” is common, as is an integer representation of the form [YYYYMMDD], among others. The key is to always check the description of the variable in the header, and adjust your methods accordingly until it’s in a format you want it in. If you're using python, the `xarray` package has the ability to interpret some of these time representations for you and translates them into the `datetime64` class, which makes some kinds of manipulation, like [averaging over months](http://xarray.pydata.org/en/stable/time-series.html), easier.
+The `time` variable can also be listed in a few different formats. An integer representation of “days since [some date, often 1850-01-01]” is common, as is an integer representation of the form [YYYYMMDD], among others. The key is to always check the description of the variable in the header, and adjust your methods accordingly until it’s in a format you want it in. If you're using Python, the `xarray` package has the ability to interpret some of these time representations for you and translates them into the `datetime64` class, which makes some kinds of manipulation, like [averaging over months](http://xarray.pydata.org/en/stable/time-series.html), easier.
 
 Now that you have loaded your weather and climate data, a good practice is to double-check that it is downloaded and processed correctly. Common red flags include suspiciously many `NA` / `NaN` values, suspiciously high or low values, or variables that unexpectedly don't line up with geographic features. Plotting your data can be a good first-order check: 
 
@@ -434,24 +435,29 @@ See also UCAR's Model Data Guide [summary](https://climatedataguide.ucar.edu/cli
 ### Regional Datasets
 Observational datasets exist with both global coverage (e.g., GISTEMP, HadCRUT, etc.) or regional coverage (e.g., PRISM in North America, TRMM in the tropics, etc.). Global datasets attempt to build a self-consistent database spanning the whole globe, and are therefore more likely to have sparser data coverage in specific regions - both as a logistical limitation, but also to ensure data pre-processing is as standardized as possible. Regional datasets may provide higher-resolution coverage and more specialized methodologies by incorporating local climatological knowledge or data sources that are not publicly available or parsable by global datasets (see e.g., the discussion in [Dinku et al. 2019](http://www.sciencedirect.com/science/article/pii/B9780128159989000075)). 
 
-### Using Gridded Weather Data
[…]
