+++ tutorial-content/content/weighting-schemes.md
@@ -83,7 +83,7 @@ CSV, spaces for ASC).
 
 The spatial gridding scheme is determined by 6 numbers: a latitude and longitude of an origin point, a horizontal and vertical cell lengths, and a number of rows and columns.
     - The most common origin point is the location of the lower-left corner of the lower-left grid cell. For example, for a global dataset, that might be 90°S, 180°W, which is represented in x, y coordinates as (-180, -90). Sometimes (particularly with NetCDF files), grid cell center locations will be used instead.
-    - Grid cell sizes are often given as decimal representation of fractions of a degree, such as 0.0083333333333 = 1 / 120 of a degree. This is the grid cell size needed globally to ensure a km-scale resolution. Usually the horizontal and vertical grid cell lengths are the same, and reported as a single number.
+    - Grid cell sizes are often given as decimal representation of fractions of a degree, such as $0.0083333333333 = 1 / 120$ of a degree. This is the grid cell size needed globally to ensure a km-scale resolution. Usually the horizontal and vertical grid cell lengths are the same, and reported as a single number.
     - The number of grid cells is the most common way to describe the spatial coverage of the dataset. A global dataset will have 180 / cellsize rows and 360 / cellsize columns.
 
 
@@ -126,15 +126,15 @@ Use the `raster` library. For example:
 library(raster)
 rr <- raster(filename)
 ```
+
+``` {seealso}
+If you are using R, take a look at the [Introduction to Geospatial Raster and Vector Data with R](https://datacarpentry.org/r-raster-vector-geospatial/),
+which has some extensive examples of working with geospatial raster data.
+```
 ````
 ````{tabbed} Python
 Take a look at <https://github.com/jrising/research-common/tree/master/python/geogrid>.
 ````
-
-If you are using R, take a look at
-the
-[Introduction to Geospatial Raster and Vector Data with R](https://datacarpentry.org/r-raster-vector-geospatial/),
-which has some extensive examples of working with geospatial raster data.
 
 In some cases, it is appropriate and possible to use time-varying weighting schemes. For example, if population impacts are being studied, and the scale of the model is individuals, annual estimate of population can be used. This kind of data is often either in NetCDF format (see above), or as a collection of files.
 
@@ -220,9 +220,9 @@ The following recipe should work for most cases to align weighting data with a w
 1. **Upsample the weighting data until the grid of the weighting data evenly divides up the weather data.** Start by considering the southwest corner of each dataset. Even if the
 datasets are of the same resolution, upsampling may be required to
 make the grid cells line up. For example, if the weighting data has a
-southwest corner of (0 N, 0 E) and a resolution of 1 degree, while the weather data starts at (0.5
-N, 0.5 E) and a resolution of 1 degree, the weighting data will need
-to be upsampled by a factor of 2, so that it provides a grid line at
+southwest corner of (0 N, 0 E) and a resolution of $1$ degree, while the weather data starts at (0.5
+N, 0.5 E) and a resolution of $1$ degree, the weighting data will need
+to be upsampled by a factor of $2$, so that it provides a grid line at
 (0.5 N, 0.5 E).
 
 2. **Clip the two datasets so that they line up.** After step 1, it should be possible to clip the two datasets to the
[…]
