+++ _sources/content/working-with-data-product.md
@@ -36,11 +36,11 @@ Say you're looking at agriculture in Ethiopia. You would like both temperature a
   - BEST
 
 * - *1. Understand the Data Product*
-  - CHIRPS is unfortunately not covered on the UCAR Climate Data Guide. _However_, you find several articles specifically validating it in Ethiopia (e.g. [Dinku et al. 2018](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3244) or [Gebrechorkos et al. 2018](https://eprints.soton.ac.uk/435188/)). You see that satellite data products are more biased South of the Rift Valley than North. You also see that CHIRPS tends to overestimate rainfall. You consider how these biases may affect your results.
+  - CHIRPS is unfortunately not covered on the UCAR Climate Data Guide. _However_, you find several articles specifically validating it in Ethiopia (e.g., [Dinku et al. 2018](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3244) or [Gebrechorkos et al. 2018](https://eprints.soton.ac.uk/435188/)). You see that satellite data products are more biased South of the Rift Valley than North. You also see that CHIRPS tends to overestimate rainfall. You consider how these biases may affect your results.
   - BEST _is_ [covered in the Climate Data Guide](https://climatedataguide.ucar.edu/climate-data/global-surface-temperatures-best-berkeley-earth-surface-temperatures). You see that it is able to provide high-resolution data because it includes incomplete and partial station records that other global data products may throw out. However, you also see that the data is highly smoothed, meaning that it will likely be more biased in areas with large heterogeneity in temperature - for example in the mountainous highlands of Ethiopia. You resolve to use different sources to check for robustness.
 
 * - *2. Prepare to Download the Data*
-  - CHIRPS data is stored in a [publicly accessible directory](https://data.chc.ucsb.edu/products/CHIRPS-2.0/). You navigate to the `africa_daily/bils/` directory, and choose between 0.5 degree resolution and 2.5 degree resolution. However, you realize that you may have to write a shell script to download this data, to avoid clicking every file separately (using `ftplib` in python and similar packages is also an option).  
+  - CHIRPS data is stored in a [publicly accessible directory](https://data.chc.ucsb.edu/products/CHIRPS-2.0/). You navigate to the `africa_daily/bils/` directory, and choose between 0.5 degree resolution and 2.5 degree resolution. However, you realize that you may have to write a shell script to download this data, to avoid clicking every file separately (using `ftplib` in Python and similar packages is also an option).  
   - Click on 'Get Data (external)' on the Climate Data Guide website, taking you to Berkley Earth's data overview page. You navigate down to the section on 'Gridded Data'. You'll have to click on every decade separately, but without further ado, clean NetCDF files are being downloaded to your machine.
 
 * - *3. Accessing the Data*
@@ -52,7 +52,7 @@ Say you're looking at agriculture in Ethiopia. You would like both temperature a
 Most weather products will require some bureaucracy (creating accounts, signing data agreements, etc.) to download data, and most have their own quirks about how they want data to be downloaded. CHIRPS and BEST do not require bureaucracy, but CHIRPS will require some scripting to download.
 ```
 
-These datasets are stored in different geographical grids and will need to be regridded to a common grid, using tools like `xesmf` in python. See also [weighting schemes](weighting-schemes). 
+These datasets are stored in different geographical grids and will need to be regridded to a common grid, using tools like `xesmf` in Python. See also [weighting schemes](weighting-schemes). 
 
 (content:working-with-era5)=
 ## Getting Started with a Data Product: ERA-5
@@ -93,7 +93,7 @@ Many datasets, especially those from smaller institutions, will not give up thei
 
 ## Thinking Ahead to Climate Projections
 
[…]
