### 2021-02-24 atrisovic/weather-panel.github.io#41 @jrising
Thank you @jwagemann for these questions! Since this is the first response that you've head from me in a while, I wanted to say that we have worked carefully with all of your comments and in many cases they have helped us to rework and greatly improve the tutorial. I'm including all the responses to these points below, but feel free to follow up on them however is convenient for you.

> what do you refer to with ‘first case’ and ‘second case’?

We have dropped the phrases “In the first case” and “In the second case”, since the sentences already described the situations we were concerned with (local vs. regional processes). For example, we have replaced the lines “In the first case, the phenomenon occurs locally, in response to local weather. In this case, we perform weighted aggregations…” with “When the phenomenon occurs locally, in response to local weather, we perform weighted aggregations…”.

> where do weighting data files come from? Which files are these?
> I don’t understand the ‘Kinds of weight schemes and data sources’ section

I will address these together, since they prompted a useful reorganization. There was a section on where to get weighting data at the end of the next section. To make this clearer, we have moved this into its own section, labelled “Where to get spatial weighting data”, above the discussion of how to read the files.

The rest of the content in the “Kinds of weighting schemes and data sources” section offers information about how weighting data is organized. We have divided this into several shorter subsections, to make the structure of the section clearer. These subsections are: “The format of the data values”, “The spatial gridding scheme”, “The geographic projection”, and “Handling of missing data”.

> The example comes somehow out of the blue and it is difficult to follow what you are actually trying to explain

Thank you for this point. We have split the example out from the description of the necessary steps. So, there are now subsections on upsampling, downsampling, and cropping, followed by a section labelled “Example” where these are each used.

> it would be helpful to define what the authors mean with weighting schemes

We have added a description of weighting schemes to the top of the section. This includes the text, “As we use the term, weighting schemes assign a weight to each grid cell or regional observation. They have two major uses: (1) when running regressions, to weight observations to accurately estimate processes, or (2) to perform a weighted aggregation of gridded data to data regions.”

> the topic with projections is introduced with a comic and then not elaborated further which projection is best → hard to understand for a learner

We have added a link for learners to find additional information about map projections. This is split out in a “See also” box with the text “Choosing a map projection is beyond the scope of this tutorial, however, you can take a look at this overview from Jochen Albrecht or how Randall Munroe thinks about them.”
