### 2021-02-24 atrisovic/weather-panel.github.io#38 @jrising
Thank you for these comments! We have extended the introduction considerably and improved the organization based on your suggestions here.

> Adjust the Introduction of the tutorial more with the learning objectives described in the paper

We have incorporated more of the material from the paper here. We have also added learning objectives at the top of the main tutorial sections, in “Key objectives and decision points” boxes, which relate closely to the discussion in the paper.

> Differentiate sections clearer: e.g. introduce a section for audience, learning objectives, how to use this tutorial, etc.

We have added sections on “Who is this tutorial for?”, and “What we hope you will learn”, “How to use this tutorial”, and other helpful online resources.

> Also telling the story - at the moment, the image is introduced, but not explained at all. If you include the image, then also highlight some of the most important findings and integrate the challenges / difficulties associated with this type of research and the reason why one needs this tutorial.

Thank you for pointing this out. We have added additional discussion on some of the interesting features of those graphs, and how they reflect different answers to the questions posed in this tutorial. This is now included in additional paragraphs directly following the figure.
