### 2021-02-24 atrisovic/weather-panel.github.io#40 @jrising
These comments have been handled by @azharhsain, and I'm including his responses below.

Thank you for your helpful comments and close reading of the tutorial. We have provided more information and clarified various points, as detailed below.

To set expectations and objectives right, we have added a new sub-section called “Key objectives and decision points” in the beginning of every section. This could help further guide the readers through the tutorial, especially the ones that are looking for an answer to a specific problem.

> Define ‘weather-panel data’

This has been done by providing a weblink to understand what “panel data” means. Since, the reader of this tutorial is expected to know basic econometrics, we have not explicitly defined it in the tutorial.

> Also a short definition what a reduced-form specification is.

Same as above.

> Under precipitation - what does non-normal mean?

Clearly this was not well-defined before. We have explained it in the text now.

> General section ‘Choosing weather variables’ is not about choosing the variables but more a list of possible variables.

We definitely agree that this section is not only about choosing weather variables, but also introduces different forms in which they can be used. Moreover, the list is indicative and not comprehensive to include all possible weather variables. The aim of this section is to introduce some important and frequently used weather variables and different forms in which they are generally used by researchers. We have now changed the heading to “Choosing weather variables and their forms” to signal not only the choice of weather variables, but also various forms in which they could be used.

> How does this section relate to developing a reduced-form specification?

Since, we require the readers to be equipped with basic econometrics knowledge, they would understand that the variables here would represent candidates for regressors and other control variables. Later, in the tutorial we re-emphasize this in the section “Spatial and temporal scales of economic processes.”

> In general - this section would be more descriptive, if you tell the story with accompanying images, e.g. what is the problem and why do we need transformations and what do the different solutions do?

This is an excellent feedback. We have incorporated this in the example.

> I further recommend to show actual coding examples of the different transformations based on the chosen programming language (e.g. Python or R).

Again, a very good remark. We have added an actual coding example in the revised version of this tutorial.

> Instead of using an image from Carleton et al. (2019), you can show the different outcomes based on the example provided.

This has been included in the example now.

> The same also for section 2.3, show an actual example how cross-validation is done.

This has been included in the example now.
