+++ _sources/content/example-step2.md
@@ -160,7 +160,17 @@ df_mort2 = pd.DataFrame(df_mort.input.apply(
 df_mort3 = df_mort2.apply(pd.to_numeric, errors='coerce')
 
 df_mort4 = df_mort3.groupby(['fips', 'year']).sum()
+df_mort4.head()
+```
+| fips, year   |   deaths |
+|:-------------|---------:|
+| (1001, 1979) |      225 |
+| (1001, 1980) |      221 |
+| (1001, 1981) |      221 |
+| (1001, 1982) |      223 |
+| (1001, 1983) |      267 |
 
+```{code-block} python
 df_pop = pd.read_csv("../data/cmf/Pop7988.txt", names = ['input'])
 
 slices = [(0, 5), (5, 9), (9,18)] + \
@@ -180,13 +190,24 @@ df_pop4 = df_pop3[df_pop3.type == 3]
 df_pop5 = df_pop4.groupby(['fips', 'year', 'type']).sum()
 df_pop5['pop'] = df_pop5.pop1 + df_pop5.pop2 + df_pop5.pop3 + df_pop5.pop4 + df_pop5.pop5 + df_pop5.pop6 + df_pop5.pop7 + df_pop5.pop8 + df_pop5.pop9 + df_pop5.pop10 + df_pop5.pop11 + df_pop5.pop12
 
-df = df_pop5.merge(df_mort4, how='left', on=['fips', 'year'])
+df_pop5.head()
+```
+| fips, year, type |   pop1 |   pop2 |   pop3 |   pop4 |   pop5 |   pop6 |   pop7 |   pop8 |   pop9 |   pop10 |   pop11 |   pop12 |   pop |
+|:----------------|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|--------:|--------:|--------:|------:|
+| (1001, 1979, 3) |   2022 |   2982 |   3248 |   3491 |   2640 |   4414 |   4211 |   3310 |   2457 |    1813 |     779 |     178 | 31545 |
+| (1001, 1980, 3) |   2021 |   2952 |   3184 |   3495 |   2663 |   4463 |   4293 |   3373 |   2487 |    1848 |     795 |     181 | 31755 |
+| (1001, 1981, 3) |   2037 |   2776 |   3132 |   3320 |   2664 |   4646 |   4210 |   3330 |   2516 |    1829 |     824 |     192 | 31476 |
+| (1001, 1982, 3) |   2042 |   2707 |   3098 |   3190 |   2651 |   4714 |   4343 |   3327 |   2565 |    1835 |     856 |     201 | 31529 |
+| (1001, 1983, 3) |   2044 |   2670 |   3054 |   3063 |   2625 |   4815 |   4408 |   3325 |   2613 |    1833 |     882 |     215 | 31547 |
 
+
+```{code-block} python
+df = df_pop5.merge(df_mort4, how='left', on=['fips', 'year'])
 df.to_csv("../data/cmf/merged.csv", header=True)
 ```
 ````
 
-Final dataset (`merged.csv`) should look like:
+The final dataset (`merged.csv`) should look like:
 		
 | fips, year |   pop1 |   pop2 |   pop3 |   pop4 |   pop5 |   pop6 |   pop7 |   pop8 |   pop9 |   pop10 |   pop11 |   pop12 |   pop |   deaths |
 |:-------------|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|--------:|--------:|--------:|------:|---------:|
