### 2021-02-24 atrisovic/weather-panel.github.io#39 @jrising
Thank you for these extensive comments! You will see that we have made extensive changes, including splitting this section up into two easier-to-handle sections. @ks905383 has gone through them and here are his responses.

> make the description of the section more explaining. You introduce a learner to different weather and climate data and also the netcdf format, but a learner without pre-knowledge will have a hard time to understand what you intend to do

We have restructured the introduction to address this point. We now start with explaining the NetCDF format, so that people can better understand what they will be working with.

> instead of saying almost all weather and climate data are released in netcdf, I would say that a common data format used with weather and climate data is netCDF.

This is a good point; we’ve adjusted the language and added more context.

> I don’t understand this sentence: ‘Through this section, when possible relevant commands for working with netCDF files are listed for:’

This section generally has been clarified, and is clearer now with the use of tabs. That line in particular now reads, "Throughout this section, we introduce relevant commands whenever possible for the following languages and packages (click on the tab names for details):".

> include coding examples as ‘code blocks’ instead of integrating them into the sentence or in a table? This makes it easier for a learner to simply copy paste some examples

Thanks for the suggestion; we have adjusted the formatting to be easier to read. We also now use tabs, so that people can quickly get just the code that they need.

> instead of matrices, maybe call them arrays?

Changed.

> The ‘NetCDF content’ description is very abstract for someone who never used netCDF data before. Maybe show the content based on an example data set?

We have now added a full example using an actual weather dataset.

> CMIP data is all of a sudden introduced which might confuse a reader, also don’t think that the naming convention is already appropriate in the place where it is introduced. I would add this when the data is introduced as a ‘Hint’ or ‘Note’ section

We agree; the note on CMIP data is now de-emphasized in a ‘note’ section with a bit more context.

> Diagnostic Maps of Climate data
> this section feels it comes out of the blue and I doubt learners can really follow, because the section before did not introduce how to read the data properly

We have added some more context to the beginning of this section.

> 1.2 Gridded data
> rephrase ‘Weather stations are imperfect’ and explain more in detail and more accurately the challenges with data coming from weather stations

We have added some more context here and referred to later parts of the same section where we deal with some of these biases more explicitly.

> add hyperlinks to data documentations and data organisations

Thank you for the note - this has been updated.

> cross-reference and link content within the tutorial (not only say ‘see content below without linking to it’)

Thank you for the note - this has been updated.

> too many aspects are touched - e.g. location information is mentioned, but then only global climate data is introduced

We have re-organized this section and added more structural text - we hope this adequately addresses this issue. We have also split up the section into two pages, to make them more digestible.

> What is the main learning outcome this module shall have?

The introduction to the module now begins with learning goals and outcomes.
