### 2021-02-24 atrisovic/weather-panel.github.io#42 @jrising
Thank you for these comments! @atrisovic has gone through them, with responses below.

> maybe use a table to describe administrative units

Thanks for the suggestion! A table is added to illustrate the administrative units.

> the section does not tell me why it is important to generate geographical unit data

An explanation is added in the opening paragraph.

> here examples of different programming languages are shown again and I'd say it would be better to stick to one example in one programming language and to refer to packages in other programming languages

As folks in the field frequently use Stata, R and Python, we decided to keep different languages throughout the tutorial. However, this should now be much ‘cleaner’ and easier to understand, as we make use of tabs, which is a new feature of Jupyter Book.
