### 2021-02-24 atrisovic/weather-panel.github.io#43 @jrising
@atrisovic has also provided responses on this section. Let us know if you have more suggestions.

> do not only provide the comic without explanation for reasons why Git is good, provide a list of bullet points yourself why you use git and how it benefits projects

An enumerated list has now been added.

> 5.4 workflow automation
> if you make this as a part of your tutorial, use an actual example and show how the workflow automation can be implemented and also list advantages why one would want to do so

We have now developed a common example throughout the tutorial, split out into steps for each of the main sections of the tutorial. While we think this is a significant improvement in the tutorial, we have not provided a single script to readers along the lines suggested here. This is because we want to emphasize the unique decision points researchers need to make (now listed at the tops of the main sections) and want to avoid people taking our example process as a whole. The ‘suggestions’ page also now includes more diverse (and unrelated) content regarding good practices.
