### 2021-02-24 #90 @jrising
Thank you @labarba for the prompt! We have been going through the suggestions and thinking broadly about how to improve the structure of the tutorial in light of these comments. I have just replied to the full list of helpful comments. Here are responses to the points above from @jwagemann.

> at the current state, it fails to meet the aim to be a step-by-step guide on how to carry out climatic econometrics. Sections often seem arbitrary and for a learner or external person it is often hard to understand what the section aims to describe or to do

We have made several improvements to address this. We added "learning objectives and decision points" boxes to the top of each main section of the tutorial. We added a consistent "hands-on exercise" to show how the various points can be put together. And we have made a number of improvements in the organization of sections, splitting up sections to make the key points clearer.

> Many parts are not easy to follow, as authors try to capture too many aspects. Also many headlines seem a bit out of place, e.g the headline ‘Geographic Information Systems’ is used when the shapefile format is explained

This comment drove us to reconsider the structure of each section. In many cases,we have split up a large section (such as the one relating to GIS concepts) into multiple smaller sections. For example, we first introduce GIS concepts, then have a section on how to get shapefiles, and then a section on how to use them.

> at the current state, the tutorial cannot be used by anyone other than the authors in order to teach students on the important topic of climatic econometrics

We have tried to make the tutorial much more accessible, and have shared it with more people. Crucially, we have also been more explicit about the background knowledge (mainly on econometrics) that we are expecting readers to have (in a section "Who is this tutorial for?"), while also including more references to other resources that will help readers get needed knowledge throughout the tutorial.

> The smart use of images and explanatory media is missing. The tutorial is heavily based on text and hardly works with additional images that help to understand the complex content

While the tutorial is still quite "wordy", we have tried to make it more accessible in a few ways. First, we have made better use of Jupyter Book features like tabs (used now to split out different programming languages) and colored boxes (used to reference additional resources and highlight notes). We have also included a "Hands-On Exercise" that will allow readers to engage directly with the concepts in the tutorial. Finally, we have added more explanatory material to go with our images.

> The general writing needs to be spell-checked and the general tone feels somehow a bit sloppy: e.g. ‘here are some suggestions for dealing with the mess that is political geography’ . I would not use the word ‘mess’ in any type of tutorial.

We apologize for these errors. We have done a more thorough editing process, and we have also passed the tutorial content through spell-checking and readability tools. We have also gotten comments back from other users, and adjusted the language to address their confusion.

> The tutorial does not follow a consistent instructional design. Some sections are accompanied with code examples, others explain the content only in a very abstract way.

We hope that our inclusion of a common example will help make even the very general parts of the text more concrete. We have not used the same level of coding examples throughout the text, because of the wide range of challenges we hope that our tutorial can help students address: some challenges are conceptual, some informational, some mathematical, and some coding. We have worked to make the tutorial as a whole more consistent however.

> Coding examples are either in R or try to show examples in multiple programming languages, which is quite confusing for the learner. I recommend to focus only on one programming language and provide only links to similar packages in other programming languages

It is quite common in this field to use multiple languages within a single project, and R, python, and Stata are all in common use amongst our audience. For this reason, we have kept the multiple languages, but we have improved how they are presented. In the previous version, they were included in the text in different ways. Now, we consistently use Jupyter Book tabs, to allow readers to quickly get the example in the language they need.

> The tutorial misses an accompanying coding script in one of the programming languages. Without a coding script, it is very hard that a learner will be able to implement the content explained

We have now added this, in the form of a common hands-on exercise. The exercise is split into four steps, following the structure of the tutorial, and takes readers from downloading the necessary data through to getting out a dose-response function similar to the ones shown on the first page of the tutorial.
