+++ tutorial-content/content/getting-started.md
@@ -33,17 +33,17 @@ aimed at helping you make decisions appropriate to your research.
 # Who is this tutorial for?
 
 This tutorial is created for students and researchers new to climate
-econometrics. However, it assumes a knowledge of panel econometrics. You
+econometrics. However, it assumes a knowledge of panel econometric methods. You
 should be familiar with specifications that factor-out unobserved
-heterogeneity, implementing flexible trends, weighted regressions,
+heterogeneity, flexible trends, weighted regressions,
 approaches to deal with heteroskedasticity and correlated errors, and
-diagnosing efficiency problems. Some concepts in panel econometrics
+diagnostics for efficiency problems. Some concepts in panel econometrics
 are more specific to weather regressions, such as dose-response
 curves, data-generating processes, and neighborhood effects, and we
 will cover those in this tutorial.
 
 We also assume basic experience with one scientific programming
-language (Stata, R, Matlab, Julia, Python). In most examples, we provide code snippets in more than one programming language, so you can get started.
+language (Stata, R, Python, Matlab, Julia). In most examples, we provide code snippets in more than one programming language, so you can get started.
 
 You should also already have a research question. There are plenty of
 ways to find important questions, and maybe we will try to offer a
@@ -71,12 +71,20 @@ tutorial is aimed at helping you do that.
 The tutorial is organized along the rough outline of a research
 project: 
 
-1. *Using weather and climate data*: introduces the data and its attributes, the NetCDF file format, supported programming languages, and common limitations of the data.
+1. *Weather and climate data*: introduces environmental data and its attributes, the NetCDF file format, supported programming languages, and common limitations of the data.
 2. *Developing a reduced-form specification*: provides a number of considerations before starting an analysis, such as choosing weather variables, functions for creating a model, and caveats when working with spatial and temporal processes.
-3. *Weighting schemes*: explains the importance of weighting schemes and how to work with them based on their file formats and origins. 
-4. *Generating geographical unit data*: introduces geographic information systems, shapefiles, and how to work with them effectively.
+3. *Geographical unit data*: introduces geographic information
+   systems, shapefiles, and how to work with them effectively, and explains the importance of weighting schemes and how to work with them based on their file formats and origins. 
+4. *Bringing it all together*: offers a range of suggestions for
+   organizing your work and producing final results.
 
-You can go through the sections one at a time or reference them independently.
+You can go through the sections one at a time or reference them
+independently.
+
+We also include a Hands-On Exercise, which takes you through each of
+these steps, from collecting data to producing a final weather
+relationship. Please note that this is just one example, and not a
+template for other research projects.
 
 Most importantly, if you learn something in your own research process
 that you think would be of broad interest, see [the contributor's guide](https://github.com/atrisovic/weather-panel.github.io) for how to include it here. Help us make this tutorial more
@@ -84,19 +92,23 @@ useful for more people!
 
 # Useful introductory resources
 
+Here are some resources to read to understand climate
+econometrics. This tutorial is a kind of practical counterpart to
+these excellent documents.
+
 📚  [An Economist’s Guide to Climate Change Science](https://www.aeaweb.org/articles?id=10.1257/jep.32.4.3)
 : is a useful resource to better understand the basics of weather, climate,
 and the physical changes occurring in the climate system. If
 you have not had experience with climate (or meteorological) science,
 that is a great place to start. 
 
-📚  Check out the [Encyclopedia of Earth](https://editors.eol.org/eoearth/wiki/Weather_%26_Climate) and the [IPCC WGI report](https://www.ipcc.ch/site/assets/uploads/2017/09/WG1AR5_Chapter01_FINAL.pdf)
[…]
