+++ requirements.txt
@@ -26,7 +26,7 @@ ipython==7.19.0
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
 jedi==0.17.2
-Jinja2==2.11.2
+Jinja2==2.11.3
 jsonschema==3.2.0
 jupyter-book==0.8.3
 jupyter-cache==0.4.1
