+++ requirements.txt
@@ -63,7 +63,7 @@ pyparsing==2.4.7
 pyrsistent==0.17.3
 python-dateutil==2.8.1
 pytz==2020.4
-PyYAML==5.3.1
+PyYAML==5.4
 pyzmq==19.0.2
 requests==2.25.0
 Send2Trash==1.5.0
