### 2021-03-30 a3d5571
Bump pygments from 2.7.2 to 2.7.4
Bumps [pygments](https://github.com/pygments/pygments) from 2.7.2 to 2.7.4.
- [Release notes](https://github.com/pygments/pygments/releases)
- [Changelog](https://github.com/pygments/pygments/blob/master/CHANGES)
- [Commits](https://github.com/pygments/pygments/compare/2.7.2...2.7.4)

Signed-off-by: dependabot[bot] <support@github.com>
