+++ requirements.txt
@@ -58,7 +58,7 @@ pybtex==0.23.0
 pybtex-docutils==0.2.2
 pycparser==2.20
 pydata-sphinx-theme==0.4.1
-Pygments==2.7.2
+Pygments==2.7.4
 pyparsing==2.4.7
 pyrsistent==0.17.3
 python-dateutil==2.8.1
