+++ requirements.txt
@@ -90,7 +90,7 @@ terminado==0.9.1
 testpath==0.4.4
 tornado==6.1
 traitlets==5.0.5
-urllib3==1.26.4
+urllib3==1.26.5
 wcwidth==0.2.5
 webencodings==0.5.1
 widgetsnbextension==3.5.1
