### 2021-09-16 #90 @labarba
hi @howardjp, @jwagemann 👋  — It looks like the author of this JOSE submission made improvements and they are waiting for you to take a second look. Can you let us know when you might be able to do so? Thanks!
