### 2021-10-21 7f00618
Bump babel from 2.8.1 to 2.9.1
Bumps [babel](https://github.com/python-babel/babel) from 2.8.1 to 2.9.1.
- [Release notes](https://github.com/python-babel/babel/releases)
- [Changelog](https://github.com/python-babel/babel/blob/master/CHANGES)
- [Commits](https://github.com/python-babel/babel/compare/v2.8.1...v2.9.1)

---
updated-dependencies:
- dependency-name: babel
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
