+++ requirements.txt
@@ -3,7 +3,7 @@ appnope==0.1.0
 argon2-cffi==20.1.0
 async-generator==1.10
 attrs==20.3.0
-Babel==2.8.1
+Babel==2.9.1
 backcall==0.2.0
 beautifulsoup4==4.9.3
 bleach==3.3.0
