### 2021-11-07 c3b1943
New script to calculate a global temparature average
