### 2021-11-08 75873a7
Bump nbdime from 2.1.0 to 2.1.1
Bumps [nbdime](https://github.com/jupyter/nbdime) from 2.1.0 to 2.1.1.
- [Release notes](https://github.com/jupyter/nbdime/releases)
- [Commits](https://github.com/jupyter/nbdime/compare/nbdime-jupyterlab@2.1.0...nbdime-jupyterlab@2.1.1)

---
updated-dependencies:
- dependency-name: nbdime
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
