+++ requirements.txt
@@ -41,7 +41,7 @@ myst-nb==0.10.1
 myst-parser==0.12.10
 nbclient==0.5.1
 nbconvert==5.6.1
-nbdime==2.1.0
+nbdime==2.1.1
 nbformat==5.0.8
 nest-asyncio==1.4.3
 notebook==6.4.1
