### 2022-01-05 90e0694
Update reduced-form-specification.md
