### 2022-01-21 9a79911
Bump ipython from 7.19.0 to 7.31.1
Bumps [ipython](https://github.com/ipython/ipython) from 7.19.0 to 7.31.1.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/7.19.0...7.31.1)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
