### 2022-04-12 #90 @jwagemann
Hi @kyleniemeyer ,
I finally had a chance to review the reviewed tutorial and I congratulate @jrising , @atrisovic and the other authors for producing this great resource. It is now well-structured and guides learners nicely through the content.

I recommend to accept. One small proposal of change: Under ERA5, instead of linking to UCAR’s ERA5 description, link to the description on the Climate Data Store: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=overview
