+++ requirements.txt
@@ -44,7 +44,7 @@ nbconvert==5.6.1
 nbdime==2.1.1
 nbformat==5.0.8
 nest-asyncio==1.4.3
-notebook==6.4.10
+notebook==6.4.1
 oset==0.1.3
 packaging==20.4
 pandocfilters==1.4.3
