### 2022-06-16 d736198
Bump notebook from 6.4.1 to 6.4.12
Bumps [notebook](http://jupyter.org) from 6.4.1 to 6.4.12.

---
updated-dependencies:
- dependency-name: notebook
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
