+++ requirements.txt
@@ -36,7 +36,7 @@ jupyter-sphinx==0.3.2
 latexcodec==2.0.1
 markdown-it-py==0.5.6
 MarkupSafe==1.1.1
-mistune==0.8.4
+mistune==2.0.3
 myst-nb==0.10.1
 myst-parser==0.12.10
 nbclient==0.5.1
