+++ requirements.txt
@@ -40,7 +40,7 @@ mistune==0.8.4
 myst-nb==0.10.1
 myst-parser==0.12.10
 nbclient==0.5.1
-nbconvert==5.6.1
+nbconvert==6.3.0
 nbdime==2.1.1
 nbformat==5.0.8
 nest-asyncio==1.4.3
