### 2022-08-23 2fc5b7f
Bump nbconvert from 5.6.1 to 6.5.1
Bumps [nbconvert](https://github.com/jupyter/nbconvert) from 5.6.1 to 6.5.1.
- [Release notes](https://github.com/jupyter/nbconvert/releases)
- [Commits](https://github.com/jupyter/nbconvert/compare/5.6.1...6.5.1)

---
updated-dependencies:
- dependency-name: nbconvert
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
