### 2022-12-08 7fa5c2c
Bump certifi from 2020.11.8 to 2022.12.7
Bumps [certifi](https://github.com/certifi/python-certifi) from 2020.11.8 to 2022.12.7.
- [Release notes](https://github.com/certifi/python-certifi/releases)
- [Commits](https://github.com/certifi/python-certifi/compare/2020.11.08...2022.12.07)

---
updated-dependencies:
- dependency-name: certifi
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
