+++ requirements.txt
@@ -7,7 +7,7 @@ Babel==2.9.1
 backcall==0.2.0
 beautifulsoup4==4.9.3
 bleach==3.3.0
-certifi==2020.11.8
+certifi==2022.12.7
 cffi==1.14.3
 chardet==3.0.4
 click==7.1.2
