+++ requirements.txt
@@ -22,7 +22,7 @@ idna==2.10
 imagesize==1.2.0
 importlib-metadata==2.0.0
 ipykernel==5.3.4
-ipython==7.31.1
+ipython==8.10.0
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
 jedi==0.17.2
