### 2023-02-23 9b9bdd0
Bump markdown-it-py from 0.5.6 to 2.2.0
Bumps [markdown-it-py](https://github.com/executablebooks/markdown-it-py) from 0.5.6 to 2.2.0.
- [Release notes](https://github.com/executablebooks/markdown-it-py/releases)
- [Changelog](https://github.com/executablebooks/markdown-it-py/blob/master/CHANGELOG.md)
- [Commits](https://github.com/executablebooks/markdown-it-py/compare/v0.5.6...v2.2.0)

---
updated-dependencies:
- dependency-name: markdown-it-py
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
