+++ requirements.txt
@@ -34,7 +34,7 @@ jupyter-client==6.1.7
 jupyter-core==4.6.3
 jupyter-sphinx==0.3.2
 latexcodec==2.0.1
-markdown-it-py==0.5.6
+markdown-it-py==2.2.0
 MarkupSafe==1.1.1
 mistune==0.8.4
 myst-nb==0.10.1
