### 2023-06-06 atrisovic/weather-panel.github.io#76 @kls2177
JOSE Review - comments on Introduction
Nicely laid out. Just a few comments:
- The link to the IPCC report should be updated to AR6 (I realize that AR5 was the most up-to-date when this review started). You can link to the entire WGI landing page rather than a specific chapter. You may also consider adding a link to the WGII report (not just WGI) as this may be more relevant to your readers. Although, I agree that some WGI background knowledge is essential.
- For the region data section, do you want to add an example, e.g. census tract.
- Mathematical notation: as a climate scientist, I would prefer that T not be used as the dummy variable name as it may make the reader think of temperature. What about X? I haven't read the entire tutorial yet, so maybe the choice will make more sense later.
