### 2023-06-16 atrisovic/weather-panel.github.io#77 @ks905383
Thank you for the helpful comments! Will take a look at this shortly.
