### 2023-06-16 atrisovic/weather-panel.github.io#77 @kls2177
The full review is taking me a bit longer than I thought, so I probably won't be finished until next week. Sorry about that!
