### 2023-06-20 atrisovic/weather-panel.github.io#78 @kls2177
Now that I have read the entire tutorial, I have one other comment about the Hands-On Exercise, Step 2. You introduce the linear and quadratic forms of the temperature data in this section, but the actual exercise is downloading the population and mortality data. There seems to be a bit of a disconnect here. I think it would be helpful to provide a bit more narrative - let the students know that you will come back to the temperature transformations in Hands-On Exercise, Step 3.
