### 2023-06-20 atrisovic/weather-panel.github.io#79 @kls2177
JOSE Review - comment on Geographical Unit Data section
This section provides a good introduction to the different formats of socio-economic data, but without too much unnecessary detail. My main comment here is simply the sequencing of material. The Hands-On Exercise, Step 2 makes use of geographical unit data, but this data is not discussed until after. I realize that you want the structure of the tutorial to be such that there is an Exercise at the end of each Chapter, but Step 2 seems out of place. Please consider how you might restructure the Exercises to better align with the flow of the content.

===============================
Section: Generating Geographical Unit Data

- first paragraph: remove "and" -> "collected corresponding to..."
- See Also box: link just goes to the current issue -> needs to be revised
- Note that some of the packages for GIS-type data contain the word "raster". Perhaps you should introduce this term in the GIS data section as well.
- update GDAL Georeferencer link -> there is a note on the website indicating the the tutorial is now obsolete.
- Caution Box: would have been good to know this for Hands-On Exercise, Step 2.

==============================
Section: Constructing weather averages within spatial units

Sub-section: Approach 1
- I am not an R user, so I don't really know what you are doing without doing a lot of googling. Is it possible to add the python (or matlab) implementation? Also it would help if you defined some of your variables: longitude0, longitude1, gridwidth, etc.
- paragraph after second code block: "upright" -> "up right"

Sub-section: Approach 2
- can you show a diagram of this? I think this would help students visualize what's going on. There are some similar ones online. I don't think this is exactly the right image, but maybe something similar to this: https://www.esri.com/arcgis-blog/products/spatial-analyst/analytics/getting-the-most-out-of-zonal-statistics/

=============================
Section: Weighting Schemes

-....will come back to this later...
