### 2023-06-21 atrisovic/weather-panel.github.io#80 @kls2177
JOSE Review - comment on Bringing it all together
This Chapter nicely wraps up the tutorial. Just a few comments below.

Section: Producing Results

This section is out of my area, so I only have a few comments.

- could you potentially create a more descriptive title. "Producing Results" is quite generic
- I don't quite understand the learning outcome -> who is "offering"? The verb should refer to what the student will be able to do by the end of the chapter.
- is there a python implementation for the error term? I found something that might be relevant on stackoverflow: https://stackoverflow.com/questions/30553838/getting-statsmodels-to-use-heteroskedasticity-corrected-standard-errors-in-coeff. Perhaps the statsmodels or linearmodels packages might have relevant functionality.

Section: Hands-On Exercise, Step 4
- note that I wasn't able to test the code in this section because I had an error in Step 3.
- Again, I would appreciate a python implementation of this section. Perhaps, using pandas and then one of the linearmodels packages would be the most straightforward.
- in the "Running the regression" sub-section, what do you mean by "state trends"? Why do you use state trends versus county trends? Why have county fixed effects but not state fixed effects? It would be helpful if you explained your choices. Perhaps, you could build in some sensitivity tests for students to do on their own to justify the choices - for example, what happens to the results if they don't include FEs or use different ones? What about the linear and quadratic temperature terms? Does having both improve the results? How do they know which model is more robust (not sure if you use an adjusted R^2 for this type of thing)?
- after the figure, it might be useful to describe what is shown. How do you want students to interpret the figure? (Don't forget to revise C to deg C in the figure x-axis as well).

Section: Suggestions for work organization
- second note: I don't think you introduced the term "vector data" in the pervious GIS section.
