### 2023-06-21 #90 @kls2177
Hello @kyleniemeyer, I apologize for the delay with my review.

The tutorial, "ClimateEstimate.net - A practical guide to climate econometrics", is a valuable contribution to the econometrics community. Thank you to the authors - I learned a lot while reading it and feel that most relevant topics are covered. I was able to work through the most of the Hands-On Exercises, but had some trouble in Step 3 (see the links to my issues below) and consequently, Step 4.

I feel that this tutorial is suitable for publication after some revision. I recommend that there be more examples and/or activities in order to make it a more valuable learning tool. I also agree with the previous reviewer that more visuals would help students better engage with certain concepts and would also help to break up the text. Also, as a python-user, given that python is used quite a bit in the tutorial, I feel that providing python code examples for ALL the steps would be very helpful. I am not that familiar with R.

Finally, although the tutorial seems relatively complete for someone who already has a background in econometrics (this is key!), I feel that the guide would benefit from integrating the Hands-On Exercises into the narrative of the content to a greater extent. Let the Exercises tell the story. This is just a suggestion.

Major Comments:
- The first section of Chapter 1 is much too long and needs to be broken up
- The narrative around the Hands-On Exercises need to be spelled out in Step 1. Otherwise, Step 1 just seems like a "make-work" exercise. Let students know that the exercises are scaffolded and have a particular end goal in mind.
- The first Section of Chapter 2 is mostly a list of variables and their descriptions, most of which could be put in an appendix (or this section needs to be revised)
- Hands-On Exercise, Step 2 does not align well with the Chapter content
- I had problems with Hands-On Exercise, Step 3 (see below)
- Hands-On Exercise, Step 4 could benefit from exploring sensitivity to reduced form specification, FEs, etc.

Here are my detailed comments:

[Introduction](https://github.com/atrisovic/weather-panel.github.io/issues/76)
[Chapter 1](https://github.com/atrisovic/weather-panel.github.io/issues/77)
[Chapter 2](https://github.com/atrisovic/weather-panel.github.io/issues/78)
[Chapter 3](https://github.com/atrisovic/weather-panel.github.io/issues/79)
[Chapter 4](https://github.com/atrisovic/weather-panel.github.io/issues/80)

Thank you very much to the authors for your patience! I would be happy to review a revised version of the tutorial.
