### 2023-06-21 #90 @jrising
Thank you @kls2177 for the extensive and detailed comments! These will be very useful for improving the tutorial. We will work through them over the course of the next few weeks, and we will respond first within each github Issue.
