### 2023-06-29 2cc9796
fixes for issue 76
