### 2023-06-29 atrisovic/weather-panel.github.io#79 @ks905383
Looks like this points to a bug in [xagg](https://github.com/ks905383/xagg) as well. I can start an issue there and hopefully take a stab at it.
