### 2023-07-03 atrisovic/weather-panel.github.io#79 @ks905383
> couldn't get the conda install of xagg to work. Used pip but had to also add xesmf.
> got a "warning" for the weightmap = xa.pixel.overlaps() step (see attached screenshot)
> got an error for the xa.aggregate step (see attached screenshot)

This is now fixed in `xagg` v0.3.1, which is the latest version on conda-forge. Currently `conda` and `mamba` still want to download v0.3.0.2 (which has a few bugs) unless explicitly told `xagg==0.3.1`. Perhaps we should make a note of that - I don't think we currently have an environment creation step in the tutorial anyway?
