### 2023-07-05 de4184e
fixes for issue 80
