### 2023-07-05 atrisovic/weather-panel.github.io#76 @jrising
@kls2177 Thank you for the comments, across all of our sections! I have edited the Introduction under this commit:https://github.com/atrisovic/weather-panel.github.io/commit/2cc9796484c38addc214080081f0c20dc039e908

Responding to your comments individually:
 - Thanks for the note about AR6-- we definitely want to keep this up to date. I've linked to the AR6 WGI and WGII chapters, but I also included a comment pointing readers to WGI Chapter 1, since that provides the background information we think readers will particularly benefit from at this point in the tutorial.
 - I added examples for each region type: "political units (e.g. municipalities), statistical units (e.g., census tracts), and jurisdictions (e.g., educational districts)".
 - I switched this to X. It's somewhat common in the climate-economics field to use T as a proxy for any weather variable, since temperature is often the one that is of most interest, but I can see the potential for confusion. I clarified that we will use T and P for temperature and precip. elsewhere in the tutorial. We will also update the comments on functional forms to use X rather than T as our generic variable.
