### 2023-07-07 32657a2
Address comments from #77, other updates
