+++ tutorial-content/content/weather-and-climate-data.md
@@ -1,4 +1,4 @@
-# Using Weather and Climate Data
+# Introduction to Weather and Climate Data
 
 ```{admonition} Key objectives and decision points
 :class: note
@@ -13,6 +13,10 @@ Decision points:
 
  - How to choose climate or weather data to use in your research project?
 ```
+```{figure} https://www.earthdata.nasa.gov/s3fs-public/2023-01/product-distribution-volume-discipline-2.jpg?VersionId=Tor97BJIz5dyuZofS5swA7RGwdccByVe 
+The growing amount of data distributed by NASA's EOSDIS cloud ([source](https://www.earthdata.nasa.gov/esds/esds-highlights/2022-esds-highlights)). We're live through a golden age of climate data use creation. See also the classic [xkcd](https://xkcd.com/2582/). 
+
+```
 
 This section will introduce you to the right questions to ask when deciding on climate or weather data to use in your research. It will cover how to deal with a commonly used weather and climate data format in multiple languages. It will present the differences between gridded weather data products, evaluations of weather data products, and examples of downloading several different weather data products, and a few warnings on common biases in weather data, especially precipitation. 
 
@@ -22,467 +26,8 @@ When using weather data as independent variables in an economic model, or climat
 - Every weather or climate data product has its use cases, limitations, uncertainties, and quirks.
 
 ```{seealso}
-[Auffhammer et al. (2013)](https://core.ac.uk/download/pdf/189626825.pdf) provide an excellent discussion of the challenges to consider when using weather data in econometric studies.
+[Auffhammer et al. (2013)](https://doi.org/10.1093/reep/ret016) provide an excellent discussion of the challenges to consider when using weather data in econometric studies.
 ```
 
-## The NetCDF Data Format
 
-We start this section with a guide to the [NetCDF](https://climatedataguide.ucar.edu/climate-data-tools-and-analysis/NetCDF-overview) format, a common data format used for weather and climate data. Most weather and climate datasets will be published primarily or additionally in the NetCDF format. It's efficient, self-describing, and supported in major programming languages, though you’ll have to pre-process data into another format (`.csv`, etc.) before you can use it in STATA. If you get familiar with the commands to read the header and access data in the language you’re most comfortable with, you will be able to work with most existing climate or weather datasets.
 
-### Your code environment
-
-Throughout this section, we introduce relevant commands whenever
-possible for the following languages and packages (click on the tab
-names for details):
-
-````{tabbed} Python (xarray)
-
-[`xarray`](http://xarray.pydata.org/en/stable/) (recommended) is a package for working with N-dimensional data that natively supports NetCDF files.
-
-For any Python (xarray) code chunks, it's assumed that the `xarray` package is loaded as `xr`: 
-
-```{code-block} python
-import xarray as xr
-```
-````
-
-````{tabbed} Matlab
-MATLAB has native support for working with N-dimensional data.
-````
-
-````{tabbed} R
-Support through the [ncdf4](https://cran.r-project.org/web/packages/ncdf4/index.html) package.
-
-For any R code chunks, it's assumed the `ncdf4` package is loaded with: 
-
-```{code-block} R 
-library(ncdf4)
-```
-````
-
-````{tabbed} Python (NetCDF4)
-Support through the [netCDF4](https://unidata.github.io/netcdf4-python/netCDF4/index.html) module.
-
-For any Python (NetCDF4) code chunks, it's assumed that the `NetCDF4` package is loaded as `nc`:
-
-```{code-block} python
-import netCDF4 as nc
-```
-````
-
-````{tabbed} nco
[…]
