### 2023-07-11 2b1fc4f
refactor tabs and panels
