+++ tutorial-content/content/weighting-schemes.md
@@ -119,7 +119,8 @@ All of these allow missing data to be handled. Typically, a specific numerical r
 
 ### Implementation Notes: Reading gridded data.
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 Use the `raster` library. For example:
 
 ```{code-block} R
@@ -132,9 +133,10 @@ If you are using R, take a look at the [Introduction to Geospatial Raster and Ve
 which has some extensive examples of working with geospatial raster data.
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 Take a look at <https://github.com/jrising/research-common/tree/master/python/geogrid>.
 ````
+`````
 
 In some cases, it is appropriate and possible to use time-varying weighting schemes. For example, if population impacts are being studied, and the scale of the model is individuals, annual estimate of population can be used. This kind of data is often either in NetCDF format (see above), or as a collection of files.
 
@@ -145,10 +147,8 @@ conform to data grid definition used by your weather data.  This
 generally requires resampling the weighting data, increasing its
 resolution by some factor. You want to do this so that two conditions to be met after resampling: (A) The new resolution should be an integer multiple of the weather resolution. (B) The horizontal and vertical grid lines of the weather data coincide with the resampled grid lines of the weighting data.
 
-Here we
-assume that both are regular latitude-longitude
-grids. See
-[working with gridded weighting data](content:working-with-gridded-data) to
+Here we assume that both are regular latitude-longitude
+grids. See [working with gridded weighting data](content:working-with-gridded-data) to
 understand the grid scheme for your weighting file; note that gridded
 weather data often reports the center of each grid cell, rather than
 the corner.
@@ -166,28 +166,31 @@ upsampling, downsampling, and cropping.
 To increase the resolution of a grid `rr` by a factor of `N` without increasing the
 sum of grid cells:
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
 rr2 <- disaggregate(rr, fact=N) / (N^2)
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 ``` python
 rr2 = np.repeat(np.repeat(rr, N, axis=0), N, axis=1) / (N*N)
 ```
 ````
+`````
 
 ### Downsampling - decreasing data resolution
 
 To decrease the resolution of a grid `rr` by a factor of `N` without
 decreasing the sum of grid cells:
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
 rr2 <- aggregate(rr, fact=N, fun=sum)
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 ``` python
 rr2 = np.zeros((int(rr.shape[0] / N), int(rr.shape[1] / N)))
 for ii in range(N):
@@ -195,6 +198,7 @@ for ii in range(N):
     rr2 += rr[ii::N, ii::N] / (N*N)
 ```
 ````
+`````
 
 ### Cropping - adjusting the extent of data
 
@@ -202,18 +206,20 @@ To adjust the spatial extent of grid `rr` to conform as closely as possible
 to a box from the longitudinal range `WW` to `EE` and the latitudinal
 range from `SS` to `NN`.
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
[…]
