### 2023-07-11 94a9087
jose feedback: typos
Ref: #79
