+++ tutorial-content/content/weighting-schemes.md
@@ -118,8 +118,8 @@ need to project these back to latitude-longitude and regrid them.
 All of these allow missing data to be handled. Typically, a specific numerical representation, like -9999, will be used. This is specified the same way that the gridding scheme is.
 
 ### Implementation Notes: Reading gridded data.
-
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 Use the `raster` library. For example:
 
 ```{code-block} R
@@ -132,9 +132,10 @@ If you are using R, take a look at the [Introduction to Geospatial Raster and Ve
 which has some extensive examples of working with geospatial raster data.
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 Take a look at <https://github.com/jrising/research-common/tree/master/python/geogrid>.
 ````
+`````
 
 In some cases, it is appropriate and possible to use time-varying weighting schemes. For example, if population impacts are being studied, and the scale of the model is individuals, annual estimate of population can be used. This kind of data is often either in NetCDF format (see above), or as a collection of files.
 
@@ -166,28 +167,31 @@ upsampling, downsampling, and cropping.
 To increase the resolution of a grid `rr` by a factor of `N` without increasing the
 sum of grid cells:
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
 rr2 <- disaggregate(rr, fact=N) / (N^2)
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 ``` python
 rr2 = np.repeat(np.repeat(rr, N, axis=0), N, axis=1) / (N*N)
 ```
 ````
+`````
 
 ### Downsampling - decreasing data resolution
 
 To decrease the resolution of a grid `rr` by a factor of `N` without
 decreasing the sum of grid cells:
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
 rr2 <- aggregate(rr, fact=N, fun=sum)
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 ``` python
 rr2 = np.zeros((int(rr.shape[0] / N), int(rr.shape[1] / N)))
 for ii in range(N):
@@ -195,26 +199,27 @@ for ii in range(N):
     rr2 += rr[ii::N, ii::N] / (N*N)
 ```
 ````
-
+`````
 ### Cropping - adjusting the extent of data
 
 To adjust the spatial extent of grid `rr` to conform as closely as possible
 to a box from the longitudinal range `WW` to `EE` and the latitudinal
 range from `SS` to `NN`.
 
-````{tabbed} R
+`````{tab-set}
+````{tab-item} R
 ```{code-block} R
 rr2 <- crop(rr, extent(WW, SS, EE, NN))
 ```
 ````
-````{tabbed} Python
+````{tab-item} Python
 ``` python
 # We need to know the original extent of the data (W0, S0, E0, N0) and the spatial resolution (D)
 
 rr2 = rr[int((SS - S0)/D):int((NN - S0)/D), int((WW - W0)/D):int((EE - W0)/D)]
 ```
 ````
-
+`````
 The following recipe should work for most cases to align weighting data with a weather grid.
 
 1. **Upsample the weighting data until the grid of the weighting data evenly divides up the weather data.** Start by considering the southwest corner of each dataset. Even if the
@@ -234,7 +239,8 @@ exact same extent.
 
[…]
