+++ tutorial-content/content/gridded-data.md
@@ -83,7 +83,7 @@ Much of our observations for entire classes of variables (clouds, land surface c
 
 
 ### Temporal and Spatial Homogeneity
-The number and locations of weather stations, the technologies used, and the methods for processing data change over time. The creators of gridded datasets will have chosen a strategy for dealing with these differences. Some datasets, such as the [IMERG](https://climatedataguide.ucar.edu/climate-data/gpm-global-precipitation-measurement-mission) rainfall dataset, provide high resolution data but at the expense of an inhomogenous data record, by using all satellites they can get data from at a given time. Other datasets (such as the temperature uncertainty ensemble of [GISTEMP](https://climatedataguide.ucar.edu/climate-data/global-surface-temperature-data-gistemp-nasa-goddard-institute-space-studies-giss)) emphasize temporal homogeneity to better study long-term trends. 
+The number and locations of weather stations, the technologies used to collect data, and the methods for processing data change over time. The creators of gridded datasets will have chosen a strategy for dealing with these differences. Some datasets, such as the [IMERG](https://climatedataguide.ucar.edu/climate-data/gpm-global-precipitation-measurement-mission) rainfall dataset, provide high resolution data but at the expense of an inhomogenous data record, by using all satellites they can get data from at a given time. Other datasets (such as the temperature uncertainty ensemble of [GISTEMP](https://climatedataguide.ucar.edu/climate-data/global-surface-temperature-data-gistemp-nasa-goddard-institute-space-studies-giss)) emphasize temporal homogeneity to better study long-term trends. 
 
 
 Next, we will get into how to [choose and work with weather data products](content:working-with-data) - but before that, we'd like to leave you with two warnings on using [hydrological variables](content:warning-on-hydrological) and using [station data](content:station-data).
