### 2023-07-17 0023fac
add note on projections
