### 2023-07-17 8a8299d
add instructions to print header in hands-on
