+++ _sources/content/gridded-data.md
@@ -1,0 +1,132 @@
+(content:gridded-data)=
+# Working with Gridded Data
+## Gridded Data
+
+Weather data is traditionally collected at weather stations. Weather stations are imperfect, unevenly distributed point sources of data whose raw output may not be suitable for economic and policy applications. Weather readings may be biased - for example, rain gauges tend to [underestimate](https://journals.ametsoc.org/view/journals/apme/58/10/jamc-d-19-0049.1.xml) peak rainfall, and air temperature sensors often become [more inaccurate](https://journals.ametsoc.org/view/journals/atot/21/7/1520-0426_2004_021_1025_saeeia_2_0_co_2.xml) at extreme temperatures.
+
+Weather stations are more likely to be located in wealthier and more populated areas, which makes them less useful for work in developing countries or for non-human variables such as agriculture. Their number and coverage constantly change, making it difficult to compare across regions or time ranges. Despite being the most accurate tool for measuring the current weather at their location, they may hide microclimates nearby.
+
+Thankfully, a large suite of data products has been developed to mitigate these issues. These generally consist of combining or ‘assimilating’ many data sources and analysis method into a ‘gridded dataset’ - the earth is divided into a latitude x longitude (x height) grid, and one value for a variable (temperature, precipitation, etc.) is provided at each gridpoint and timestep. These data products generally cover either the whole globe (or all global land), or are specialized to a certain region, and provide consistent coverage at each grid point location. 
+
+```{note}
+Some variables, especially relating to hydrology, may be better suited to station data, by providing single values for large regions such as river basins.
+```
+
+However, since the world is not made up of grids (i.e., the world is not broken up into 50 x 50 km chunks, within which all weather conditions are identical), some processing has to be done even for historical “weather” data, and other limitations arise. For historical data, this processing is one of the sources of differences between data products, and for climate model output, the simulation of sub-grid processes is the greatest source of uncertainty between models.
+
+```{caution}
+Keep in mind that just because a dataset exists at a certain resolution does not mean it is accurate at that resolution! 
+```
+
+The next section will briefly introduce how these products are generated, how to choose between them, and best practices for using “historical” data.
+
+## Gridded Weather Data Products
+
+**The Interpolation - Reanalysis Spectrum:**
+Historical data products differ by how they ["assimilate"](https://climatedataguide.ucar.edu/climate-data/atmospheric-reanalysis-overview-comparison-tables) (join observational with model data) or combine data, and how much “additional” information is added beyond (pre-processed) station data. They can be thought of as a rough spectrum ranging from ‘observational’ data products that merely statistically interpolate data into a grid to ‘reanalysis’ products that feed data products into a sort of climate model to produce a more complete set of variables. Some datasets are observational but include topographic and other physical information in their statistical methods, while some reanalysis datasets use pure model output for only some variables.
+
+Both ends of their spectrum have tradeoffs, and generalizable statements about these tradeoffs are hard to make because of differences in methodologies. The following are a few simplified rules of thumb:
+
+### “Observational” / Interpolated Datasets
+Examples: GISTEMP, GHCN, Wilmot and Matsuura (aka “UDel”), Berkeley Earth (aka “BEST”), HadCrut4, PRISM, CHIRPS etc.
+
+- Observations (from weather stations, satellites, etc.) are statistically interpolated into a grid with little or no physical information added (though topography and - less commonly - wind speed are occasionally included)
+- Products generally differ by which stations or other data sources are included and excluded
+
+```{panels}
+Strengths
+^^^
+- Simple, biases well-understood
[…]
