### 2023-07-18 atrisovic/weather-panel.github.io#77 @ks905383
Thank you for the comments! I have addressed them in merge #85.

**Responses to major comments**
> One example could be a time series of the volume of weather the climate data available

Added.

> In the Hands-On Excercise, Step 1 section, it might be useful to ask students to reflect on the components of variability - is there a trend? a clear seasonal cycle? other low and high-frequency variability?
> Plotting the data is a good way for them to check that the steps they have taken make sense. Maybe asking them to plot a map of the time mean would be a useful exercise.

This was a great idea - I've now added a section asking them to plot and reflect on a sample time series and a sample map, to think about their data (and also to make sure it has been pre-processed correctly).

> I suggest a note about satellite data products.

Good point.

**Responses to (major) minor comments**
> Overall, this section seems a bit too long.

Yeah, I think you're right - I've split up the section into more manageable chunks. I think it flows better now as well.

> when you list "run" in your terminology bullet list, I don't agree with your statement "Don't worry about this".

Yes, fully agree, thanks for catching. The wording has been changed, and also made reference to uncertainty ensembles that are sometimes provided with observational data products. If we expand this guide to include a section on future climate data, this will certainly be expanded on as well.

**Further changes**
In addition to the comments made above, I've also:
- split up the section into more and shorter subsections for ease of reading
- added more introductions and concluding texts to better guide the reader
- replaced links to papers with DOIs whenever possible, which will hopefully be more stable in the long run
