+++ tutorial-content/content/example-step3.md
@@ -2,11 +2,19 @@
 
 ## Aggregating the weather data
 
-One of the tutorial authors has produced a new tool for aggregating
-gridded data according to a shapefile: `xagg`. Since it handles
-partial grid-cell overlapping and arbitrary weighting grid, we
-recommend that it be used irrespective of the language being used
-elsewhere. Install it according to [the instructions from GitHub](https://github.com/ks905383/xagg).
+There exist tools in R and Python which aggregate gridded data
+according to a shapefile. In particular, `xagg` in Python and `stagg`
+in R handle partial grid-cell overlapping and arbitrary weighting
+grid, and we recommend that it be used irrespective of the language
+being used elsewhere. 
+
+## Setting up the environment
+
+````{tabbed} Python
+The approach uses `xarray` for gridded data, `geopandas` to work with shapefiles, and `xagg` to aggregate gridded data onto shapefiles. 
+
+Install `xagg` according to [the instructions from
+GitHub](https://github.com/ks905383/xagg).
 
 You will need to download a shapefile for the US counties. We use one
 from the ESRI community website:
@@ -15,11 +23,7 @@ from the ESRI community website:
 Save its contents to a folder `geo_data` in the `data` directory. The
 following code should be run from the `data` directory.
 
-## Setting up the environment
-
-The approach uses `xarray` for gridded data, `geopandas` to work with shapefiles, and `xagg` to aggregate gridded data onto shapefiles. 
-
-````{tabbed} Python
+Begin your code as follows:
 ```python
 import xarray as xr
 import geopandas as gpd
@@ -28,7 +32,22 @@ import numpy as np
 ```
 ````
 
-First, load the data from the previous steps:
+````{tabbed} R
+The approach uses `raster` to work with gridded data, `tigris` to work
+with shapefiles, and `stagg` to aggregate gridded data onto shapefiles.
+
+Install `stagg` according to [the instructions from
+GitHub](https://github.com/tcarleton/stagg).
+
+Begin your code as follows:
+```{code-block} R
+library(stagg)
+library(raster)
+library(tigris)
+```
+````
+
+Now, load the data from the previous steps and the county shapefile:
 
 ````{tabbed} Python
 ```python
@@ -41,6 +60,22 @@ ds_pop = xr.open_dataset('pcount/usap90ag.nc')
 
 # Load county shapefiles using geopandas
 gdf_counties = gpd.read_file('geo_data/UScounties.shp')
+```
+````
+
+````{tabbed} R
+```{code-block} R
+## Load data and expand to global grid
+global.extent <- extent(-180, 180, -90, 90)
+
+rr.tas <- stack("../data/climate_data/tas_day_BEST_historical_station_19800101-19891231.nc")
+rr.tas.padded <- extend(rr.tas, global.extent)
+rr.pop <- raster("../data/pcount/usap90ag.nc")
+rr.pop.padded <- extend(rr.pop, global.extent)
+rr.pop.padded[is.na(rr.pop.padded)] = 0
+
+## Load counties
+counties <- tigris::counties()
 ```
 ````
 
@@ -63,15 +98,20 @@ ds_tas = ds_tas.drop('land_mask')
 ```
 ````
 
+````{tabbed} R
+With `stagg`, we will let the library handle this with the
+`staggregate_polynomial` function (below).
+````
+
 ## Create map of pixels onto polygons
 
 We need to create a weight map of pixels to polygons. For each
 polygon, we need to know which pixels overlap it and by how much.
 
-`xagg` does this by creating polygons for each pixel in the gridded
-dataset, taking the intersect between each county polygon and all
-pixel polygons, and calculating the average area of overlap between
-the pixels that touch the polygon and the polygon.
[…]
