+++ tutorial-content/content/example-step4.md
@@ -1,50 +1,72 @@
-# Hands-On Exercise, Step 4: Producing results
+# Hands-On Exercise, Step 4: Regression results
 
-## Constructing annual polynomials summed over days
+## Necessary libraries
 
-We have county aggregated data from the previous step, but we also
-need to sum it to the annual level to match the mortality data. Again,
-this code should be run from a sister directory to the `data`
-directory.
+In this step, we will perform econometric regressions and plot the
+results. This requires some libraries, and we will also be calculating
+confidence intervals by hand which requires some additional code or
+libraries.
 
-Let's first turn the data from wide form into long form and label the
-years.
 
-````{tabbed} R
-```R
-library(reshape2)
+````{tabbed} Python
 
-clim <- read.csv("../data/climate_data/agg_vars.csv")
+We will use `PanelOLS` from `linearmodels` to perform the regression,
+but we also need the `t` function to compute t values for the
+confidence intervals. We will plot the result with `matplotlib`.
 
-clim2 <- melt(clim[, c(-1:-5)], id.vars='FIPS')
-clim2$date <- as.Date("1980-01-01") + as.numeric(gsub("tas_adj|tas_sq", "", clim2$variable))
-clim2$year <- as.numeric(substring(clim2$date, 1, 4))
+```python
+import pandas as pd
+import numpy as np
+from linearmodels.panel import PanelOLS
+from scipy.stats import t
+import matplotlib.pyplot as plt
 ```
 ````
 
-Now we can sum over years. To do this, we will need to base this on
-the first several characters of the column names (not variable row
-values).
-
 ````{tabbed} R
+We will use the `lfe` package to perform the regression and `ggplot2`
+to graph the result.
+
+To get confidence intervals from the `felm` function in `lfe`, we use
+the `predict.felm` function from the `felm-tools.R` library at
+https://github.com/jrising/research-common/blob/master/R/felm-tools.R
+
+
 ```R
 library(dplyr)
-
-clim3 <- clim2 %>% group_by(FIPS, year) %>% summarize(tas_adj=sum(value[substring(variable, 1, 7) == 'tas_adj']), tas_sq=sum(value[substring(variable, 1, 6) == 'tas_sq']))
+library(lfe)
+library(ggplot2)
+source("~/projects/research-common/R/felm-tools.R")
 ```
 ````
 
 ## Merging weather and outcome data
 
-And now we can merge in the mortality data! This is by county (FIPS
+Now we can merge in the mortality data! This is by county (FIPS
 code) and year. We also construct the death rate, as deaths per
 100,000 people in the population.
 
+````{tabbed} Python
+```python
+# Read data
+clim = pd.read_csv("../data/climate_data/agg_vars.csv")
+df = pd.read_csv("../data/cmf/merged.csv")
+
+# Merge datasets
+df2 = pd.merge(df, clim, how='left', left_on=['fips', 'year'], right_on=['FIPS', 'year'])
+
+# Create new variable
+df2['deathrate'] = 100000 * df2['deaths'] / df2['pop']
+df2.loc[df2['deathrate'] == float("inf"), 'deathrate'] = np.nan
+```
+````
+
 ````{tabbed} R
 ```R
+clim <- read.csv("../data/climate_data/agg_vars.csv")
 df <- read.csv("../data/cmf/merged.csv")
 
-df2 <- df %>% left_join(clim3, by=c('fips'='FIPS', 'year'))
[…]

[…]
