### 2023-08-13 e42382b
Update reduced-form-specification.md
Addressed all the reviewer comments for this file.
