+++ tutorial-content/content/example-step3.md
@@ -54,13 +54,13 @@ Now, load the data from the previous steps and the county shapefile:
 ```python
 # Load temperature data using xarray
 ds_tas = xr.open_dataset(
-    'climate_data/tas_day_BEST_historical_station_19800101-19891231.nc')
+    '../data/climate_data/tas_day_BEST_historical_station_19800101-19891231.nc')
 
 # Load population data using xarray 
-ds_pop = xr.open_dataset('pcount/usap90ag.nc')
+ds_pop = xr.open_dataset('../data/pcount/usap90ag.nc')
 
 # Load county shapefiles using geopandas
-gdf_counties = gpd.read_file('geo_data/UScounties.shp')
+gdf_counties = gpd.read_file('../data/geo_data/UScounties.shp')
 ```
 ````
 
@@ -195,7 +195,7 @@ ds2['FIPS'] = xr.apply_ufunc(np.char.add, ds2.STATE_FIPS, ds2.CNTY_FIPS)
 ds2['FIPS'] = ds2.FIPS.isel(year=0).drop('year')
 
 ds2.swap_dims({'poly_idx': 'FIPS'}).drop('poly_idx')
-ds2.to_dataframe().to_csv("climate_data/agg_vars.csv")
+ds2.to_dataframe().to_csv("../data/climate_data/agg_vars.csv")
 ```
 ````
 
@@ -217,6 +217,6 @@ county.tas.final$tas_adj <- county.tas.final$order_1 - county.tas.final$order_1.
 county.tas.final$tas_sq <- county.tas.final$order_2 - county.tas.final$order_2.base
 names(county.tas.final)[2] <- 'FIPS'
 
-write.csv(county.tas.final, 'climate_data/agg_vars.csv', row.names=F)
+write.csv(county.tas.final, '../data/climate_data/agg_vars.csv', row.names=F)
 ```
 ````
