### 2023-08-14 atrisovic/weather-panel.github.io#79 @jrising
@atrisovic I made the `../data/` edits on the PR #87, since I make other edits to the example step 3 there.

For the warnings and errors from `xagg`, here's how things look with the new version:
```
>>> weightmap = xa.pixel_overlaps(ds_tas, gdf_counties, weights=ds_pop.Population, subset_bbox=False)
creating polygons for each pixel...
lat/lon bounds not found in dataset; they will be created.
calculating overlaps between pixels and output polygons...
success!
>>> aggregated = xa.aggregate(ds_tas, weightmap)
aggregating tas_adj...
aggregating tas_sq...
all variables aggregated to polygons!
>>>
```
