### 2023-08-14 6b24d84
Update reduced-form-specification.md
