### 2023-08-16 43a5d43
Update functional-forms.md
Editing the bin section remains.
