+++ tutorial-content/content/functional-forms.md
@@ -28,10 +28,10 @@ which the temperature falls into that bin.
 To calculate bin observations, count up the number of timesteps where
 the weather predictor falls into each bin:
 
-$$T_{it} = \sum_{p \in i} \psi_{p} \sum \mathbf{1} \left \{  {T_{p i t} \in Bin_k} \right \}$$
+$$X_{it} = \sum_{p \in i} \psi_{p} \sum \mathbf{1} \left \{{X_{p i t} \in Bin_k} \right \}$$
 
 where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
-    
+
 ## [Polynomial](https://en.wikipedia.org/wiki/Polynomial_regression)
 
 Polynomial specifications balance smooth dose-response curves with the
@@ -39,7 +39,7 @@ flexibility of increasing the effective resolution of the curve by
 increasing the degree of the polynomial. Using more degrees improves
 the ability of the curve to fit the data, but may lead to
 over-fitting. To choose the number of terms in the polynomial,
-consider cross-validation.
+consider [cross-validation](#cross-validation).
 
 Another benefit of polynomials for climate change estimates is that
 they provide smooth extrapolation to higher temperatures. Again, it is
@@ -55,18 +55,18 @@ high-resolution weather values should be raised to powers before
 aggregating up to the regions in the economic data. That is, the 
 predictor for the $k$-th power of the polynomial is 
 
-$$f(T_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} T_{p i t}^k$$ 
+$$f(X_{it}^k)=\sum_{p \in \Theta(i)} \psi_{p} X_{p i t}^k$$ 
 
 where $\psi_{p}$ is the weight assigned to the $p$ grid cell.  
     
 The dose-response regression would then be applied as follows
 
-$$F(T_{it})=\sum_{k} \beta_k f(T_{it}^k)$$
+$$F(X_{it})=\sum_{k} \beta_k f(X_{it}^k)$$
 	
 while the coefficients can be interpreted as describing a local
 dose-response relationship:
 
-$$F(T_{pit})=\sum_{k} \beta_k T_{pit}^k$$
+$$F(X_{pit})=\sum_{k} \beta_k X_{pit}^k$$
 
 ## [Restricted cubic spline](https://support.sas.com/resources/papers/proceedings16/5621-2016.pdf)
 
@@ -86,19 +86,17 @@ title on cubic splines can be helpful in deciding the knot
 specifications.
 
 Once knot locations are determined, the weather data needs to be
-translated into RCS terms. As before let the gridded weather be $T_{p
-i t}$ and let there be $n$ knots, placed at $T_1<T_2<...<T_n$. Then we
+translated into RCS terms. As before let the gridded weather be $X_{p
+i t}$ and let there be $n$ knots, placed at $X_1<X_2<...<X_n$. Then we
 have a set of $(n-2)$ terms, here indexed by $k$ and defined as:
 
-$$f(T_{i t})_k= \sum_{p \in \Theta(i)} \psi_{p} \{(T_{p i
-	t}-T_k)^3_+ - (T_{p i t} - T_{n-1})^3_+
-	\frac{T_n-T_k}{T_n-T_{n-1}}+(T_{p i t} - T_{n})^3_+ \frac{T_{n-1}-T_k}{T_{n}-T_{n-1}}\}$$ 
[…]
