### 2023-08-16 8cc23c3
Update functional-forms.md
1. Updated reference link used in the bin section header to provide some direct reference to binning procedure. Added a few lines on binning.

2. Added a line in the first para on why we are avoiding timestep, s, and instead working with more aggregated time value, t.
