### 2023-08-21 5be888a
Update reduced-form-specification.md
Updated precip subsection
