### 2023-08-25 atrisovic/weather-panel.github.io#77 @ks905383
Detailed responses to minor comments:

***Sub-section: The NetCDF Data Format***
> Is there an assumption that students use STATA specifically? This reference to STATA kind of came out of nowhere.

This has been clarified; STATA is commonly used in economic analysis.

> - Python(NetCDF4) link doesn't work

Fixed.

> nco -> maybe direct students to use ncdump -h for meta data only rather than ncdump because it is usually way too much of a data dump.

Good point, this has been updated.

> As an open-source alternative to MATLAB, Python(NetCDF4) + numpy works very well and was typically the way Python users worked before xarray was developed.

Agreed; we're definitely keeping the mention in this section - though I think, at least for the purposes of this tutorial, we'll stick with `xarray` for simplicity for the rest of the code chunks.

***Sub-section: NetCDF Contents***
> it might be helpful to show a schematic of the netcdf data model: https://docs.unidata.ucar.edu/netcdf-c/current/netcdf_data_model.html

Added, thanks for the suggestion.

***Sub-section: NetCDF Header***
> I realize that you want students to do most of the work themselves and not provide them with data sets as examples, but a picture is often worth a thousand words. It would be nice to show the file header - you can use the xarray sample data so that you don't have to rely on external data sources: https://tutorial.xarray.dev/fundamentals/04.1_basic_plotting.html. You could also use this sample data in the plotting section (which I would also recommend).

Agreed - showing an image of the header is definitely a good idea. Using arrays built-in sample data

***Sub-section: Attributes***
> no_sleap -> no_leap

Fixed.

***Sub-section: Basic Vis...***
> the correct title is "An Introduction to Earth and Environmental Data Science"

Fixed.

> Cartopy section link not working

Fixed

***Sub-section: 2-D plotting***
> - Remember that Earth is a sphere and for most grids you cannot average over all lat/lon points in this way: https://docs.xarray.dev/en/stable/examples/area_weighted_temperature.html

Definitely worth bringing up - I've changed the example to note this.

***Sub-section: Maps***
> - Maybe add a note that other projections are available: https://scitools.org.uk/cartopy/docs/v0.15/crs/projections.html

I've added the following text:
> Note that which [map projection](https://www.usgs.gov/faqs/how-are-different-map-projections-used) you use will influence how you read the map. In the code examples below, we will use an equal-area projection, in which every grid cell in the gridded data is shown with its accurate relative area, to avoid visually overemphasizing data in regions with smaller geographic extent. To see which other projections are available, see the relevant parts of the documentations ([here](https://scitools.org.uk/cartopy/docs/v0.15/crs/projections.html) for cartopy/python, and [here](https://www.mathworks.com/discovery/map-projection.html) for Matlab)

***Sub-section: Gridded Data***
> - General note: it seems that the term “climate data” is being used to refer to “climate model output”. Climate data is not exclusively model output. For example, a 30-year climate normal from a weather station would be considered climate data.

I've changed references to 'data' to be more clear to what they're referring to (specifying climate model output, or historical "observational" output).

***Sub-section: Reanalysis Datasets***
> - Products also differ by which assimilation scheme is used.

I've clarified:
> Historical data products differ by how they [“assimilate”](https://climatedataguide.ucar.edu/climate-data/atmospheric-reanalysis-overview-comparison-tables) (join observational with model data) or combine data, and how much “additional” information is added beyond (pre-processed) station data.

***Sub-section: Warning, Station Data***
> - GHCN link not working

Fixed.
