### 2023-08-30 99861f4
Updates for issue #78
