+++ _sources/content/example-step2.md
@@ -1,4 +1,4 @@
-# Hands-On Exercise, Step 2: Understanding the outcomes
+# Hands-On Exercise, Step 2: Prepping the Demographic Data
 
 ## Thinking about the data-generating process
 
@@ -39,13 +39,19 @@ World, since the weather data is not very high resolution. Download it
 from
 <https://sedac.ciesin.columbia.edu/data/set/gpw-v3-population-count>.
 
-The code below assumes that you download the global population count
-grid as a `.bil` format at 2.5' resolution for 1990.
+The code below assumes that you download the USA population count grid
+as a `.bil` format at 2.5' resolution for 1990 (these are all options
+on the Gridded Population of the World download form). The zip file
+produced will contain a `usap90ag.bil` file, along with other
+associated files that are required for this file to be loaded. For the
+code below to work, the contents of the zip file should be placed in a
+`data/pcount` directory.
 
-This is global data, so it will be useful to clip it to the US and
-aggregate it to the scale of the weather. We also need it in NetCDF
-format, for the aggregation step. Again, this code assumes that it is
-being run from a directory `code`, sister to the `data` directory.
+Although this is US-specific data, the coverage extends far beyond the
+contiguous US. It will be useful to clip it more tightly and aggregate
+it to the scale of the weather. We also need it in NetCDF format, for
+the aggregation step. Again, this code assumes that it is being run
+from a directory `code`, sister to the `data` directory.
 
 `````{tab-set}
 ````{tab-item} R
@@ -63,10 +69,12 @@ writeRaster(rr3, "../data/pcount/usap90ag.nc4",
 ````
  
 ````{tab-item} Python
+You will need to install the `rioxarray` package, using `pip install
+rioxarray`, to `.bil` files.
+
 ```{code-block} python
-#!pip install rasterio # will work in Jupyter Notebook
-import xarray as xr
-rr = xr.open_rasterio("../data/pcount/usap90ag.bil")
+import rioxarray
+rr = rioxarray.open_rasterio("../data/pcount/usap90ag.bil")
 rr = rr.sel(band=1, drop=True)
 
 # Crop to US bounding box
@@ -87,6 +95,8 @@ rr3.to_dataset(name="Population").to_netcdf("../data/pcount/usap90ag.nc4")
 ```
 ````
 `````
+
+Further details on the use of geographic data are discussed in the next chapter.
 
 ## Downloading the mortality data
 
@@ -119,6 +129,12 @@ each of 12 age groups is reported with 8 characters per age group
 number (so, the first age group, 1 - 4 year-olds, is characters 19 -
 26; then 5 - 9 year-olds is reported in characters 27 - 34; and so
 on).
+
+The population file also contains US-wide population numbers,
+state-wide numbers, and county-level numbers. These are identified by
+a column in the data with a 1 (US), 2 (state), or 3 (county). In the
+code below, we label this the `type` column and only include type = 3
+data.
 
 ## Preparing the mortality data
 
[…]
