### 2023-09-05 atrisovic/weather-panel.github.io#78 @jrising
These were merged under PRs #86, #88, and #89.
