### 2023-09-06 eb0c4f1
add Python code and diagram
