### 2023-09-06 60a342a
add raster definition
