+++ tutorial-content/content/example-step2.md
@@ -39,6 +39,20 @@ World, since the weather data is not very high resolution. Download it
 from
 <https://sedac.ciesin.columbia.edu/data/set/gpw-v3-population-count>.
 
+```{note}
+Raster data is a type of digital image represented by reducible and 
+enlargeable grids. Each cell (often referred to as a pixel) within this 
+grid contains a value representing information, such as temperature, 
+elevation, or a color value for image data. Each pixel stores a single 
+value, and when visualized collectively, these pixels can represent 
+complex images or spatial information. See more 
+[here](https://desktop.arcgis.com/en/arcmap/latest/manage-data/raster-and-images/what-is-raster-data.htm) 
+and [here](https://datacarpentry.org/organization-geospatial/01-intro-raster-data).
+
+Common file formats for raster data include GeoTIFF, NetCDF, HDF, JPEG2000, 
+and BIL, among others.
+```
+
 The code below assumes that you download the USA population count grid
 as a `.bil` format at 2.5' resolution for 1990 (these are all options
 on the Gridded Population of the World download form). The zip file
@@ -55,6 +69,10 @@ from a directory `code`, sister to the `data` directory.
 
 `````{tab-set}
 ````{tab-item} R
+
+The R library `raster` is used for reading, writing, manipulating, analyzing 
+and modeling of spatial data. 
+
 ```{code-block} R
 library(raster)
 rr <- raster("../data/pcount/usap90ag.bil")
@@ -70,7 +88,7 @@ writeRaster(rr3, "../data/pcount/usap90ag.nc4",
  
 ````{tab-item} Python
 You will need to install the `rioxarray` package, using `pip install
-rioxarray`, to `.bil` files.
+rioxarray`, to open the `.bil` files.
 
 ```{code-block} python
 import rioxarray
