### 2023-09-06 9034bf6
add Python code snippets
