+++ tutorial-content/content/weighting-schemes.md
@@ -68,9 +68,9 @@ CSV, spaces for ASC).
 
 ### The spatial gridding scheme
 
-The spatial gridding scheme is determined by 6 numbers: a latitude and longitude of an origin point, a horizontal and vertical cell lengths, and a number of rows and columns.
+The spatial gridding scheme is determined by 6 numbers: a latitude and longitude of an origin point, meridional (north-south) and zonal (east-west) cell lengths, and a number of rows and columns.
     - The most common origin point is the location of the lower-left corner of the lower-left grid cell. For example, for a global dataset, that might be 90°S, 180°W, which is represented in x, y coordinates as (-180, -90). Sometimes (particularly with NetCDF files), grid cell center locations will be used instead.
-    - Grid cell sizes are often given as decimal representation of fractions of a degree, such as $0.0083333333333 = 1 / 120$ of a degree. This is the grid cell size needed globally to ensure a km-scale resolution. Usually the horizontal and vertical grid cell lengths are the same, and reported as a single number.
+    - Grid cell sizes are often given as decimal representation of fractions of a degree, such as $0.0083333333333 = 1 / 120$ of a degree. This is the grid cell size needed globally to ensure a km-scale resolution. Usually, the meridional and zonal grid cell lengths are the same and are reported as a single number.
     - The number of grid cells is the most common way to describe the spatial coverage of the dataset. A global dataset will have 180 / cellsize rows and 360 / cellsize columns.
 
 
@@ -121,7 +121,7 @@ which has some extensive examples of working with geospatial raster data.
 ```
 ````
 ````{tab-item} Python
-Take a look at <https://github.com/jrising/research-common/tree/master/python/geogrid>.
+Take a look at the Python package [`rasterio`](https://rasterio.readthedocs.io/en/stable/). GeoGrid may also be useful: <https://github.com/jrising/research-common/tree/master/python/geogrid>.
 ````
 `````
 
@@ -142,13 +142,20 @@ the corner.
 
 If your datasets cover the same geographic extent, and if the
 resolution of your weighting dataset is an even multiple of the
-resolution of your weather dataset, your job is easy. Just either
-aggregate or disaggregate grid cells so that they have the same
-resolution as the weather data, using the code below. In other cases,
-you will need to follow three steps, and generally require
-upsampling, downsampling, and cropping.
+resolution of your weather dataset, your job is easy. 
+You'll simply aggregate or disaggregate grid cells so that they have the 
+same resolution as the weather data. 
+
+In situations that don't meet these conditions, you will need to follow 
+the three methods illustrated below: upsampling, downsampling, and cropping.
 
 ### Upsampling - increasing data resolution
+
+The variable `rr` represents the data on the grid, or a matrix or 2D array where each 
+element corresponds to a value (like temperature, elevation, or some other measurement) 
+at a particular grid cell. The structure of this matrix (`rr` in this case) essentially 
+overlays a spatial grid, where the row and column indices of the matrix can be mapped to 
[…]
