### 2023-09-07 c4767a3
change psi to w and remove reference to geogrid
