### 2023-09-11 bfaaeb7
add exercise intro
