### 2023-09-11 25719db
add exercise intro
