### 2023-09-11 #90 @jrising
@kls2177 We have now made changes in response to all of the items you identified. Thank you again for all of your time on the tutorial-- we believe that it has improved significantly with your suggestions.

Responding to your major comments here:

 - The first section of chapter 1 is now split into more manageable chunks.
 - We have now added an introduction to the exercise, at the beginning of step 1. This lays out the goal, and some important considerations that will inform our decisions throughout the exercise.
 - Section 2 has been split up to highlight the most important variables, putting the rest after the main content.
 - We appreciate the challenge creating a sensible introduction across the chapters while also including relevant steps of the exercise. The main element "work" of step 2 of the exercise is actually the first part of that section, "Thinking about the data-generating process", which matches the goal of chapter 2. However, since this is the point where we think about the dependent variable, it makes sense to prepare that data. Unfortunately, this auxiliary work ends up dominating the length of step 2. However, we think it is still useful to break it up, rather than include this in step 3.
 - Thank you so much for letting us know about the problems with step 3 and 4. We have now fixed bugs in the underlying library and provided an alternative implementation using a recently-developed R package.
 - We have now added a discussion of the main considerations that go into our regression (and what we miss), and provide a list of ideas for robustness checks.

More information is available in the issues. If anything is unclear, do not hesitate to let us know!
