### 2023-09-11 #90 @kls2177
Thanks very much @jrising for responding to my comments. I am just ramping up with the first week of classes, so I will likely not be able to take another look over everything for a week or two. Thanks for your patience.
