### 2024-01-30 atrisovic/weather-panel.github.io#77 @kls2177
@ks905383

I appreciate these updates. I particularly like the updates to the Hands-On Exercise.
