### 2024-01-30 atrisovic/weather-panel.github.io#78 @kls2177
@azharhsain

Thanks for these updates. In the Hands-On Exercise, Step 2, there are still a few 20C rather than 20$^{\circ}$C. Please update these.
