### 2024-01-30 atrisovic/weather-panel.github.io#79 @kls2177
@atrisovic

Thanks for these updates. A couple of outstanding comments:

==============================
Section: Constructing weather averages within spatial units

Sub-section: Approach 1

It would help if you defined some of your variables: longitude0, longitude1, gridwidth, etc.

==============================
Hands-On Exercise, Step 3

During the `aggregated = xa.aggregate(ds_tas, weightmap)` step, I am getting the following repeated warning:

**RuntimeWarning: invalid value encountered in divide
  a2 = a2/a2.sum()**
