### 2024-01-30 atrisovic/weather-panel.github.io#80 @kls2177
@jrising

Thanks for the updates.

After the bulleted list in the "Running the Regression" section, there is an equation that is not rendering correctly. Please update.

**Major Issue:**

I am having some issues with the code that likely relate to my new issue in Step 3 of the Hands-On Exercise:

When I try to run the regression using python, I get the following error:

**ValueError: exog does not have full column rank. If you wish to proceed with model estimation irrespective of the numerical accuracy of coefficient estimates, you can set check_rank=False.**

If I add, "check_rank=False", I get the following error:

**AbsorbingEffectError:
The model cannot be estimated. The included effects have fully absorbed
one or more of the variables. This occurs when one or more of the dependent
variable is perfectly explained using the effects included in the model.

The following variables or variable combinations have been fully absorbed
or have become perfectly collinear after effects are removed:

          tas_adj
          tas_sq

Set drop_absorbed=True to automatically drop absorbed variables.**

If I add "drop_absorbed=True", this still gives me the same error. I think this is related to the RunTimeError I was getting in Step 3. Please revise.

Also, in the plotting section (and throughout), please revise C to deg C using Markdown as follows:

`plt.xlabel('Daily temperature ($^{\circ}$C)')`
