### 2024-01-30 #90 @kls2177
Hi @atrisovic, @jrising, I have added some new comments. Overall, I think the updates have addressed most of my questions.

However, **I am still have technical problems with the python code**. I have suggested some edits, but I am still stuck. I am getting a RunTimeError in Step 3 and I think this error is causing issues in Step 4.

Please see my comments for Step 4 [here](https://github.com/atrisovic/weather-panel.github.io/issues/80)

and Step 3 [here](https://github.com/atrisovic/weather-panel.github.io/issues/79)

I have also added **new comments** to my other issues for the other section [here](https://github.com/atrisovic/weather-panel.github.io/issues/78), [here](https://github.com/atrisovic/weather-panel.github.io/issues/77) and [here](https://github.com/atrisovic/weather-panel.github.io/issues/76)
