### 2024-01-31 atrisovic/weather-panel.github.io#80 @kls2177
Here is some additional information:

@ks905383 The population data has a lot of zeros and the temperature data has masked values over ocean.

Also, note that when I plot the map of summer temperature (Hands-On Exercise, Step 1), there are a few spots over the US that are also missing data (figure attached)
.
