### 2024-02-14 atrisovic/weather-panel.github.io#80 @kls2177
Hello! I noticed that xagg has been updated and I gave it a try. It now works, but the figure that I get looks different from the one in the jupyterbook.

![fit](https://github.com/atrisovic/weather-panel.github.io/assets/13908222/6e6164ad-9d40-458a-8637-f1f2c49b2567)
