### 2024-02-15 28ad4a9
Minor fixes to hands-on exercises - switches lingering references to a directory "sources" to a directory "data" instead in paths in step 1 - switches ".nc4" to ".nc" in saved filenames - changes `.drop()` to `.drop_vars()` to avoid xarray error - renames "latitude"/"longitude" to lat/lon in python example to better fit conventions and make plotting example code work - adds `.mean()` call, minor syntax fixes, to plotting example
