### 2024-02-27 atrisovic/weather-panel.github.io#80 @kls2177
hi @ks905383,

I went through all the code and redownloaded all the datasets again and now I get the right figure - yay! I can't pinpoint what the issue was, but all seems to be working well now.
