### 2024-02-27 #90 @kls2177
@kyleniemeyer, @jrising, @atrisovic, I have gone through the revisions and tested the code and all is working well. I recommend this resource for publication in JOSE after all the revisions are complete.

Thanks for your patience!
