### 2024-02-28 #90 @atrisovic
Hi @kls2177 wonderful news! Thank you very much for your help!

A question for my co-authors: what do you think about removing `.net` from the title in case we want to move the domain in the next few years? I opened an [issue](https://github.com/atrisovic/weather-panel.github.io/issues/92) where we could quickly revisit it.
