### 2024-03-03 df79311
replace 20 C with 2 degC
