### 2024-03-04 atrisovic/weather-panel.github.io#79 @jrising
@kls2177 Thanks for all of your suggestions. We have now added definitions for the variables under Approach 1 and addressed the warming in `xagg`.
