### 2024-03-04 atrisovic/weather-panel.github.io#78 @jrising
@kls2177 Good catch. We have now fixed the `20 C` texts, under commit df79311bdb7c31385e5e8568ab0859b65e0f3b25.
