### 2024-03-04 atrisovic/weather-panel.github.io#77 @jrising
@kls2177 Thanks for identifying these. We meant to convert the variable names, but have now decided to leave the conversion to the end. In particular,

 - We have made the first three specific changes you recommend, changing `lat`/`lon` to `latitude`/`longitude`.
 - In the `cartopy`, we have added the `.mean('time')`, as you point out.
 - In our `contourf` line, we were referring to `df` rather than `df_summer`, which might have caused trouble for you. But concerning the `ax` variable, in our test, we do not need to explicitly refer to it because `matplotlib` overlays the subsequent `contourf` command to the pre-defined axes.
 - We have switched a couple of lingering references to a `sources` directory to `data`.

The relevant commits are 28ad4a9edd63775f8fe7cdc7808aadb3e26d6007 and 7815a99294b405a3c4983bf6f2be4897bcd230f4.
