### 2024-03-04 atrisovic/weather-panel.github.io#76 @jrising
@kls2177 Good point. I now changed all of the table of contents headings and within-section headings to use first-word-style capitalization.
