### 2024-03-04 atrisovic/weather-panel.github.io#80 @jrising
@kls2177 Thank you for going through the whole thing again. I really appreciate all the time you have spent testing this.

I have also fixed the rendering of that equation.
