### 2024-03-04 atrisovic/weather-panel.github.io#92 @jrising
I think your second option is really nice. I don't love asking to switch the paper title at this stage, but I guess there's no harm in asking.
