### 2024-03-13 atrisovic/weather-panel.github.io#92 @jrising
@atrisovic Actually, did we already do this a long time ago? In https://github.com/atrisovic/weather-panel.github.io/blob/master/paper.md, I see the title "A practical guide to climate econometrics: Navigating key decision points in weather and climate data analysis". I see where this was previously the title in that paper, but we changed it after the initial submission. Am I looking in the wrong spot?
