### 2024-03-13 atrisovic/weather-panel.github.io#92 @atrisovic
Ah, you are completely right. We did it - I forgot about it and thought that title was for the website logo. 😅
