### 2024-03-15 #90 @jrising
@kyleniemeyer We realized that we actually already did this, shortly after the initial submission. The paper title is "A practical guide to climate econometrics: Navigating key decision points in weather and climate data analysis", which is perfect. It's ready for you!
