### 2024-03-18 #90 @atrisovic
Hi @kyleniemeyer I just created a new release for JOSE with the DOI https://doi.org/10.5281/zenodo.10833818 =)
