### 2024-05-22 #90 @labarba
I see that the version number is [v2.0](https://github.com/atrisovic/weather-panel.github.io/releases/tag/v2.0) in the repository, but we see v1.2 at the top of this issue. As I don't see a version command was run here, I will run it now.
