### 2024-05-22 7158a1a
Update paper.md
I found a broken link to Jupyter Book and took the chance to make a couple of small editing fixes too.
