+++ paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 The use of econometrics to study how social, economic, and biophysical systems respond to weather has started a torrent of new research [@carleton2016social]. It is allowing researchers to better understand the impacts of climate change, disaster risk and responses, resource management, human behavior, and sustainable development.
 
-The research often features weather regression modeling and analyzing socioeconomic outcomes as they vary over time. The researchers need to collect, integrate, and analyze many different datasets such as historical climate data, future climate models, GIS information, and administrative data, in order to obtain meaningful and comprehensive results. While the challenges of climate econometrics are similar across all problems, the appropriate use of data, its integration and aggregation is unique to each research question, thereby presenting a stumbling block to early-career researchers. No single sequence of steps can be widely applied for resolving these issues. Similarly, software aimed at facilitating steps in this process needs to be fully understood and operated with care. 
+The research often features weather regression modeling and analyzing socioeconomic outcomes as they vary over time. The researchers need to collect, integrate, and analyze many different datasets such as historical climate data, future climate models, GIS information, and administrative data, in order to obtain meaningful and comprehensive results. While the challenges of climate econometrics are similar across all problems, the appropriate use of data, its integration and aggregation are unique to each research question, thereby presenting a stumbling block to early-career researchers. No single sequence of steps can be widely applied for resolving these issues. Similarly, software aimed at facilitating steps in this process needs to be fully understood and operated with care. 
 
 In light of this, we present a new tutorial that offers step-by-step guidance on carrying out a climate econometric analysis. Moreover, it features reusable code snippets in multiple programming languages and points to several pre-existing packages while providing some introductory comments on their appropriate use. The tutorial is structured into the following sections:
 
@@ -50,11 +50,11 @@ The audiences for this tutorial are researchers and students trained in economet
 
 Over the past year, the tutorial has been actively used by Ph.D. and Master's students at the London School of Economics, and it has been highlighted at multiple workshops. Several independent research projects have been developed by following the steps of the tutorial, which provides additional evidence for the success of its learning objectives.
 
-In this rapidly evolving and interdisciplinary field, our tutorial is free and open-source. It is implemented as a [JupyterBook](https://jupyterbook.org/intro) and available online through [Github Pages](https://pages.github.com) at [climateestimate.net](https://climateestimate.net/getting-started.html). Its implementation also features a deployment workflow through [GitHub Actions](https://github.com/features/actions), meaning that every time there is an update to the `master` branch, the action will automatically build a new JupyterBook and update the tutorial website within minutes. As a result, contributors can change the tutorial content in markdown files without worrying about updating it on the web.
+In this rapidly evolving and interdisciplinary field, our tutorial is free and open-source. It is implemented as a [JupyterBook](https://jupyterbook.org/) and available online through [Github Pages](https://pages.github.com) at [climateestimate.net](https://climateestimate.net/getting-started.html). Its implementation also features a deployment workflow through [GitHub Actions](https://github.com/features/actions), meaning that every time there is an update to the `master` branch, the action will automatically build a new JupyterBook and update the tutorial website within minutes. As a result, contributors can change the tutorial content in markdown files without worrying about updating it on the web.
 
 # Statement of Need
 
-The existing academic work provides overviews of climate science for econometricians [@hsiang2018economist] and the theoretical backing of these methods [@hsiang2016climate]. However, information that focuses on analysis methods is still lacking, and new resources are needed to offer practical help to researchers. For example, several steps in the research process can be challenging to students and social science researchers because of the unfamiliarity of weather data [@nissan2019use]. As a result, some of the most common mistakes are: 
[…]
