### 2024-05-22 #90 @labarba
I found a broken link to the JupyterBook landing pages, and in fixing it, added a couple of tiny edits.
https://github.com/atrisovic/weather-panel.github.io/pull/93
