### 2024-05-23 #90 @labarba
I notice that the [LICENSE](https://github.com/atrisovic/weather-panel.github.io/blob/master/LICENSE) on the repository is CC-BY-SA. The footer of the JupyterBook, however, has a copyright symbol and year only. Would you edit the `_config.yml` to include the license information in the footer?
