### 2024-05-23 6acc7db
Update _config.yml
I believe this change will show the license information correctly in the footer
