### 2024-05-23 #90 @labarba
PR here: https://github.com/atrisovic/weather-panel.github.io/pull/94
