### 2024-05-23 #90 @labarba
Congratulations, @jrising and @atrisovic, your JOSE paper is now published! 🚀
Thank you so much for your patience.

For [citations](https://climateestimate.net/content/contributors.html#citing-this-tutorial), you may use:

> Rising et al., (2024). A practical guide to climate econometrics: Navigating key decision points in weather and climate data analysis. Journal of Open Source Education, 7(75), 90, https://doi.org/10.21105/jose.00090

Note that TheOJ's Crossref deposit is more conducive to collecting citations than a Zenodo (Datacite) DOI, and should be captured by Google Scholar.

Hugely grateful to our Editor: @kyleniemeyer, and Reviewers: @jwagemann, @kls2177 — this adventure in publishing is made possible thanks to your contributions 🙏
