### 2024-05-23 #90 @labarba
@openjournals/dev Can I request your assistance here? I'm not sure whey the JOSE "published" badge is failing…
