### 2024-05-24 #90 @atrisovic
Thank you very much @labarba  and all! The badge looks fine in the repository's readme: https://github.com/atrisovic/weather-panel.github.io :))
